Ganeti.BasicTypes

Documentation

data GenericResult a b

Generic monad for our error handling mechanisms.

Constructors

Bad a

Ok b

genericResult :: (a -> c) -> (b -> c) -> GenericResult a b -> c

Sum type structure of GenericResult.

type Result = GenericResult String

Type alias for a string Result.

class FromString a where

Type class for things that can be built from strings.

Methods

mkFromString :: String -> a

newtype ResultT a m b

This is a monad transformation for Result. It's implementation is based on the implementations of MaybeT and ErrorT.

ResultT is very similar to ErrorT, but with one subtle difference: If mplus combines two failing operations, errors of both of them are combined.

Constructors

ResultT

Fields

runResultT :: m (GenericResult a b)

elimResultT :: Monad m => (a -> ResultT a' m b') -> (b -> ResultT a' m b') -> ResultT a m b -> ResultT a' m b'

withError :: MonadError e m => (e' -> e) -> GenericResult e' a -> m a

Changes the error message of a result value, if present. Note that since GenericResult is also a MonadError, this function is a generalization of (Error e') => (e' -> e) -> GenericResult e' a -> GenericResult e a

withErrorT :: (Monad m, FromString e) => (e' -> e) -> ResultT e' m a -> ResultT e m a

Changes the error message of a ResultT value, if present.

toError :: MonadError e m => GenericResult e a -> m a

Lift a Result value to any MonadError. Since ResultT is itself its instance, it's a generalization of Monad m => GenericResult a b -> ResultT a m b.

toErrorBase :: (MonadBase b m, MonadError e m) => ResultT e b a -> m a

Lift a ResultT value into any MonadError with the same base monad.

toErrorStr :: (MonadError e m, FromString e) => Result a -> m a

An alias for withError mkFromString, which is often used to lift a pure error to a monad stack. See also annotateResult.

tryError :: MonadError e m => m a -> m (Either e a)

Run a given computation and if an error occurs, return it as Left of Either. This is a generalized version of try.

mkResultT :: (Monad m, FromString e) => m (Result a) -> ResultT e m a

Converts a monadic result with a String message into a ResultT with an arbitrary Error.

Expects that the given action has already taken care of any possible errors. In particular, if applied on IO (Result a), any exceptions should be handled by the given action.

See also toErrorStr.

mkResultT' :: (Monad m, FromString e, Show s) => m (GenericResult s a) -> ResultT e m a

Generalisation of mkResultT accepting any showable failures.

isOk :: GenericResult a b -> Bool

Simple checker for whether a GenericResult is OK.

isBad :: GenericResult a b -> Bool

Simple checker for whether a GenericResult is a failure.

justOk :: [GenericResult a b] -> [b]

Simple filter returning only OK values of GenericResult

justBad :: [GenericResult a b] -> [a]

Simple filter returning only Bad values of GenericResult

eitherToResult :: Either a b -> GenericResult a b

Converter from Either to GenericResult.

isLeft :: Either a b -> Bool

Check if an either is Left. Equivalent to isLeft from Data.Either version 4.7.0.0 or higher.

isRight :: Either a b -> Bool

Check if an either is Right. Equivalent to isRight from Data.Either version 4.7.0.0 or higher.

annotateResult :: (MonadError e m, FromString e) => String -> Result a -> m a

Annotate an error with an ownership information, lifting it to a MonadError. Since Result is an instance of MonadError itself, it's a generalization of type String -> Result a -> Result a. See also toErrorStr.

annotateError :: (MonadError e m, FromString e, Monoid e) => String -> m a -> m a

Annotate an error with an ownership information inside a MonadError. See also annotateResult.

failError :: (MonadError e m, FromString e) => String -> m a

Throws a String message as an error in a MonadError. This is a generalization of Bad. It's similar to fail, but works within a MonadError, avoiding the unsafe nature of fail.

handleErrorT :: (Monad m, FromString e) => (e' -> ResultT e m a) -> ResultT e' m a -> ResultT e m a

A synonym for flip catchErrorT.

catchErrorT :: (Monad m, FromString e) => ResultT e' m a -> (e' -> ResultT e m a) -> ResultT e m a

Catches an error in a ResultT value. This is similar to catchError, but in addition allows to change the error type.

orElse :: MonadError e m => m a -> m a -> m a

If the first computation fails, run the second one. Unlike mplus instance for ResultT, this doesn't require the Monoid constrait.

iterateOk :: (a -> GenericResult b a) -> a -> [a]

Iterate while Ok.

Misc functionality

select

Arguments

:: a
default result

-> [(Bool, a)]
list of "condition, result"

-> a
first result which has a True condition, or default

Return the first result with a True condition, or the default otherwise.

runListHead :: a -> (b -> a) -> [b] -> a

Apply a function to the first element of a list, return the default value, if the list is empty. This is just a convenient combination of maybe and listToMaybe.

Lookup of partial names functionality

data MatchPriority

The priority of a match in a lookup result.

Constructors

ExactMatch

MultipleMatch

PartialMatch

FailMatch

data LookupResult

The result of a name lookup in a list.

Constructors

LookupResult

Fields

lrMatchPriority :: MatchPriority

The result type | Matching value (for ExactMatch, PartialMatch), Lookup string otherwise

lrContent :: String

prefixMatch

Arguments

:: String
Lookup

-> String
Full name

-> Bool
Whether there is a prefix match

Check for prefix matches in names. Implemented in Ganeti core utils.text.MatchNameComponent as the regexp r^%s(..*)?$ % re.escape(key)

goodMatchPriority :: MatchPriority -> Bool

Is the lookup priority a good one?

goodLookupResult :: LookupResult -> Bool

Is the lookup result an actual match?

compareNameComponent

Arguments

:: String
Canonical (target) name

-> String
Partial (lookup) name

-> LookupResult
Result of the lookup

Compares a canonical name and a lookup string.

chooseLookupResult :: String -> String -> LookupResult -> LookupResult

lookupName

Arguments

:: [String]
List of keys

-> String
Lookup string

-> LookupResult
Result of the lookup

Find the canonical name for a lookup string in a list of names.

newtype ListSet a

Wrapper for a Haskell Set

This type wraps a Set and it is used in the Haskell to Python opcode generation to transform a Haskell Set into a Python list without duplicate elements.

Constructors

ListSet

Fields

unListSet :: Set a

emptyListSet :: ListSet a

newtype Down a

Constructors

Down a