Ganeti.Compat

Description

Compatibility helper module.

This module holds definitions that help with supporting multiple library versions or transitions between versions.

Documentation

rwhnf :: Strategy a

Wrapper over the function exported from Control.Parallel.Strategies.

This wraps either the old or the new name of the function, depending on the detected library version.

finiteBitSize :: Bits a => a -> Int