Ganeti.Confd.Client

Description

Implementation of the Ganeti Confd client functionality.

Documentation

getConfdClient :: IO ConfdClient

Builds a properly initialized ConfdClient

query :: ConfdClient -> ConfdRequestType -> ConfdQuery -> IO (Maybe ConfdReply)

Sends a query to all the Confd servers the client is connected to. Returns the most up-to-date result according to the serial number, chosen between those received before the timeout.

updateConfdReply :: ConfdReply -> Maybe ConfdReply -> Maybe ConfdReply

queryOneServer :: MVar () -> MVar (Maybe ConfdReply) -> ConfdRequestType -> ConfdQuery -> HashKey -> (String, PortNumber) -> IO ()