Ganeti.Config

Description

Implementation of the Ganeti configuration database.

Documentation

type LinkIpMap = Map String (Map String String)

Type alias for the link and ip map.

Operations on the whole configuration

readConfig :: FilePath -> IO (Result String)

parseConfig :: String -> Result ConfigData

encodeConfig :: ConfigData -> String

loadConfig :: FilePath -> IO (Result ConfigData)

Wrapper over readConfig and parseConfig.

saveConfig :: Handle -> ConfigData -> IO ()

Wrapper over hPutStr and encodeConfig.

Query functions

withMissingParam :: String -> (a -> ErrorResult b) -> Maybe a -> ErrorResult b

computeDiskNodes :: Disk -> Set String

instDiskNodes :: ConfigData -> Instance -> Set String

instNodes :: ConfigData -> Instance -> Set String

Computes all nodes of an instance.

instSecondaryNodes :: ConfigData -> Instance -> Set String

getNodeInstances :: ConfigData -> String -> ([Instance], [Instance])

Get instances of a given node. The node is specified through its UUID.

getNodeRole :: ConfigData -> Node -> NodeRole

Computes the role of a node.

getMasterNodes :: ConfigData -> [Node]

Get the list of the master nodes (usually one).

getMasterCandidates :: ConfigData -> [Node]

Get the list of master candidates, not including the master itself.

getMasterOrCandidates :: ConfigData -> [Node]

Get the list of master candidates, including the master.

getMasterNetworkParameters :: ConfigData -> MasterNetworkParameters

Get the network parameters for the master IP address.

getOnlineNodes :: ConfigData -> [Node]

Get the list of online nodes.

getDefaultNicLink :: ConfigData -> String

Returns the default cluster link.

getDefaultHypervisor :: ConfigData -> Hypervisor

Returns the default cluster hypervisor.

getInstancesIpByLink :: LinkIpMap -> String -> [String]

Returns instances of a given link.

getItem :: String -> String -> Map String a -> ErrorResult a

getItem' :: String -> String -> Map ByteString a -> ErrorResult a

getNode :: ConfigData -> String -> ErrorResult Node

Looks up a node by name or uuid.

getInstance :: ConfigData -> String -> ErrorResult Instance

Looks up an instance by name or uuid.

getInstanceByName :: ConfigData -> String -> ErrorResult Instance

getDisk :: ConfigData -> String -> ErrorResult Disk

Looks up a disk by uuid.

getFilterRule :: ConfigData -> String -> ErrorResult FilterRule

Looks up a filter by uuid.

getGroup :: ConfigData -> String -> ErrorResult NodeGroup

Looks up a node group by name or uuid.

getGroupNdParams :: ConfigData -> NodeGroup -> FilledNDParams

Computes a node group's node params.

getGroupIpolicy :: ConfigData -> NodeGroup -> FilledIPolicy

Computes a node group's ipolicy.

getGroupDiskParams :: ConfigData -> NodeGroup -> GroupDiskParams

Computes a group's (merged) disk params.

getGroupNodes :: ConfigData -> String -> [Node]

Get nodes of a given node group.

getGroupInstances :: ConfigData -> String -> ([Instance], [Instance])

Get (primary, secondary) instances of a given node group.

defaultHvStateParams :: FilledHvStateParams

getFilledHvStateParams :: ConfigData -> Node -> FilledHvState

Retrieves the node's static hypervisor state parameters, missing values filled with group's parameters, missing group parameters are filled with cluster's parameters. Currently, returns hvstate parameters only for the default hypervisor.

getFilledInstHvParams :: [String] -> ConfigData -> Instance -> HvParams

Retrieves the instance hypervisor params, missing values filled with cluster defaults.

getFilledInstBeParams :: ConfigData -> Instance -> ErrorResult FilledBeParams

Retrieves the instance backend params, missing values filled with cluster defaults.

getFilledInstOsParams :: ConfigData -> Instance -> OsParams

Retrieves the instance os params, missing values filled with cluster defaults. This does NOT include private and secret parameters.

getInstPrimaryNode :: ConfigData -> String -> ErrorResult Node

Looks up an instance's primary node.

getDrbdDiskNodes :: ConfigData -> Disk -> [Node]

getInstAllNodes :: ConfigData -> String -> ErrorResult [Node]

Retrieves all the nodes of the instance.

As instances not using DRBD can be sent as a parameter as well, the primary node has to be appended to the results.

getInstDisks :: ConfigData -> String -> ErrorResult [Disk]

Get disks for a given instance. The instance is specified by name or uuid.

getInstDisksFromObj :: ConfigData -> Instance -> ErrorResult [Disk]

Get disks for a given instance object.

collectFromDrbdDisks :: Monoid a => (String -> String -> Int -> Int -> Int -> Private DRBDSecret -> a) -> Disk -> a

getDrbdSecretsForDisk :: Disk -> [DRBDSecret]

getDrbdMinorsForDisk :: Disk -> [(Int, String)]

Returns the DRBD minors of a given Disk

getDrbdMinorsForNode :: String -> Disk -> [(Int, String)]

getDrbdMinorsForInstance :: ConfigData -> Instance -> ErrorResult [(Int, String)]

Returns the DRBD minors of a given instance

rolePrimary :: String

roleSecondary :: String

getInstMinorsForNode

Arguments

:: ConfigData

-> String
The UUID of a node.

-> Instance

-> [(String, Int, String, String, String, String)]

Gets the list of DRBD minors for an instance that are related to a given node.

buildLinkIpInstnameMap :: ConfigData -> LinkIpMap

Builds link -> ip -> instname map. For instances without a name, we insert the uuid instead.

TODO: improve this by splitting it into multiple independent functions:

abstract the "fetch instance with filled params" functionality

abstsract the [instance] -> [(nic, instance_name)] part

etc.

getGroupOfNode :: ConfigData -> Node -> Maybe NodeGroup

Returns a node's group, with optional failure if we can't find it (configuration corrupt).

getNodeNdParams :: ConfigData -> Node -> Maybe FilledNDParams

Returns a node's ndparams, filled.

Network

getNetwork :: ConfigData -> String -> ErrorResult Network

Looks up a network. If looking up by uuid fails, we look up by name.

MACs

type MAC = String

getAllMACs :: ConfigData -> [MAC]

Returns all MAC addresses used in the cluster.

DRBD secrets

getAllDrbdSecrets :: ConfigData -> [DRBDSecret]

LVs

type NodeLVsMap = MultiMap String LogicalVolume

A map from node UUIDs to

FIXME: After adding designated types for UUIDs, use them to replace String here.

getInstanceLVsByNode :: ConfigData -> Instance -> ErrorResult NodeLVsMap

getAllLVs :: ConfigData -> ErrorResult (Set LogicalVolume)

ND params

class NdParamObject a where

Type class denoting objects which have node parameters.

Methods

getNdParamsOf :: ConfigData -> a -> Maybe FilledNDParams