Ganeti.ConfigReader

Description

Implementation of configuration reader with watching support.

Documentation

type ConfigReader = IO (Result ConfigData)

A type for functions that can return the configuration when executed.

data ReloadModel

Constructors

ReloadNotify

ReloadPoll Int

data ServerState

Constructors

ServerState

Fields

reloadModel :: ReloadModel

reloadTime :: Integer

reloadFStat :: FStat

maxIdlePollRounds :: Int

watchInterval :: Int

pollInterval :: Int

reloadRatelimit :: Integer

initialPoll :: ReloadModel

data ConfigReload

Constructors

ConfigToDate

ConfigReloaded

ConfigIOError

Configuration handling

Helper functions

moveToPolling :: String -> INotify -> FilePath -> (Result ConfigData -> IO ()) -> MVar ServerState -> IO ReloadModel

moveToNotify :: IO ReloadModel

Configuration loading

updateConfig :: FilePath -> (Result ConfigData -> IO ()) -> IO ()

safeUpdateConfig :: FilePath -> FStat -> (Result ConfigData -> IO ()) -> IO (FStat, ConfigReload)

Watcher threads

We have three threads/functions that can mutate the server state:

the long-interval watcher (onWatcherTimer)

the polling watcher (onPollTimer)

the inotify event handler (onInotify)

All of these will mutate the server state under modifyMVar or modifyMVar_, so that server transitions are more or less atomic. The inotify handler remains active during polling mode, but checks for polling mode and doesn't do anything in this case (this check is needed even if we would unregister the event handler due to how events are serialised).

onWatcherTimer :: IO Bool -> FilePath -> (Result ConfigData -> IO ()) -> MVar ServerState -> IO ()

onWatcherInner :: FilePath -> (Result ConfigData -> IO ()) -> ServerState -> IO ServerState

onPollTimer :: IO Bool -> FilePath -> (Result ConfigData -> IO ()) -> MVar ServerState -> IO ()

onPollInner :: IO Bool -> FilePath -> (Result ConfigData -> IO ()) -> ServerState -> IO (ServerState, Bool)

addNotifier :: INotify -> FilePath -> (Result ConfigData -> IO ()) -> MVar ServerState -> IO Bool

onInotify :: INotify -> String -> (Result ConfigData -> IO ()) -> MVar ServerState -> Event -> IO ()

initConfigReader :: (Result ConfigData -> a) -> IORef a -> IO ()