Ganeti.Curl.Internal

Description

Hsc2hs definitions for Storable interfaces.

Documentation

data CurlMsgCode

Data representing a CURLMSG enum.

Constructors

CurlMsgNone

CurlMsgDone

CurlMsgUnknown CInt
Haskell specific code for unknown codes

data CurlMsg

Data representing a struct CURLMsg.

Constructors

CurlMsg

Fields

cmMessage :: CurlMsgCode

The message type

cmHandle :: CurlH

The internal curl handle to which it applies

cmResult :: CurlCode

The message-specific result

errorBufferSize :: Int

Minimum buffer size for CurlErrorBuffer.

data CurlMCode

Multi interface error codes.

Constructors

CurlmCallMultiPerform

CurlmOK

CurlmBadHandle

CurlmBadEasyHandle

CurlmOutOfMemory

CurlmInternalError

CurlmBadSocket

CurlmUnknownOption

CurlmUnknown CInt
Haskell specific code denoting undefined codes (e.g. when libcurl has defined new codes that are not implemented yet)

toMsgCode :: CInt -> CurlMsgCode

Convert a CInt CURLMSG code (as returned by the C library) to a CurlMsgCode. When an unknown code is received, the special CurlMsgUnknown constructor will be used.

fromMsgCode :: CurlMsgCode -> CInt

Convert a CurlMsgCode to a CInt.

toMCode :: CInt -> CurlMCode

Convert a CInt CURLMcode (as returned by the C library) to a CurlMCode. When an unknown code is received, the special CurlmUnknown constructor will be used.