Ganeti.Daemon

Description

Implementation of the generic daemon functionality.

Constants

devNull :: FilePath

daemonStartupErr :: String -> String

Data types

data DaemonOptions

Command line options structure.

Constructors

DaemonOptions

Fields

optShowHelp :: Bool

Just show the help

optShowVer :: Bool

Just show the program version

optShowComp :: Bool

Just show the completion info

optDaemonize :: Bool

Whether to daemonize or not

optPort :: Maybe Word16

Override for the network port

optDebug :: Bool

Enable debug messages

optNoUserChecks :: Bool

Ignore user checks

optBindAddress :: Maybe String

Override for the bind address

optSyslogUsage :: Maybe SyslogUsage

Override for Syslog usage

optForceNode :: Bool

Ignore node checks

optNoVoting :: Bool

skip voting for master

optYesDoIt :: Bool

force dangerous options

defaultOptions :: DaemonOptions

Default values for the command line options.

type OptType = GenericOptType DaemonOptions

Abrreviation for the option type.

type CheckFn a = DaemonOptions -> IO (Either ExitCode a)

Check function type.

type PrepFn a b = DaemonOptions -> a -> IO b

Prepare function type.

type MainFn a b = DaemonOptions -> a -> b -> IO ()

Main execution function type.

Command line options

oNoDaemonize :: OptType

oDebug :: OptType

oNoUserChecks :: OptType

oPort :: Int -> OptType

oBindAddress :: OptType

oSyslogUsage :: OptType

oForceNode :: OptType

oNoVoting :: OptType

oYesDoIt :: OptType

genericOpts :: [OptType]

ioErrorToResult :: String -> IOError -> IO (Result a)

parseArgs :: String -> [OptType] -> IO (DaemonOptions, [String])

Small wrapper over getArgs and parseOpts.

Daemon-related functions

pidFileMode :: FileMode

pidFileFlags :: OpenFileFlags

writePidFile :: FilePath -> IO Fd

cleanupSocket :: FilePath -> IO ()

Helper function to ensure a socket doesn't exist. Should only be called once we have locked the pid file successfully.

setupDaemonEnv :: FilePath -> FileMode -> IO ()

finalCleanup :: FilePath -> IO ()

handleSigTerm :: ThreadId -> IO ()

handleSigHup :: FilePath -> IO ()

setupDaemonFDs :: Maybe FilePath -> IO ()

defaultBindAddr :: Int -> Family -> Result (Family, SockAddr)

parseAddress

Arguments

:: DaemonOptions
Command line options

-> Int
Default port for this daemon

-> IO (Result (Family, SockAddr))

Based on the options, compute the socket address to use for the daemon.

vClusterHostNameEnvVar :: String

getFQDN' :: Maybe AddrInfo -> IO String

getFQDNwithHints :: Maybe AddrInfo -> IO String

getFQDN :: IO String

Return the full qualified host name, honoring the vcluster setup.

isMaster :: IO Bool

ensureNode :: GanetiDaemon -> DaemonOptions -> IO ()

describeError :: String -> Maybe Handle -> Maybe FilePath -> IO a -> IO a

Run an I/O action that might throw an I/O error, under a handler that will simply annotate and re-throw the exception.

daemonize :: FilePath -> (Maybe Fd -> IO ()) -> IO ()

genericMain

Arguments

:: GanetiDaemon
The daemon we're running

-> [OptType]
The available options

-> CheckFn a
Check function

-> PrepFn a b
Prepare function

-> MainFn a b
Execution function

-> IO ()

Generic daemon startup.

fullPrep :: GanetiDaemon -> DaemonOptions -> SyslogUsage -> a -> PrepFn a b -> IO (FilePath, b)

innerMain :: GanetiDaemon -> DaemonOptions -> SyslogUsage -> a -> PrepFn a b -> MainFn a b -> Maybe Fd -> IO ()

handlePrepErr :: Bool -> Maybe Fd -> IOError -> IO a

maybeCloseFd :: Maybe Fd -> IO ()