Ganeti.DataCollectors.CPUload

Description

procstat data collector.

Documentation

defaultFile :: FilePath

bufferSize :: Int

windowSizeInUSec :: Integer

defaultCharNum :: Int

dcName :: String

The name of this data collector.

dcVersion :: DCVersion

The version of this data collector.

dcFormatVersion :: Int

The version number for the data format of this data collector.

dcCategory :: Maybe DCCategory

The category of this data collector.

dcKind :: DCKind

The kind of this data collector.

dcReport :: Maybe CollectorData -> IO DCReport

The data exported by the data collector, taken from the default location.

type Buffer = Seq (ClockTime, [Integer])

computeLoad :: CPUstat -> Integer

dcCollectFromFile :: FilePath -> IO (ClockTime, [Integer])

dcCollect :: IO Buffer

formatData :: [Double] -> CPUavgload

updateEntry :: Buffer -> Buffer -> Buffer

dcUpdate :: Maybe CollectorData -> IO CollectorData

Updates the given Collector data.

computeAverage :: Buffer -> Integer -> Integer -> Result [Double]

buildJsonReport :: Buffer -> IO JSValue

buildDCReport :: Buffer -> IO DCReport