Ganeti.DataCollectors.Drbd

Description

DRBD data collector.

Documentation

defaultFile :: FilePath

defaultCharNum :: Int

dcName :: String

The name of this data collector.

dcVersion :: DCVersion

The version of this data collector.

dcFormatVersion :: Int

The version number for the data format of this data collector.

dcCategory :: Maybe DCCategory

The category of this data collector.

dcKind :: DCKind

The kind of this data collector.

dcReport :: IO DCReport

The data exported by the data collector, taken from the default location.

Command line options

options :: IO [OptType]

arguments :: [ArgCompletion]

The list of arguments supported by the program.

getPairingInfo :: Maybe String -> IO (Result [DrbdInstMinor])

computeStatus :: DRBDStatus -> DCStatus

computeDevStatus :: DeviceInfo -> (DCStatusCode, String)

buildJsonReport :: FilePath -> Maybe FilePath -> IO JSValue

buildDCReport :: FilePath -> Maybe FilePath -> IO DCReport

main :: Options -> [String] -> IO ()

Main function.