Ganeti.DataCollectors.InstStatusTypes

Description

Type declarations specific for the instance status data collector.

Documentation

data InstStatus

Data type representing the status of an instance to be returned.

Constructors

InstStatus

Fields

iStatName :: String

iStatUuid :: String

iStatAdminState :: AdminState

iStatActualState :: ActualState

iStatUptime :: Maybe String

iStatMtime :: Double

iStatStateReason :: ReasonTrail

iStatStatus :: DCStatus

loadInstStatus :: JSValue -> Result InstStatus

saveInstStatus :: InstStatus -> JSValue

toDictInstStatus :: InstStatus -> [(String, JSValue)]

data ReportData

Constructors

ReportData

Fields

rDataInstances :: [InstStatus]

rDataStatus :: DCStatus

loadReportData :: JSValue -> Result ReportData

saveReportData :: ReportData -> JSValue

toDictReportData :: ReportData -> [(String, JSValue)]