Ganeti.DataCollectors.KvmRSS

Description

kvm resident set size collector

It collects the resident set size (RSS) for all kvm processes managed by Ganeti, i.e., the number of pages the process has in RAM. The value is obtained by taking the corresponding value from proc$pid/memstat.

Documentation

dcName :: String

The name of this data collector for the resident set size (RSS).

dcFormatVersion :: Int

The version number for the data format of this data collector.

dcVersion :: DCVersion

The version of this data collector.

dcCategory :: Maybe DCCategory

The category of this data collector.

dcKind :: DCKind

The kind of this data collector.

parsePid :: Monad m => String -> m Int

parseRss :: Monad m => String -> m Int

collectInstanceRSS :: String -> IO (Result (String, JSValue))

dcReport :: IO DCReport

The data exported by the data collector.