Ganeti.DataCollectors.Types

Description

Implementation of the Ganeti data collector types.

Documentation

data DCCategory

The possible classes a data collector can belong to.

Constructors

DCInstance

DCStorage

DCDaemon

DCHypervisor

getCategoryName :: DCCategory -> String

Get the category name and return it as a string.

categoryNames :: Map String DCCategory

data DCStatusCode

The possible status codes of a data collector.

Constructors

DCSCOk
Everything is OK

DCSCTempBad
Bad, but being automatically fixed

DCSCUnknown
Unable to determine the status

DCSCBad
Bad. External intervention required

data DCStatus

The status of a "status reporting data collector".

Constructors

DCStatus

Fields

dcStatusCode :: DCStatusCode

dcStatusMessage :: String

loadDCStatus :: JSValue -> Result DCStatus

saveDCStatus :: DCStatus -> JSValue

data DCKind

The type representing the kind of the collector.

Constructors

DCKPerf
Performance reporting collector

DCKStatus
Status reporting collector

data DCVersion

Type representing the version number of a data collector.

Constructors

DCVerBuiltin

DCVersion String

data CollectorData

Type for the value field of the CollectorMap below.

Constructors

CPULoadData (Seq (ClockTime, [Int]))

InstanceCpuLoad (Map String (Seq (ClockTime, Double)))

type CollectorMap = Map String CollectorData

Type for the map storing the data of the statefull DataCollectors.

data DCReport

This is the format of the report produced by each data collector.

Constructors

DCReport

Fields

dcReportName :: String

dcReportVersion :: DCVersion

dcReportFormatVersion :: Int

dcReportTimestamp :: Integer

dcReportCategory :: Maybe DCCategory

dcReportKind :: DCKind

dcReportData :: JSValue

loadDCReport :: JSValue -> Result DCReport

saveDCReport :: DCReport -> JSValue

addStatus :: DCStatus -> JSValue -> JSValue

Add the data collector status information to the JSON representation of the collector data.

mergeStatuses :: (DCStatusCode, String) -> (DCStatusCode, [String]) -> (DCStatusCode, [String])

Helper function for merging statuses.

buildReport :: String -> DCVersion -> Int -> Maybe DCCategory -> DCKind -> JSValue -> IO DCReport

Utility function for building a report automatically adding the current timestamp (rounded up to seconds). If the version is not specified, it will be set to the value indicating a builtin collector.

data ReportBuilder

A report of a data collector might be stateful or stateless.

Constructors

StatelessR (IO DCReport)

StatefulR (Maybe CollectorData -> IO DCReport)

type Name = String

data DataCollector

Type describing a data collector basic information

Constructors

DataCollector

Fields

dName :: Name

Name of the data collector

dCategory :: Maybe DCCategory

Category (storage, instance, ecc) of the collector

dKind :: DCKind

Kind (performance or status reporting) of the data collector

dReport :: ReportBuilder

Report produced by the collector

dUpdate :: Maybe (Maybe CollectorData -> IO CollectorData)

Update operation for stateful collectors.

dActive :: Name -> ConfigData -> Bool

Checks if the collector applies for the cluster.

dInterval :: Name -> ConfigData -> Integer

Interval between collection in microseconds