Ganeti.Errors

Description

Implementation of the Ganeti error types.

This module implements our error hierarchy. Currently we implement one identical to the Python one; later we might one to have separate ones for frontend (clients), master and backend code.

Documentation

data ErrorCode

Error code types for OpPrereqError.

Constructors

ECodeResolver

ECodeNoRes

ECodeTempNoRes

ECodeInval

ECodeState

ECodeNoEnt

ECodeExists

ECodeNotUnique

ECodeFault

ECodeEnviron

errorCodeFromRaw :: forall m. Monad m => String -> m ErrorCode

errorCodeToRaw :: ErrorCode -> String

data GanetiException

Constructors

GenericError String

LockError String

PidFileLockError String

HypervisorError String

ProgrammerError String

BlockDeviceError String

ConfigurationError String

ConfigVersionMismatch Int Int

ReservationError String

RemoteError String

SignatureError String

ParameterError String

ResultValidationError String

OpPrereqError String ErrorCode

OpExecError String

OpResultError String

OpCodeUnknown String

JobLost String

JobFileCorrupted String

ResolverError String Int String

HooksFailure String

HooksAbort [(String, String, String)]

UnitParseError String

ParseError String

TypeEnforcementError String

X509CertError String String

TagError String

CommandError String

StorageError String

InotifyError String

JobQueueError String

JobQueueDrainError String

JobQueueFull

ConfdMagicError String

ConfdClientError String

UdpDataSizeError String

NoCtypesError String

IPAddressError String

LuxiError String

QueryFilterParseError String

RapiTestResult String

FileStoragePathError String

saveGanetiException :: GanetiException -> JSValue

loadGanetiException :: JSValue -> Result GanetiException

type ErrorResult = GenericResult GanetiException

Error monad using GanetiException type alias.

excName :: GanetiException -> String

errorExitCode :: GanetiException -> ExitCode

Returns the exit code of a program that should be used if we got back an exception from masterd.

formatError :: GanetiException -> String

Formats an exception.

errToResult :: ErrorResult a -> Result a

Convert from an ErrorResult to a standard Result.