Ganeti.HTools.Backend.MonD

Description

Monitoring daemon backend

This module holds implements the querying of the monitoring daemons for dynamic utilisation data.

General definitions

data Report

The actual data types for MonD's Data Collectors.

Constructors

CPUavgloadReport CPUavgload

InstanceCpuReport (Map String Double)

InstanceRSSReport (Map String Double)

data DataCollector

Type describing a data collector basic information.

Constructors

DataCollector

Fields

dName :: String

Name of the data collector

dCategory :: Maybe DCCategory

dMkReport :: DCReport -> Maybe Report

dUse :: [(Node, Report)] -> (List, List) -> Result (List, List)

Node-total CPU load average data collector

mkCpuReport :: DCReport -> Maybe Report

updateNodeCpuFromReport :: (Node, Report) -> Node

updateCpuUtilDataFromNode :: List -> Node -> List

useNodeTotalCPU :: [(Node, Report)] -> (List, List) -> Result (List, List)

totalCPUCollector :: DataCollector

The node-total CPU collector.

Xen instance CPU-usage collector

mkXenCpuReport :: DCReport -> Maybe Report

useInstanceCpuData :: [(Node, Report)] -> (List, List) -> Result (List, List)

xenCPUCollector :: DataCollector

Collector for per-instance CPU data as observed by Xen

kvm instance RSS collector

mkKvmRSSReport :: DCReport -> Maybe Report

pagesPerMiB :: Double

useInstanceRSSData :: [(Node, Report)] -> (List, List) -> Result (List, List)

Update cluster data based on per-instance RSS data. Also set the node's memoy util pool correctly. Our unit of memory usage is pages; there are 256 pages per MiB of node memory not used by the node itself.

kvmRSSCollector :: DataCollector

Update cluster data based on the per-instance CPU usage

scaleMemoryWeight :: Double -> (List, List) -> (List, List)

Scale the importance of the memory weight in dynamic utilisation, by multiplying the usage with the given factor. Note that the underlying model for dynamic utilisation is that they are reported in arbitrary units.

Collector choice

collectors :: Options -> [DataCollector]

Querying infrastructure

mkReport :: DataCollector -> Maybe DCReport -> Maybe Report

Return the data from correct combination of a Data Collector and a DCReport.

type MonDData = (String, [DCReport])

type MapMonDData = Map String [DCReport]

fromFile :: DataCollector -> Node -> MapMonDData -> Maybe DCReport

getDCCName :: Maybe DCCategory -> String

prepareUrl :: DataCollector -> Node -> URLString

fromCurl :: DataCollector -> Node -> IO (Maybe DCReport)

Query a specified MonD for a Data Collector.

pMonDN :: JSRecord -> Result MonDData

pMonDData :: String -> Result [MonDData]

Parse MonD data file contents.

queryAMonD :: Maybe MapMonDData -> DataCollector -> Node -> IO (Maybe Report)

queryAllMonDs :: Maybe MapMonDData -> (List, List) -> DataCollector -> WriterT All IO (List, List)

queryAllMonDDCs :: ClusterData -> Options -> WriterT All IO ClusterData

Query all MonDs for all Data Collector. Return the cluster enriched by dynamic data, as well as a bit indicating wether all collectors could be queried successfully.