Ganeti.HTools.Backend.Simu

Description

Parsing data from a simulated description of the cluster.

This module holds the code for parsing a cluster description.

Documentation

apolAbbrev :: String -> Result AllocPolicy

parseDesc :: String -> [String] -> Result (AllocPolicy, Int, Int, Int, Int, Int)

createGroup :: Int -> String -> Result (Group, [Node])

parseData

Arguments

:: [String]
Cluster description in text format

-> Result ClusterData

Builds the cluster data from node/instance files.

loadData

Arguments

:: [String]
Cluster description in text format

-> IO (Result ClusterData)

Builds the cluster data from node/instance files.