Ganeti.HTools.Backend.Text

Description

Parsing data from text-files.

This module holds the code for loading the cluster state from text files, as produced by gnt-node and gnt-instance list command.

Helper functions

commaSplit :: String -> [String]

Serialisation functions

serializeGroup :: Group -> String

Serialize a single group.

serializeGroups :: List -> String

serializeNode

Arguments

:: List
The list of groups (needed for group uuid)

-> Node
The node to be serialised

-> String

Serialize a single node.

serializeNodes :: List -> List -> String

Generate node file data from node objects.

serializeInstance :: List -> Instance -> String

serializeInstances :: List -> List -> String

Generate instance file data from instance objects.

serializeISpec :: ISpec -> String

Generate a spec data from a given ISpec object.

serializeDiskTemplates :: [DiskTemplate] -> String

serializeIPolicy :: String -> IPolicy -> String

Generate policy data from a given policy object.

serializeAllIPolicies :: IPolicy -> List -> String

serializeCluster :: ClusterData -> String

Generate complete cluster data from node and instance lists.

Parsing functions

loadGroup :: Monad m => [String] -> m (String, Group)

loadNode

Arguments

:: Monad m

=> NameAssoc
Association list with current groups

-> [String]
Input data as a list of fields

-> m (String, Node)
The result, a tuple o node name and node object

Load a node from a field list.

loadInst

Arguments

:: NameAssoc
Association list with the current nodes

-> [String]
Input data as a list of fields

-> Result (String, Instance)
A tuple of instance name and the instance object

Load an instance from a field list.

loadISpec :: String -> [String] -> Result ISpec

Loads a spec from a field list.

loadIPolicy :: [String] -> Result (String, IPolicy)

Loads an ipolicy from a field list.

loadOnePolicy :: (IPolicy, List) -> String -> Result (IPolicy, List)

loadAllIPolicies :: List -> [String] -> Result (IPolicy, List)

loadTabular :: (Monad m, Element a) => [String] -> ([String] -> m (String, a)) -> m (NameAssoc, Container a)

readData :: String -> IO String

parseData

Arguments

:: String
Text data

-> Result ClusterData

Builds the cluster data from text input.

loadData

Arguments

:: String
Path to the text file

-> IO (Result ClusterData)

Top level function for data loading.