Ganeti.HTools.Cluster.AllocatePrimitives

Description

Implementation of the primitives of instance allocation

Documentation

allocateOnSingle :: AlgorithmOptions -> List -> Instance -> Ndx -> OpResult AllocElement

Tries to allocate an instance on one given node.

allocateOnPair :: AlgorithmOptions -> ClusterStatistics -> List -> Instance -> Ndx -> Ndx -> OpResult AllocElement

Tries to allocate an instance on a given pair of nodes.