Ganeti.HTools.Cluster.Metrics

Description

Implementation of the cluster metric

Documentation

reservedMemRtotalCoeff :: Double

detailedCVInfoExt :: [((Double, String), ([Double] -> Statistics, Bool))]

optimalCVScore :: List -> Double

Compute the lower bound of the cluster score, i.e., the sum of the minimal values for all cluster score values that are not 0 on a perfectly balanced cluster.

detailedCVInfo :: [(Double, String)]

detailedCVWeights :: [Double]

detailedCVAggregation :: [([Double] -> Statistics, Bool)]

detailedCVOnlineStatus :: [Bool]

compDetailedCVNode :: Node -> [Double]

compClusterStatistics :: [Node] -> [Statistics]

Compute the statistics of a cluster.

updateClusterStatistics :: [Statistics] -> (Node, Node) -> [Statistics]

updateClusterStatisticsTwice :: [Statistics] -> (Node, Node) -> (Node, Node) -> [Statistics]

Update a cluster statistics twice.

compDetailedCV :: [Node] -> [Double]

compCVfromStats :: [Statistics] -> Double

Compute the cluster score from its statistics

compCVNodes :: [Node] -> Double

Compute the total variance.

compCV :: List -> Double

Wrapper over compCVNodes for callers that have a List.

printStats :: String -> List -> String

Shows statistics for a given node list.