Ganeti.HTools.Cluster.Moves

Description

Implementation of instance moves in a cluster.

Documentation

instanceNodes :: List -> Instance -> (Ndx, Ndx, Node, Node)

setInstanceLocationScore

Arguments

:: Instance
the original instance

-> Node
the primary node of the instance

-> Node
the secondary node of the instance

-> Instance
the instance with the location score updated

Sets the location score of an instance, given its primary and secondary node.

applyMoveEx

Arguments

:: Bool
whether to ignore soft errors

-> List

-> Instance

-> IMove

-> OpResult (List, Instance, Ndx, Ndx)

Applies an instance move to a given node list and instance.

move :: (List, List) -> (Idx, IMove) -> OpResult (List, List)

Apply a move to an instance, ignoring soft errors. This is a variant of `applyMoveEx True` suitable for folding.