Ganeti.HTools.ExtLoader

Description

External data loader.

This module holds the external data loading, and thus is the only one depending (via the specialized Text/Rapi/Luxi modules) on the actual libraries implementing the low-level protocols.

Documentation

wrapIO :: IO (Result a) -> IO (Result a)

parseUtilisation :: String -> Result (String, DynUtil)

loadExternalData :: Options -> IO ClusterData

External tool data loader from a variety of sources.

maybeSaveData

Arguments

:: Maybe FilePath
The file prefix to save to

-> String
The suffix (extension) to add

-> String
Informational message

-> ClusterData
The cluster data

-> IO ()

Function to save the cluster data to a file.

data DataCollector

Constructors

DataCollector

Fields

dName :: String

dCategory :: Maybe DCCategory

data Report

Constructors

CPUavgloadReport CPUavgload

collectors :: Options -> [DataCollector]

type MonDData = (String, [DCReport])

type MapMonDData = Map String [DCReport]

pMonDData :: String -> Result [MonDData]

Parse MonD data file contents.

pMonDN :: JSRecord -> Result MonDData

queryAllMonDDCs :: ClusterData -> Options -> IO ClusterData

Query all MonDs for all Data Collector.

queryAllMonDs :: Maybe MapMonDData -> (List, List) -> DataCollector -> IO (List, List)

fromCurl :: DataCollector -> Node -> IO (Maybe DCReport)

mkReport :: DataCollector -> Maybe DCReport -> Maybe Report

fromFile :: DataCollector -> Node -> MapMonDData -> Maybe DCReport

queryAMonD :: Maybe MapMonDData -> DataCollector -> Node -> IO (Maybe Node)

updateUtilData :: List -> Node -> List

prepareUrl :: DataCollector -> Node -> URLString

getDCCName :: Maybe DCCategory -> String