Ganeti.HTools.Group

Description

Module describing a node group.

Type declarations

data Group

The node group type.

Constructors

Group

Fields

name :: String

The node name

uuid :: GroupID

The UUID of the group

idx :: Gdx

Internal index for book-keeping

allocPolicy :: AllocPolicy

The allocation policy for this group

iPolicy :: IPolicy

The instance policy for this group

allTags :: [String]

The tags for this group

type AssocList = [(Gdx, Group)]

A simple name for the int, node association list.

type List = Container Group

A simple name for a node map.

Initialization functions

create :: String -> GroupID -> AllocPolicy -> IPolicy -> [String] -> Group

Create a new group.

setIdx :: Group -> Gdx -> Group

Sets the group index.

This is used only during the building of the data structures.

setName :: Group -> String -> Group

isAllocable :: Group -> Bool

Checks if a group is allocable.