Ganeti.HTools.Instance

Description

Module describing an instance.

The instance data type holds very few fields, the algorithm intelligence is in the Node and Cluster modules.

Type declarations

data Disk

Constructors

Disk

Fields

dskSize :: Int

Size in bytes

dskSpindles :: Maybe Int

Number of spindles

data Instance

The instance type.

Constructors

Instance

Fields

name :: String

The instance name

alias :: String

The shortened name

mem :: Int

Memory of the instance

dsk :: Int

Total disk usage of the instance

disks :: [Disk]

Sizes of the individual disks

vcpus :: Int

Number of VCPUs

runSt :: InstanceStatus

Original run status

pNode :: Ndx

Original primary node

sNode :: Ndx

Original secondary node

idx :: Idx

Internal index

util :: DynUtil

Dynamic resource usage

movable :: Bool

Can and should the instance be moved?

autoBalance :: Bool

Is the instance auto-balanced?

diskTemplate :: DiskTemplate

The disk template of the instance

spindleUse :: Int

The numbers of used spindles

allTags :: [String]

List of all instance tags

exclTags :: [String]

List of instance exclusion tags

locationScore :: Int

The number of common-failures between primary and secondary node of the instance

arPolicy :: AutoRepairPolicy

Instance's auto-repair policy

nics :: [Nic]

NICs of the instance

forthcoming :: Bool

Is the instance is forthcoming?

isRunning :: Instance -> Bool

Check if instance is running.

isOffline :: Instance -> Bool

Check if instance is offline.

notOffline :: Instance -> Bool

Helper to check if the instance is not offline.

instanceDown :: Instance -> Bool

Check if instance is down.

applyIfOnline :: Instance -> (a -> a) -> a -> a

Apply the function if the instance is online. Otherwise use the initial value

usesSecMem :: Instance -> Bool

Helper for determining whether an instance's memory needs to be taken into account for secondary memory reservation.

localStorageTemplates :: [DiskTemplate]

Constant holding the local storage templates.

Note: Currently Ganeti only exports node total/free disk space for LVM-based storage; file-based storage is ignored in this model, so even though file-based storage uses in reality disk space on the node, in our model it won't affect it and we can't compute whether there is enough disk space for a file-based instance. Therefore we will treat this template as 'foreign' storage.

movableDiskTemplates :: [DiskTemplate]

type AssocList = [(Idx, Instance)]

A simple name for the int, instance association list.

type List = Container Instance

A simple name for an instance map.

Initialization

create :: String -> Int -> Int -> [Disk] -> Int -> InstanceStatus -> [String] -> Bool -> Ndx -> Ndx -> DiskTemplate -> Int -> [Nic] -> Bool -> Instance

Create an instance.

Some parameters are not initialized by function, and must be set later (via setIdx for example).

setIdx

Arguments

:: Instance
The original instance

-> Idx
New index

-> Instance
The modified instance

Changes the index.

This is used only during the building of the data structures.

setName

Arguments

:: Instance
The original instance

-> String
New name

-> Instance
The modified instance

Changes the name.

This is used only during the building of the data structures.

setAlias

Arguments

:: Instance
The original instance

-> String
New alias

-> Instance
The modified instance

Changes the alias.

This is used only during the building of the data structures.

Update functions

setPri

Arguments

:: Instance
the original instance

-> Ndx
the new primary node

-> Instance
the modified instance

Changes the primary node of the instance.

setSec

Arguments

:: Instance
the original instance

-> Ndx
the new secondary node

-> Instance
the modified instance

Changes the secondary node of the instance.

setBoth

Arguments

:: Instance
the original instance

-> Ndx
new primary node index

-> Ndx
new secondary node index

-> Instance
the modified instance

Changes both nodes of the instance.

setMovable

Arguments

:: Instance
The original instance

-> Bool
New movable flag

-> Instance
The modified instance

Sets the movable flag on an instance.

shrinkByType :: Instance -> FailMode -> Result Instance

Try to shrink the instance based on the reason why we can't allocate it.

getTotalSpindles :: Instance -> Maybe Int

Get the number of disk spindles

specOf :: Instance -> RSpec

Return the spec of an instance.

instCompareISpec :: Ordering -> Instance -> ISpec -> Bool -> OpResult ()

instBelowISpec :: Instance -> ISpec -> Bool -> OpResult ()

Checks if an instance is smaller than a given spec.

instAboveISpec :: Instance -> ISpec -> Bool -> OpResult ()

Checks if an instance is bigger than a given spec.

instMatchesMinMaxSpecs :: Instance -> MinMaxISpecs -> Bool -> OpResult ()

instMatchesSpecs :: Instance -> [MinMaxISpecs] -> Bool -> OpResult ()

instMatchesPolicy :: Instance -> IPolicy -> Bool -> OpResult ()

Checks if an instance matches a policy.

hasSecondary :: Instance -> Bool

Checks whether the instance uses a secondary node.

Note: This should be reconciled with sNode == noSecondary.

requiredNodes :: DiskTemplate -> Int

Computed the number of nodes for a given disk template.

allNodes :: Instance -> [Ndx]

Computes all nodes of an instance.

usesLocalStorage :: Instance -> Bool

Checks whether a given disk template uses local storage.

supportsMoves :: DiskTemplate -> Bool

mirrorType :: Instance -> MirrorType

A simple wrapper over templateMirrorType.

usesMemory :: Instance -> Bool

Whether the instance uses memory on its host node. Depends on the InstanceStatus and on whether the instance is forthcoming; instances that aren't running or existent don't use memory.