Ganeti.HTools.Loader

Description

Generic data loader.

This module holds the common code for parsing the input data after it has been loaded from external sources.

Types

data RqType

The iallocator request type.

This type denotes what request we got from Ganeti and also holds request-specific fields.

Constructors

Allocate Instance AllocDetails
A new instance allocation

Relocate Idx Int [Ndx]
Choose a new secondary node

NodeEvacuate [Idx] EvacMode
node-evacuate mode

ChangeGroup [Gdx] [Idx]
Multi-relocate mode

MultiAllocate [(Instance, AllocDetails)]
Multi-allocate mode

data Request

A complete request, as received from Ganeti.

Constructors

Request RqType ClusterData

data ClusterData

The cluster state.

Constructors

ClusterData

Fields

cdGroups :: List

The node group list

cdNodes :: List

The node list

cdInstances :: List

The instance list

cdTags :: [String]

The cluster tags

cdIPolicy :: IPolicy

The cluster instance policy

emptyCluster :: ClusterData

An empty cluster.

Functions

lookupNode :: Monad m => NameAssoc -> String -> String -> m Ndx

Lookups a node into an assoc list.

lookupInstance :: Monad m => NameAssoc -> String -> m Idx

Lookups an instance into an assoc list.

lookupGroup :: Monad m => NameAssoc -> String -> String -> m Gdx

Lookups a group into an assoc list.

assignIndices :: Element a => [(String, a)] -> (NameAssoc, Container a)

Given a list of elements (and their names), assign indices to them.

setMaster :: Monad m => NameAssoc -> List -> String -> m List

Given am indexed node list, and the name of the master, mark it as such.

setLocationScore :: List -> Instance -> Instance

fixNodes :: List -> Instance -> List

setNodePolicy :: List -> Node -> Node

updateExclTags :: [String] -> Instance -> Instance

Update instance with exclusion tags list.

updateMovable :: [String] -> [String] -> Instance -> Instance

disableSplitMoves :: List -> Instance -> Instance

setArPolicy :: [String] -> List -> List -> List -> ClockTime -> List

getArPolicy :: [String] -> ClockTime -> Maybe AutoRepairPolicy

longestDomain :: [String] -> String

extractExTags :: [String] -> [String]

Extracts the exclusion tags from the cluster configuration.

commonSuffix :: List -> List -> String

Extracts the common suffix from node/instance names.

addMigrationTags :: [String] -> Node -> Node

addLocationTags :: [String] -> Node -> Node

mergeData

Arguments

:: [(String, DynUtil)]
Instance utilisation data

-> [String]
Exclusion tags

-> [String]
Selected instances (if not empty)

-> [String]
Excluded instances

-> ClockTime
The current timestamp

-> ClusterData
Data from backends

-> Result ClusterData
Fixed cluster data

Initializer function that loads the data from a node and instance list and massages it into the correct format.

clearDynU :: ClusterData -> Result ClusterData

In a cluster description, clear dynamic utilisation information.

checkData :: List -> List -> ([String], List)

Checks the cluster data for consistency.

nodeImem :: Node -> List -> Int

nodeIdsk :: Node -> List -> Int

eitherLive :: Monad m => Bool -> a -> m a -> m a

Get live information or a default value