Ganeti.HTools.Tags

Description

Tags

This module holds all the tag interpretation done by htools.

Constants

exTagsPrefix :: String

The exclusion tag prefix. Instance tags starting with this prefix describe a service provided by the instance. Instances providing the same service at not places on the same node.

standbyPrefix :: String

migrationPrefix :: String

allowMigrationPrefix :: String

standbyAuto :: String

The tag to be added to nodes that were shutdown by hsqueeze.

autoRepairTagPrefix :: String

Auto-repair tag prefix

autoRepairTagEnabled :: String

autoRepairTagPending :: String

autoRepairTagResult :: String

autoRepairTagSuspended :: String

Predicates

hasStandbyTag :: Node -> Bool

Predicate of having a standby tag.

Migration restriction tags

getMigRestrictionsList :: [String] -> [String] -> [String]

getMigRestrictions :: [String] -> [String] -> Set String

Given the cluster tags extract the migration restrictions from a node tag.

splitAtColons :: String -> Maybe (String, String)

migrations :: [String] -> [(String, String)]

getRecvMigRestrictions :: [String] -> [String] -> Set String

Given the cluster tags, extract the set of migration restrictions a node is able to receive from its node tags.