Ganeti.HTools.Tags.Constants

Description

Tag constants

This module holds all the special tag prefixes honored by Ganeti's htools. The module itself does not depend on anything Ganeti specific so that it can be imported anywhere.

Documentation

exTagsPrefix :: String

The exclusion tag prefix. Instance tags starting with this prefix describe a service provided by the instance. Instances providing the same service at not places on the same node.

standbyPrefix :: String

The tag-prefix indicating that hsqueeze should consider a node as being standby.

migrationPrefix :: String

The prefix for migration tags

allowMigrationPrefix :: String

Prefix of tags allowing migration

locationPrefix :: String

The prefix for node location tags.

desiredLocationPrefix :: String

The prefix for instance desired location tags.

standbyAuto :: String

The tag to be added to nodes that were shutdown by hsqueeze.

autoRepairTagPrefix :: String

Auto-repair tag prefix

autoRepairTagEnabled :: String

autoRepairTagPending :: String

autoRepairTagResult :: String

autoRepairTagSuspended :: String