Ganeti.JQScheduler.Filtering

Description

Filtering of jobs for the Ganeti job queue.

Documentation

accessOpCodeField :: OpCode -> String -> ErrorResult JSValue

opCodesOf :: QueuedJob -> [OpCode]

reasonsOf :: QueuedJob -> [ReasonElem]

evaluateFilterComparator :: Ord field => Filter field -> (Comparator -> field -> FilterValue -> Maybe Bool) -> Bool

matchPredicate

Arguments

:: QueuedJob

-> JobId
the watermark to compare against if the predicate references it

-> FilterPredicate

-> Bool

Whether a FilterPredicate is true for a job.

matches :: QueuedJob -> FilterRule -> Bool

Whether all predicates of the filter rule are true for the job.

orderFilters :: Set FilterRule -> [FilterRule]

applyingFilter :: Set FilterRule -> QueuedJob -> Maybe FilterRule

Finds the first filter whose predicates all match the job and whose action is not Continue. This is the applying filter.

type RateLimitSlotMap = SlotMap ByteString

data FilterChainState

Constructors

FilterChainState

Fields

rateLimitSlotMap :: RateLimitSlotMap

tryFitSlots :: FilterChainState -> CountMap ByteString -> Maybe FilterChainState

queueRateLimitSlotMap :: Queue -> Set FilterRule -> RateLimitSlotMap

jobFiltering :: Queue -> Set FilterRule -> [JobWithStat] -> [JobWithStat]

Implements job filtering as specified in `doc/design-optables.rst`.

Importantly, the filter that *applies* is the first one of which all predicates match; this is implemented in applyingFilter.

The initial FilterChainState is currently not cached across selectJobsToRun invocations because the number of running jobs is typically small (< 100).