Ganeti.JQueue

Description

Implementation of the job queue.

Data types

noTimestamp :: Timestamp

Missing timestamp type.

fromClockTime :: ClockTime -> Timestamp

Obtain a Timestamp from a given clock time

currentTimestamp :: IO Timestamp

Get the current time in the job-queue timestamp format.

advanceTimestamp :: Int -> Timestamp -> Timestamp

From a given timestamp, obtain the timestamp of the time that is the given number of seconds later.

toMetaOpCode :: InputOpCode -> [MetaOpCode]

invalidOp :: String

extractOpSummary :: InputOpCode -> String

Tries to extract the opcode summary from an InputOpCode. This duplicates some functionality from the opSummary function in Ganeti.OpCodes.

queuedOpCodeFromMetaOpCode :: MetaOpCode -> QueuedOpCode

Convenience function to obtain a QueuedOpCode from a MetaOpCode

queuedJobFromOpCodes :: Monad m => JobId -> [MetaOpCode] -> m QueuedJob

From a job-id and a list of op-codes create a job. This is the pure part of job creation, as allocating a new job id lives in IO.

setReceivedTimestamp :: Timestamp -> QueuedJob -> QueuedJob

Attach a received timestamp to a Queued Job.

reasonTrailTimestamp :: Timestamp -> Integer

Build a timestamp in the format expected by the reason trail (nanoseconds) starting from a JQueue Timestamp.

extendInputOpCodeReasonTrail :: JobId -> Timestamp -> Int -> InputOpCode -> InputOpCode

extendOpCodeReasonTrail :: JobId -> Timestamp -> Int -> QueuedOpCode -> QueuedOpCode

extendJobReasonTrail :: QueuedJob -> QueuedJob

Append an element to the reason trail of all the OpCodes of a queued job.

getJobDependencies :: QueuedJob -> [JobId]

From a queued job obtain the list of jobs it depends on.

changeOpCodePriority :: Int -> QueuedOpCode -> QueuedOpCode

Change the priority of a QueuedOpCode, if it is not already finalized.

cancelOpCode :: Timestamp -> QueuedOpCode -> QueuedOpCode

changeJobPriority :: Int -> QueuedJob -> QueuedJob

Change the priority of a job, i.e., change the priority of the non-finalized opcodes.

cancelQueuedJob :: Timestamp -> QueuedJob -> QueuedJob

Transform a QueuedJob that has not been started into its canceled form.

failOpCode :: ReasonElem -> Timestamp -> QueuedOpCode -> QueuedOpCode

failQueuedJob :: ReasonElem -> Timestamp -> QueuedJob -> QueuedJob

Transform a QueuedJob that has not been started into its failed form.

jobFilePrefix :: String

jobFileName :: JobId -> FilePath

Computes the filename for a given job ID.

parseJobFileId :: Monad m => FilePath -> m JobId

liveJobFile :: FilePath -> JobId -> FilePath

Computes the full path to a live job.

archivedJobFile :: FilePath -> JobId -> FilePath

Computes the full path to an archives job. BROKEN.

opStatusToJob :: OpStatus -> JobStatus

calcJobStatus :: QueuedJob -> JobStatus

Computes a queued job's status.

jobStarted :: QueuedJob -> Bool

Determine if a job has started

jobFinalized :: QueuedJob -> Bool

Determine if a job is finalised.

jobArchivable :: Timestamp -> QueuedJob -> Bool

Determine if a job is finalized and its timestamp is before a given time.

opStatusFinalized :: OpStatus -> Bool

Determine whether an opcode status is finalized.

calcJobPriority :: QueuedJob -> Int

Compute a job's priority.

ignoreIOError :: a -> Bool -> String -> IOError -> IO a

allArchiveDirs :: FilePath -> IO [FilePath]

determineJobDirectories :: FilePath -> Bool -> IO [FilePath]

Build list of directories containing job files. Note: compared to the Python version, this doesn't ignore a potential lost+found file.

getJobIDs :: [FilePath] -> IO (GenericResult IOError [JobId])

Computes the list of all jobs in the given directories.

sortJobIDs :: [JobId] -> [JobId]

Sorts the a list of job IDs.

getDirJobIDs :: FilePath -> ResultT IOError IO [JobId]

readJobDataFromDisk :: FilePath -> Bool -> JobId -> IO (Maybe (String, Bool))

noSuchJob :: Result (QueuedJob, Bool)

Failed to load job error.

loadJobFromDisk :: FilePath -> Bool -> JobId -> IO (Result (QueuedJob, Bool))

Loads a job from disk.

writeJobToDisk :: FilePath -> QueuedJob -> IO (Result ())

Write a job to disk.

replicateJob :: FilePath -> [Node] -> QueuedJob -> IO [(Node, ERpcError ())]

replicateManyJobs :: FilePath -> [Node] -> [QueuedJob] -> IO ()

Replicate many jobs to all master candidates.

writeAndReplicateJob :: FromString e => ConfigData -> FilePath -> QueuedJob -> ResultT e IO [(Node, ERpcError ())]

Writes a job to a file and replicates it to master candidates.

readSerialFromDisk :: IO (Result JobId)

Read the job serial number from disk.

allocateJobIds :: [Node] -> Lock -> Int -> IO (Result [JobId])

Allocate new job ids. To avoid races while accessing the serial file, the threads synchronize over a lock, as usual provided by a Lock.

allocateJobId :: [Node] -> Lock -> IO (Result JobId)

Allocate one new job id.

isQueueOpen :: IO Bool

Decide if job queue is open

startJobs

Arguments

:: Livelock
Luxi's livelock path

-> Lock
lock for forking new processes

-> [QueuedJob]
the list of jobs to start

-> IO [ErrorResult QueuedJob]

Start enqueued jobs by executing the Python code.

isQueuedJobDead :: MonadIO m => Livelock -> QueuedJob -> m Bool

waitForJob :: JobId -> Int -> ResultG (Bool, String)

cancelJob

Arguments

:: Bool
if True, use sigKILL instead of sigTERM

-> Livelock
Luxi's livelock path

-> JobId
the job to cancel

-> IO (ErrorResult (Bool, String))

Try to cancel a job that has already been handed over to execution, by terminating the process.

tellJobPriority

Arguments

:: Livelock
Luxi's livelock path

-> JobId
the job to inform

-> Int
the new priority

-> IO (ErrorResult (Bool, String))

Inform a job that it is requested to change its priority. This is done by writing the new priority to a file and sending SIGUSR1.

notifyJob :: ProcessID -> IO (ErrorResult ())

Notify a job that something relevant happened, e.g., a lock became available. We do this by sending sigHUP to the process.

queueDirPermissions :: FilePermissions

Permissions for the archive directories.

archiveSomeJobsUntil :: ([JobId] -> IO ()) -> FilePath -> ClockTime -> Timestamp -> Int -> [JobId] -> [JobId] -> IO (Int, Int)

archiveJobs

Arguments

:: ConfigData
cluster configuration

-> Int
time the job has to be in the past in order to be archived

-> Int
timeout

-> [JobId]
jobs to consider

-> IO (Int, Int)

Archive jobs older than the given time, but do not exceed the timeout for carrying out this task.