Ganeti.JQueue.Objects

Description

Objects in the job queue.

Documentation

type Timestamp = (Int, Int)

The ganeti queue timestamp type. It represents the time as the pair of seconds since the epoch and microseconds since the beginning of the second.

data InputOpCode

An input opcode.

Constructors

ValidOpCode MetaOpCode
OpCode was parsed successfully

InvalidOpCode JSValue
Invalid opcode

data QueuedOpCode

Constructors

QueuedOpCode

Fields

qoInput :: InputOpCode

qoStatus :: OpStatus

qoResult :: JSValue

qoLog :: [(Int, Timestamp, ELogType, JSValue)]

qoPriority :: Int

qoStartTimestamp :: Maybe Timestamp

qoExecTimestamp :: Maybe Timestamp

qoEndTimestamp :: Maybe Timestamp

loadQueuedOpCode :: JSValue -> Result QueuedOpCode

saveQueuedOpCode :: QueuedOpCode -> JSValue

data QueuedJob

Constructors

QueuedJob

Fields

qjId :: JobId

qjOps :: [QueuedOpCode]

qjReceivedTimestamp :: Maybe Timestamp

qjStartTimestamp :: Maybe Timestamp

qjEndTimestamp :: Maybe Timestamp

qjLivelock :: Maybe FilePath

qjProcessId :: Maybe ProcessID

loadQueuedJob :: JSValue -> Result QueuedJob

saveQueuedJob :: QueuedJob -> JSValue