Ganeti.JSON

Description

JSON utility functions.

JSON-related functions

type JSField = (String, JSValue)

type JSRecord = [JSField]

A type alias for the list-based representation of J.JSObject.

readJSONWithDesc

Arguments

:: JSON a

=> String
description of a

-> Bool
include input in error messages

-> JSValue
input value

-> Result a

Annotate readJSON error messages with descriptions of what is being parsed into what.

fromJResult :: Monad m => String -> Result a -> m a

Converts a JSON Result into a monadic value.

fromJResultE :: (Error e, MonadError e m) => String -> Result a -> m a

Converts a JSON Result into a MonadError value.

readEitherString :: Monad m => JSValue -> m String

Tries to read a string from a JSON value.

In case the value was not a string, we fail the read (in the context of the current monad.

loadJSArray

Arguments

:: Monad m

=> String
Operation description (for error reporting)

-> String
Input message

-> m [JSObject JSValue]

Converts a JSON message into an array of JSON objects.

buildNoKeyError :: JSRecord -> String -> String

fromObj :: (JSON a, Monad m) => JSRecord -> String -> m a

Reads the value of a key in a JSON object.

maybeFromObj :: (JSON a, Monad m) => JSRecord -> String -> m (Maybe a)

Reads the value of an optional key in a JSON object. Missing keys, or keys that have a 'null' value, will be returned as Nothing, otherwise we attempt deserialisation and return a Just value.

fromObjWithDefault :: (JSON a, Monad m) => JSRecord -> String -> a -> m a

Reads the value of a key in a JSON object with a default if missing. Note that both missing keys and keys with value 'null' will cause the default value to be returned.

arrayMaybeFromJVal :: (JSON a, Monad m) => JSValue -> m [Maybe a]

arrayMaybeFromObj :: (JSON a, Monad m) => JSRecord -> String -> m [Maybe a]

tryArrayMaybeFromObj

Arguments

:: JSON a

=> String
Textual owner in error messages

-> JSRecord
The object array

-> String
The desired key from the object

-> Result [Maybe a]

Wrapper for arrayMaybeFromObj with better diagnostic

fromKeyValue

Arguments

:: (JSON a, Monad m)

=> String
The key name

-> JSValue
The value to read

-> m a

Reads a JValue, that originated from an object key.

fromJVal :: (Monad m, JSON a) => JSValue -> m a

Small wrapper over readJSON.

fromJValE :: (Error e, MonadError e m, JSON a) => JSValue -> m a

Small wrapper over readJSON for MonadError.

jsonHead :: JSON b => [a] -> (a -> b) -> JSValue

Helper function that returns Null or first element of the list.

getMaybeJsonHead :: JSON b => [a] -> (a -> Maybe b) -> JSValue

Helper for extracting Maybe values from a possibly empty list.

getMaybeJsonElem :: JSON b => [a] -> Int -> (a -> Maybe b) -> JSValue

Helper for extracting Maybe values from a list that might be too short.

asJSObject :: Monad m => JSValue -> m (JSObject JSValue)

Converts a JSON value into a JSON object.

asObjectList :: Monad m => [JSValue] -> m [JSObject JSValue]

Coneverts a list of JSON values into a list of JSON objects.

tryFromObj

Arguments

:: JSON a

=> String
Textual owner in error messages

-> JSRecord
The object array

-> String
The desired key from the object

-> Result a

Try to extract a key from an object with better error reporting than fromObj.

toArray :: Monad m => JSValue -> m [JSValue]

Ensure a given JSValue is actually a JSArray.

optionalJSField :: JSON a => String -> Maybe a -> Maybe JSField

Creates a Maybe JSField. If the value string is Nothing, the JSField will be Nothing as well.

optFieldsToObj :: [Maybe JSField] -> JSValue

Creates an object with all the non-Nothing fields of the given list.

Container type (special type for JSON serialisation)

class HasStringRepr a where

Class of types that can be converted from Strings. This is similar to the Read class, but it's using a different serialisation format, so we have to define a separate class. Mostly useful for custom key types in JSON dictionaries, which have to be backed by strings.

Methods

fromStringRepr :: Monad m => String -> m a

toStringRepr :: a -> String

newtype GenericContainer a b

The container type, a wrapper over Data.Map

Constructors

GenericContainer

Fields

fromContainer :: Map a b

emptyContainer :: GenericContainer a b

The empty container.

type Container = GenericContainer String

Type alias for string keys.

containerFromList :: Ord a => [(a, b)] -> GenericContainer a b

Creates a GenericContainer from a list of key-value pairs.

lookupContainer :: (Monad m, Ord a) => m b -> a -> GenericContainer a b -> m b

Looks up a value in a container with a default value. If a key has no value, a given monadic default is returned. This allows simple error handling, as the default can be mzero, failError etc.

alterContainerL :: (Functor f, Ord a) => a -> (Maybe b -> f (Maybe b)) -> GenericContainer a b -> f (GenericContainer a b)

Updates a value inside a container. The signature of the function is crafted so that it can be directly used as a lens.

readContainer :: (Monad m, HasStringRepr a, Ord a, JSON b) => JSObject JSValue -> m (GenericContainer a b)

Container loader.

showContainer :: (HasStringRepr a, JSON b) => GenericContainer a b -> JSValue

Types that (de)serialize in a special form of JSON

newtype UsedKeys

Constructors

UsedKeys (Maybe (Set String))

mkUsedKeys :: Set String -> UsedKeys

allUsedKeys :: UsedKeys

class DictObject a where

Class of objects that can be converted from and to JSObject lists-format.

Methods

toDict :: a -> [(String, JSValue)]

fromDictWKeys :: [(String, JSValue)] -> WriterT UsedKeys Result a

fromDict :: [(String, JSValue)] -> Result a

showJSONtoDict :: DictObject a => a -> JSValue

A default implementation of showJSON using toDict.

readJSONfromDict :: DictObject a => JSValue -> Result a

A default implementation of readJSON using fromDict. Checks that the input value is a JSON object and converts it using fromDict. Also checks the input contains only the used keys returned by fromDict.

class ArrayObject a where

Class of objects that can be converted from and to [JSValue] with a fixed length and order.

Methods

toJSArray :: a -> [JSValue]

fromJSArray :: [JSValue] -> Result a

General purpose data types for working with JSON

newtype MaybeForJSON a

A Maybe newtype that allows for serialization more appropriate to the semantics of Maybe and JSON in our calls. Does not produce needless and confusing dictionaries.

In particular, JSNull corresponds to Nothing. This also means that this `Maybe a` newtype should not be used with a values that themselves can serialize to null.

Constructors

MaybeForJSON

Fields

unMaybeForJSON :: Maybe a

newtype TimeAsDoubleJSON

Constructors

TimeAsDoubleJSON

Fields

unTimeAsDoubleJSON :: ClockTime

newtype Tuple5 a b c d e

Constructors

Tuple5

Fields

unTuple5 :: (a, b, c, d, e)

nestedAccessByKey :: [String] -> JSValue -> Result JSValue

Look up a value in a JSON object. Accessing [a, b, c] on an object is equivalent as accessing myobject.a.b.c on a JavaScript object.

An error is returned if the object doesn't have such an accessor or if any value during the nested access is not an object at all.

nestedAccessByKeyDotted :: String -> JSValue -> Result JSValue

Same as nestedAccessByKey, but accessing with a dotted string instead (like nestedAccessByKeyDotted a.b.c).

branchOnField

Arguments

:: String
fieldname to branch on

-> (JSValue -> Result a)
decoding function if field is present and true; field will already be removed in the input

-> (JSValue -> Result a)
decoding function otherwise

-> JSValue

-> Result a

Branch decoding on a field in a JSON object.

addField :: (String, JSValue) -> JSValue -> JSValue

Add a field to a JSON object; to nothing, if the argument is not an object.