Ganeti.Lens

Description

Provides all lens-related functions.

Documentation

lensWith :: (s -> a) -> (s -> a -> b -> t) -> Lens s t a b

Creates an optimized lens where the setter also gets the original value from the getter.

lensFieldName :: String -> String

makeCustomLensesFiltered :: (String -> Bool) -> Name -> Q [Dec]

makeCustomLenses :: Name -> Q [Dec]

Create lenses for all fields of a given data type.

makeCustomLenses' :: Name -> [Name] -> Q [Dec]

Create lenses for some fields of a given data type.

traverseOf2 :: Over (->) (Compose f g) s t a b -> (a -> f (g b)) -> s -> f (g t)

Traverses over a composition of two functors. Most often the g functor is (,) r and traverseOf2 is used to traverse an effectful computation that also returns an additional output value.

mapMOf2 :: Over (->) (Compose (WrappedMonad m) g) s t a b -> (a -> m (g b)) -> s -> m (g t)

Traverses over a composition of a monad and a functor. See traverseOf2.

atSet :: Ord a => a -> Lens' (Set a) Bool

A helper lens over sets. While a similar lens exists in the package (as Lens' Set (Maybe ())), it's available only in most recent versions. And using Bool instead of Maybe () is more convenient.