Ganeti.Logging

Description

Implementation of the Ganeti logging functionality.

This currently lacks the following (FIXME):

log file reopening

Note that this requires the hslogger library version 1.1 and above.

Documentation

data SyslogUsage

Syslog usage type.

Constructors

SyslogNo

SyslogYes

SyslogOnly

syslogUsageFromRaw :: forall m. Monad m => String -> m SyslogUsage

syslogUsageToRaw :: SyslogUsage -> String

logFormatter :: String -> Bool -> Bool -> LogFormatter a

openFormattedHandler :: LogHandler a => Bool -> LogFormatter a -> IO a -> IO [a]

setupLogging

Arguments

:: Maybe String
Log file

-> String
Program name

-> Bool
Debug level

-> Bool
Log to stderr

-> Bool
Log to console

-> SyslogUsage
Syslog usage

-> IO ()

Sets up the logging configuration.

Logging function aliases

class Monad m => MonadLog m where

A monad that allows logging.

Methods

logAt :: Priority -> String -> m ()

Log at a given level.

logDebug :: MonadLog m => String -> m ()

Log at debug level.

logInfo :: MonadLog m => String -> m ()

Log at info level.

logNotice :: MonadLog m => String -> m ()

Log at notice level.

logWarning :: MonadLog m => String -> m ()

Log at warning level.

logError :: MonadLog m => String -> m ()

Log at error level.

logCritical :: MonadLog m => String -> m ()

Log at critical level.

logAlert :: MonadLog m => String -> m ()

Log at alert level.

logEmergency :: MonadLog m => String -> m ()

Log at emergency level.

isDebugMode :: IO Bool

Check if the logging is at DEBUG level. DEBUG logging is unacceptable for production.

Logging in an error monad with rethrowing errors

withErrorLogAt :: (MonadLog m, MonadError e m, Show e) => Priority -> String -> m a -> m a

If an error occurs within a given computation, it annotated with a given message and logged and the error is re-thrown.