Ganeti.MaintD.Balance

Description

Balancing task of the maintenance daemon.

This module carries out the automated balancing done by the maintenance daemon. The actual balancing algorithm is imported from htools.

Collection of dynamic load data

data AllReports

Constructors

AllReports

Fields

rTotal :: Report

rIndividual :: Report

rMem :: Report

emptyReports :: AllReports

queryNode :: Node -> ResultT String IO AllReports

queryLoad :: List -> ResultT String IO (Container AllReports)

getXenInstances :: ResultT String IO (Set String)

findInstanceLoad :: String -> AllReports -> Maybe Double

updateCPUInstance :: List -> Container AllReports -> Set String -> [String] -> Instance -> Result Instance

updateCPULoad :: (List, List) -> Container AllReports -> Set String -> [String] -> Result (List, List)

cleanUpEvacuation :: IORef MemoryState -> List -> Container AllReports -> String -> IO ()

Balancing

moveToJob :: Timestamp -> (List, List) -> MoveJob -> [MetaOpCode]

iterateBalance :: AlgorithmOptions -> Table -> [MoveJob] -> [MoveJob]

evacuatedInsts :: (List, List) -> MoveJob -> [String]

balanceGroup :: IORef MemoryState -> Set String -> Client -> Set Int -> Double -> (Int, (List, List)) -> ResultT String IO [JobId]

Memory balancing

weightFromMemRatio :: Double -> Double

useMemData :: Double -> Container AllReports -> (List, List) -> ResultT String IO (List, List)

Interface function

balanceTask

Arguments

:: IORef MemoryState

-> (List, List)
current cluster configuration

-> Set Int
node indices on which actions may be taken

-> Double
threshold for improvement

-> ResultT String IO [JobId]
jobs submitted

Carry out all the needed balancing, based on live CPU data, only touching the available nodes. Only carry out balancing steps where the gain is above the threshold.