Ganeti.MaintD.CleanupIncidents

Description

Incident clean up in the maintenance daemon.

This module implements the clean up of events that are finished, and acknowledged as such by the user.

Documentation

cleanupIncident :: IORef MemoryState -> List -> Incident -> ResultT String IO ()

cleanupIncidents :: IORef MemoryState -> List -> ResultT String IO ()

Remove all incidents from the record that are in a final state and additionally the node tag for that incident has been removed.