Ganeti.MaintD.FailIncident

Description

Incident failing in the maintenace daemon

This module implements the treatment of an incident, once a job failed.

Documentation

markAsFailed :: IORef MemoryState -> Incident -> ResultT String IO ()

failIncident :: IORef MemoryState -> JobId -> ResultT String IO ()

Mark the incident, if any, belonging to the given job as failed after having tagged it appropriately.