Ganeti.Metad.ConfigCore

Description

Functions of the metadata daemon exported for RPC

The monad in which all the Metad functions execute

data MetadHandle

Constructors

MetadHandle

Fields

mhInstParams :: MVar InstanceParams

type MetadMonadIntType = ReaderT MetadHandle IO

A type alias for easier referring to the actual content of the monad when implementing its instances.

newtype MetadMonadInt a

The internal part of the monad without error handling.

Constructors

MetadMonadInt

Fields

getMetadMonadInt :: MetadMonadIntType a

runMetadMonadInt :: MetadMonadInt a -> MetadHandle -> IO a

Runs the internal part of the MetadMonad monad on a given daemon handle.

type MetadMonad = ResultT GanetiException MetadMonadInt

The complete monad with error handling.

Basic functions in the monad

metadHandle :: MetadMonad MetadHandle

instParams :: MetadMonad InstanceParams

modifyInstParams :: (InstanceParams -> MetadMonad (InstanceParams, a)) -> MetadMonad a

Functions available to the RPC module

echo :: String -> MetadMonad String

updateConfig :: JSValue -> MetadMonad ()

Update the configuration with the received instance parameters.

The list of all functions exported to RPC.

exportedFunctions :: [Name]