Ganeti.Metad.ConfigServer

Description

Configuration server for the metadata daemon.

Documentation

updateConfig :: MVar InstanceParams -> String -> IO ()

Update the configuration with the received instance parameters.

acceptConfig :: MVar InstanceParams -> Client -> IO ()

Reads messages from clients and update the configuration according to these messages.

acceptClients :: MVar InstanceParams -> Server -> IO ()

Loop that accepts clients and dispatches them to an isolated thread that will handle the client's requests.

start :: DaemonOptions -> MVar InstanceParams -> IO ()