Ganeti.Objects

Description

Implementation of the Ganeti config objects.

Generic definitions

fillDict :: Ord k => Map k v -> Map k v -> [k] -> Map k v

Fills one map with keys from the other map, if not already existing. Mirrors objects.py:FillDict.

Network definitions

Ipv4 types

data Ip4Address

Constructors

Ip4Address Word8 Word8 Word8 Word8

mkIp4Address :: (Word8, Word8, Word8, Word8) -> Ip4Address

readIp4Address :: (Applicative m, Monad m) => String -> m Ip4Address

ip4AddressToList :: Ip4Address -> [Word8]

ip4AddressToNumber :: Ip4Address -> Integer

Converts an address into its ordinal number. This is needed for indexing IP adresses in reservation pools.

ip4AddressFromNumber :: Integer -> Ip4Address

Converts a number into an address. This is needed for indexing IP adresses in reservation pools.

nextIp4Address :: Ip4Address -> Ip4Address

data Ip4Network

Custom type for an IPv4 network.

Constructors

Ip4Network

Fields

ip4netAddr :: Ip4Address

ip4netMask :: Word8

mkIp4Network :: Ip4Address -> Word8 -> Ip4Network

Address pools

newtype AddressPool

Currently address pools just wrap a reservation BitArray.

In future, Network might be extended to include several address pools and address pools might include their own ranges of addresses.

Constructors

AddressPool

Fields

apReservations :: BitArray

Ganeti "network" config object.

data Network

Constructors

Network

Fields

networkName :: NonEmptyString

networkMacPrefix :: Maybe String

networkNetwork :: Ip4Network

networkNetwork6 :: Maybe String

networkGateway :: Maybe Ip4Address

networkGateway6 :: Maybe String

networkReservations :: Maybe AddressPool

networkExtReservations :: Maybe AddressPool

networkUuid :: ByteString

networkCtime :: ClockTime

networkMtime :: ClockTime

networkSerial :: Int

networkTags :: TagSet

loadNetwork :: JSValue -> Result Network

saveNetwork :: Network -> JSValue

Datacollector definitions

type MicroSeconds = Integer

data DataCollectorConfig

The configuration regarding a single data collector.

Constructors

DataCollectorConfig

Fields

dataCollectorActive :: Bool

dataCollectorInterval :: MicroSeconds

loadDataCollectorConfig :: JSValue -> Result DataCollectorConfig

saveDataCollectorConfig :: DataCollectorConfig -> JSValue

IPolicy definitions

data FilledISpecParams

Constructors

FilledISpecParams

Fields

ispecMemorySize :: Int

ispecDiskSize :: Int

ispecDiskCount :: Int

ispecCpuCount :: Int

ispecNicCount :: Int

ispecSpindleUse :: Int

data PartialISpecParams

Constructors

PartialISpecParams

Fields

ispecMemorySizeP :: Maybe Int

ispecDiskSizeP :: Maybe Int

ispecDiskCountP :: Maybe Int

ispecCpuCountP :: Maybe Int

ispecNicCountP :: Maybe Int

ispecSpindleUseP :: Maybe Int

allISpecParamFields :: [String]

loadPartialISpecParams :: JSValue -> Result PartialISpecParams

savePartialISpecParams :: PartialISpecParams -> JSValue

loadFilledISpecParams :: JSValue -> Result FilledISpecParams

saveFilledISpecParams :: FilledISpecParams -> JSValue

data MinMaxISpecs

Constructors

MinMaxISpecs

Fields

mmisMinSpec :: FilledISpecParams

mmisMaxSpec :: FilledISpecParams

loadMinMaxISpecs :: JSValue -> Result MinMaxISpecs

saveMinMaxISpecs :: MinMaxISpecs -> JSValue

data PartialIPolicy

Custom partial ipolicy. This is not built via buildParam since it has a special 2-level inheritance mode.

Constructors

PartialIPolicy

Fields

ipolicyMinMaxISpecsP :: Maybe [MinMaxISpecs]

ipolicyStdSpecP :: Maybe PartialISpecParams

ipolicySpindleRatioP :: Maybe Double

ipolicyVcpuRatioP :: Maybe Double

ipolicyDiskTemplatesP :: Maybe [DiskTemplate]

loadPartialIPolicy :: JSValue -> Result PartialIPolicy

savePartialIPolicy :: PartialIPolicy -> JSValue

data FilledIPolicy

Custom filled ipolicy. This is not built via buildParam since it has a special 2-level inheritance mode.

Constructors

FilledIPolicy

Fields

ipolicyMinMaxISpecs :: [MinMaxISpecs]

ipolicyStdSpec :: FilledISpecParams

ipolicySpindleRatio :: Double

ipolicyVcpuRatio :: Double

ipolicyDiskTemplates :: [DiskTemplate]

loadFilledIPolicy :: JSValue -> Result FilledIPolicy

saveFilledIPolicy :: FilledIPolicy -> JSValue

Node definitions

data FilledNDParams

Constructors

FilledNDParams

Fields

ndpOobProgram :: String

ndpSpindleCount :: Int

ndpExclusiveStorage :: Bool

ndpOvs :: Bool

ndpOvsName :: String

ndpOvsLink :: String

ndpSshPort :: Int

ndpCpuSpeed :: Double

data PartialNDParams

Constructors

PartialNDParams

Fields

ndpOobProgramP :: Maybe String

ndpSpindleCountP :: Maybe Int

ndpExclusiveStorageP :: Maybe Bool

ndpOvsP :: Maybe Bool

ndpOvsNameP :: Maybe String

ndpOvsLinkP :: Maybe String

ndpSshPortP :: Maybe Int

ndpCpuSpeedP :: Maybe Double

allNDParamFields :: [String]

loadPartialNDParams :: JSValue -> Result PartialNDParams

savePartialNDParams :: PartialNDParams -> JSValue

loadFilledNDParams :: JSValue -> Result FilledNDParams

saveFilledNDParams :: FilledNDParams -> JSValue

type DiskState = Container JSValue

type HypervisorState = Container JSValue

data Node

Constructors

Node

Fields

nodeName :: String

nodePrimaryIp :: String

nodeSecondaryIp :: String

nodeMasterCandidate :: Bool

nodeOffline :: Bool

nodeDrained :: Bool

nodeGroup :: String

nodeMasterCapable :: Bool

nodeVmCapable :: Bool

nodeNdparams :: PartialNDParams

nodePowered :: Bool

nodeHvStateStatic :: HypervisorState

nodeDiskStateStatic :: DiskState

nodeCtime :: ClockTime

nodeMtime :: ClockTime

nodeUuid :: ByteString

nodeSerial :: Int

nodeTags :: TagSet

loadNode :: JSValue -> Result Node

saveNode :: Node -> JSValue

NodeGroup definitions

type GroupDiskParams = Container DiskParams

The cluster/group disk parameters type.

type Networks = Container PartialNicParams

data NodeGroup

Constructors

NodeGroup

Fields

groupName :: String

groupMembers :: [String]

groupNdparams :: PartialNDParams

groupAllocPolicy :: AllocPolicy

groupIpolicy :: PartialIPolicy

groupDiskparams :: GroupDiskParams

groupNetworks :: Networks

groupHvStateStatic :: HypervisorState

groupDiskStateStatic :: DiskState

groupCtime :: ClockTime

groupMtime :: ClockTime

groupUuid :: ByteString

groupSerial :: Int

groupTags :: TagSet

loadNodeGroup :: JSValue -> Result NodeGroup

saveNodeGroup :: NodeGroup -> JSValue

Job scheduler filtering definitions

data FilterAction

Actions that can be performed when a filter matches.

Constructors

Accept

Pause

Reject

Continue

RateLimit Int

data FilterPredicate

Constructors

FPJobId (Filter FilterField)

FPOpCode (Filter FilterField)

FPReason (Filter FilterField)

data FilterRule

Constructors

FilterRule

Fields

frWatermark :: JobId

frPriority :: NonNegative Int

frPredicates :: [FilterPredicate]

frAction :: FilterAction

frReasonTrail :: ReasonTrail

frUuid :: ByteString

loadFilterRule :: JSValue -> Result FilterRule

saveFilterRule :: FilterRule -> JSValue

filterRuleOrder :: FilterRule -> FilterRule -> Ordering

Order in which filter rules are evaluated, according to `doc/design-optables.rst`. For FilterRule fields not specified as important for the order, we choose an arbitrary ordering effect (after the ones from the spec).

The Ord instance for FilterRule agrees with this function. Yet it is recommended to use this function instead of compare to be explicit that the spec order is used.

data IpFamily

IP family type

Constructors

IpFamilyV4

IpFamilyV6

ipFamilyFromRaw :: forall m. Monad m => Int -> m IpFamily

ipFamilyToRaw :: IpFamily -> Int

ipFamilyToVersion :: IpFamily -> Int

Conversion from IP family to IP version. This is needed because Python uses both, depending on context.

type ClusterHvParams = GenericContainer Hypervisor HvParams

Cluster HvParams (hvtype to hvparams mapping).

type OsHvParams = Container ClusterHvParams

Cluster Os-HvParams (os to hvparams mapping).

type ClusterBeParams = Container FilledBeParams

Cluser BeParams.

type ClusterOsParams = Container OsParams

Cluster OsParams.

type ClusterOsParamsPrivate = Container (Private OsParams)

type ClusterNicParams = Container FilledNicParams

Cluster NicParams.

type UidRange = (Int, Int)

A low-high UID ranges.

formatUidRange :: UidRange -> String

type UidPool = [UidRange]

Cluster UID Pool, list (low, high) UID ranges.

type IAllocatorParams = Container JSValue

The iallocator parameters type.

type CandidateCertificates = Container String

Cluster definitions

data Cluster

Constructors

Cluster

Fields

clusterRsahostkeypub :: String

clusterDsahostkeypub :: Maybe String

clusterHighestUsedPort :: Int

clusterTcpudpPortPool :: [Int]

clusterMacPrefix :: String

clusterVolumeGroupName :: Maybe String

clusterReservedLvs :: [String]

clusterDrbdUsermodeHelper :: Maybe String

clusterMasterNode :: String

clusterMasterIp :: String

clusterMasterNetdev :: String

clusterMasterNetmask :: Int

clusterUseExternalMipScript :: Bool

clusterClusterName :: String

clusterFileStorageDir :: String

clusterSharedFileStorageDir :: String

clusterGlusterStorageDir :: String

clusterEnabledHypervisors :: [Hypervisor]

clusterHvparams :: ClusterHvParams

clusterOsHvp :: OsHvParams

clusterBeparams :: ClusterBeParams

clusterOsparams :: ClusterOsParams

clusterOsparamsPrivateCluster :: ClusterOsParamsPrivate

clusterNicparams :: ClusterNicParams

clusterNdparams :: FilledNDParams

clusterDiskparams :: GroupDiskParams

clusterCandidatePoolSize :: Int

clusterModifyEtcHosts :: Bool

clusterModifySshSetup :: Bool

clusterMaintainNodeHealth :: Bool

clusterUidPool :: UidPool

clusterDefaultIallocator :: String

clusterDefaultIallocatorParams :: IAllocatorParams

clusterHiddenOs :: [String]

clusterBlacklistedOs :: [String]

clusterPrimaryIpFamily :: IpFamily

clusterPreallocWipeDisks :: Bool

clusterIpolicy :: FilledIPolicy

clusterHvStateStatic :: HypervisorState

clusterDiskStateStatic :: DiskState

clusterEnabledDiskTemplates :: [DiskTemplate]

clusterCandidateCerts :: CandidateCertificates

clusterMaxRunningJobs :: Int

clusterMaxTrackedJobs :: Int

clusterInstallImage :: String

clusterInstanceCommunicationNetwork :: String

clusterZeroingImage :: String

clusterCompressionTools :: [String]

clusterEnabledUserShutdown :: Bool

clusterDataCollectors :: Container DataCollectorConfig

clusterSshKeyType :: SshKeyType

clusterSshKeyBits :: Int

clusterCtime :: ClockTime

clusterMtime :: ClockTime

clusterUuid :: ByteString

clusterSerial :: Int

clusterTags :: TagSet

loadCluster :: JSValue -> Result Cluster

saveCluster :: Cluster -> JSValue

ConfigData definitions

data ConfigData

Constructors

ConfigData

Fields

configVersion :: Int

configCluster :: Cluster

configNodes :: Container Node

configNodegroups :: Container NodeGroup

configInstances :: Container Instance

configNetworks :: Container Network

configDisks :: Container Disk

configFilters :: Container FilterRule

configCtime :: ClockTime

configMtime :: ClockTime

configSerial :: Int

loadConfigData :: JSValue -> Result ConfigData

saveConfigData :: ConfigData -> JSValue

Master network parameters

data MasterNetworkParameters

Constructors

MasterNetworkParameters

Fields

masterNetworkParametersUuid :: String

masterNetworkParametersIp :: String

masterNetworkParametersNetmask :: Int

masterNetworkParametersNetdev :: String

masterNetworkParametersIpFamily :: IpFamily

loadMasterNetworkParameters :: JSValue -> Result MasterNetworkParameters

saveMasterNetworkParameters :: MasterNetworkParameters -> JSValue