Ganeti.Objects.BitArray

Description

Space efficient bit arrays

The module is meant to be imported qualified (as it is common with collection libraries).

Documentation

data BitArray

A fixed-size, space-efficient array of bits.

Constructors

BitArray

Fields

size :: !Int

_bitArrayBits :: !IntSet

empty :: BitArray

zeroes :: Int -> BitArray

foldr :: (Bool -> Int -> a -> a) -> a -> BitArray -> a

Right fold over the set, including indexes of each value.

asString :: Char -> Char -> BitArray -> String

Converts a bit array into a string, given characters for 0 and 1/

count0 :: BitArray -> Int

Computes the number of zeroes in the array.

count1 :: BitArray -> Int

Computes the number of ones in the array.

(!) :: BitArray -> Int -> Bool

Test a given bit in an array. If it's outside its scope, it's always False.

setAt :: (MonadError e m, FromString e) => Int -> Bool -> BitArray -> m BitArray

Sets or removes an element from a bit array.

Sets a given bit in an array. Fails if the index is out of bounds.

(-&-) :: BitArray -> BitArray -> BitArray

An intersection of two bit arrays. The length of the result is the minimum length of the two.

(-|-) :: BitArray -> BitArray -> BitArray

A union of two bit arrays. The length of the result is the maximum length of the two.

subset :: BitArray -> BitArray -> Bool

Checks if the first array is a subset of the other.

toList :: BitArray -> [Bool]

Converts a bit array into a list of booleans.

fromList :: [Bool] -> BitArray

Converts a list of booleans to a BitArray.