Ganeti.Objects.Maintenance

Description

Implementation of the Ganeti configuration for the maintenance daemon.

Documentation

data RepairAction

Action to be taken for a certain repair event. Note that the order is important, as we rely on values higher in the derived order to be more intrusive actions.

Constructors

RANoop

RALiveRepair

RAEvacuate

RAEvacuateFailover

repairActionFromRaw :: forall m. Monad m => String -> m RepairAction

repairActionToRaw :: RepairAction -> String

data RepairStatus

Progress made on the particular repair event. Again we rely on the order in that everything larger than RSPending is finalized in the sense that no further jobs will be submitted.

Constructors

RSNoted

RSPending

RSCanceled

RSFailed

RSCompleted

repairStatusFromRaw :: forall m. Monad m => String -> m RepairStatus

repairStatusToRaw :: RepairStatus -> String

data Incident

Constructors

Incident

Fields

incidentOriginal :: JSValue

incidentAction :: RepairAction

incidentJobs :: [JobId]

incidentNode :: String

incidentRepairStatus :: RepairStatus

incidentTag :: String

incidentUuid :: ByteString

incidentCtime :: ClockTime

incidentMtime :: ClockTime

incidentSerial :: Int

loadIncident :: JSValue -> Result Incident

saveIncident :: Incident -> JSValue

data MaintenanceData

Constructors

MaintenanceData

Fields

maintRoundDelay :: Int

maintJobs :: [JobId]

maintBalance :: Bool

maintBalanceThreshold :: Double

maintEvacuated :: [String]

maintIncidents :: [Incident]

maintCtime :: ClockTime

maintMtime :: ClockTime

maintSerial :: Int

loadMaintenanceData :: JSValue -> Result MaintenanceData

saveMaintenanceData :: MaintenanceData -> JSValue