Ganeti.OpCodes

Description

Implementation of the opcodes.

Documentation

data OpCode

OpCode representation.

We only implement a subset of Ganeti opcodes: those which are actually used in the htools codebase.

Constructors

OpTestDelay

Fields

opDelayDuration :: Double

opDelayOnMaster :: Bool

opDelayOnNodes :: [NonEmptyString]

opDelayRepeat :: NonNegative Int

OpInstanceReplaceDisks

Fields

opInstanceName :: String

opEarlyRelease :: Bool

opIgnoreIpolicy :: Bool

opReplaceDisksMode :: ReplaceDisksMode

opReplaceDisksList :: [DiskIndex]

opRemoteNode :: Maybe NonEmptyString

opIallocator :: Maybe NonEmptyString

OpInstanceFailover

Fields

opInstanceName :: String

opShutdownTimeout :: NonNegative Int

opIgnoreConsistency :: Bool

opTargetNode :: Maybe NonEmptyString

opIgnoreIpolicy :: Bool

opIallocator :: Maybe NonEmptyString

opMigrationCleanup :: Bool

OpInstanceMigrate

Fields

opInstanceName :: String

opMigrationMode :: Maybe MigrationMode

opOldLiveMode :: Maybe Bool

opTargetNode :: Maybe NonEmptyString

opAllowRuntimeChanges :: Bool

opIgnoreIpolicy :: Bool

opMigrationCleanup :: Bool

opIallocator :: Maybe NonEmptyString

opAllowFailover :: Bool

OpTagsGet

Fields

opKind :: TagObject

opUseLocking :: Bool

OpTagsSearch

Fields

opTagSearchPattern :: NonEmptyString

OpTagsSet

Fields

opKind :: TagObject

opTags :: [String]

OpTagsDel

Fields

opKind :: TagObject

opTags :: [String]

OpClusterPostInit

OpClusterDestroy

OpClusterQuery

OpClusterVerify

Fields

opDebugSimulateErrors :: Bool

opErrorCodes :: Bool

opSkipChecks :: Set VerifyOptionalChecks

opIgnoreErrors :: Set CVErrorCode

opVerbose :: Bool

opOptGroupName :: Maybe NonEmptyString

OpClusterVerifyConfig

Fields

opDebugSimulateErrors :: Bool

opErrorCodes :: Bool

opIgnoreErrors :: Set CVErrorCode

opVerbose :: Bool

OpClusterVerifyGroup

Fields

opGroupName :: NonEmptyString

opDebugSimulateErrors :: Bool

opErrorCodes :: Bool

opSkipChecks :: Set VerifyOptionalChecks

opIgnoreErrors :: Set CVErrorCode

opVerbose :: Bool

OpClusterVerifyDisks

OpGroupVerifyDisks

Fields

opGroupName :: NonEmptyString

OpClusterRepairDiskSizes

Fields

opInstances :: [NonEmptyString]

OpClusterConfigQuery

Fields

opOutputFields :: [NonEmptyString]

OpClusterRename

Fields

opName :: NonEmptyString

OpClusterSetParams

Fields

opForce :: Bool

opHvState :: Maybe UncheckedDict

opDiskState :: Maybe UncheckedDict

opVgName :: Maybe String

opEnabledHypervisors :: Maybe (NonEmpty Hypervisor)

opClusterHvParams :: Maybe (Container UncheckedDict)

opClusterBeParams :: Maybe UncheckedDict

opOsHvp :: Maybe (Container UncheckedDict)

opClusterOsParams :: Maybe (Container UncheckedDict)

opDiskparams :: Maybe (GenericContainer DiskTemplate UncheckedDict)

opCandidatePoolSize :: Maybe (Positive Int)

opUidPool :: Maybe [[(Int, Int)]]

opAddUids :: Maybe [[(Int, Int)]]

opRemoveUids :: Maybe [[(Int, Int)]]

opMaintainNodeHealth :: Maybe Bool

opPreallocWipeDisks :: Maybe Bool

opNicparams :: Maybe INicParams

opNdparams :: Maybe UncheckedDict

opIpolicy :: Maybe UncheckedDict

opDrbdHelper :: Maybe String

opDefaultIallocator :: Maybe String

opMasterNetdev :: Maybe String

opMasterNetmask :: Maybe (NonNegative Int)

opReservedLvs :: Maybe [NonEmptyString]

opHiddenOs :: Maybe TestClusterOsList

opBlacklistedOs :: Maybe TestClusterOsList

opUseExternalMipScript :: Maybe Bool

opEnabledDiskTemplates :: Maybe (NonEmpty DiskTemplate)

opModifyEtcHosts :: Maybe Bool

OpClusterRedistConf

OpClusterActivateMasterIp

OpClusterDeactivateMasterIp

OpQuery

Fields

opWhat :: QueryTypeOp

opUseLocking :: Bool

opFields :: [NonEmptyString]

opQfilter :: Filter String

OpQueryFields

Fields

opWhat :: QueryTypeOp

opFields :: [NonEmptyString]

OpOobCommand

Fields

opNodeNames :: [NonEmptyString]

opCommand :: OobCommand

opTimeout :: Int

opIgnoreStatus :: Bool

opPowerDelay :: Double

OpNodeRemove

Fields

opNodeName :: NonEmptyString

OpNodeAdd

Fields

opNodeName :: NonEmptyString

opHvState :: Maybe UncheckedDict

opDiskState :: Maybe UncheckedDict

opPrimaryIp :: Maybe String

opSecondaryIp :: Maybe NonEmptyString

opReadd :: Bool

opGroup :: Maybe NonEmptyString

opMasterCapable :: Maybe Bool

opVmCapable :: Maybe Bool

opNdparams :: Maybe UncheckedDict

OpNodeQuery

Fields

opOutputFields :: [NonEmptyString]

opNames :: [NonEmptyString]

opUseLocking :: Bool

OpNodeQueryvols

Fields

opOutputFields :: [NonEmptyString]

opNodes :: [NonEmptyString]

OpNodeQueryStorage

Fields

opOutputFields :: [NonEmptyString]

opStorageType :: StorageType

opNodes :: [NonEmptyString]

opStorageName :: NonEmptyString

OpNodeModifyStorage

Fields

opNodeName :: NonEmptyString

opStorageType :: StorageType

opStorageName :: NonEmptyString

opChanges :: UncheckedDict

OpRepairNodeStorage

Fields

opNodeName :: NonEmptyString

opStorageType :: StorageType

opStorageName :: NonEmptyString

opIgnoreConsistency :: Bool

OpNodeSetParams

Fields

opNodeName :: NonEmptyString

opForce :: Bool

opHvState :: Maybe UncheckedDict

opDiskState :: Maybe UncheckedDict

opMasterCandidate :: Maybe Bool

opOffline :: Maybe Bool

opDrained :: Maybe Bool

opAutoPromote :: Bool

opMasterCapable :: Maybe Bool

opVmCapable :: Maybe Bool

opSecondaryIp :: Maybe NonEmptyString

opNdparams :: Maybe UncheckedDict

opPowered :: Maybe Bool

OpNodePowercycle

Fields

opNodeName :: NonEmptyString

opForce :: Bool

OpNodeMigrate

Fields

opNodeName :: NonEmptyString

opMigrationMode :: Maybe MigrationMode

opOldLiveMode :: Maybe Bool

opTargetNode :: Maybe NonEmptyString

opAllowRuntimeChanges :: Bool

opIgnoreIpolicy :: Bool

opIallocator :: Maybe NonEmptyString

OpNodeEvacuate

Fields

opEarlyRelease :: Bool

opNodeName :: NonEmptyString

opRemoteNode :: Maybe NonEmptyString

opIallocator :: Maybe NonEmptyString

opEvacMode :: NodeEvacMode

OpInstanceCreate

Fields

opInstanceName :: String

opForceVariant :: Bool

opWaitForSync :: Bool

opNameCheck :: Bool

opIgnoreIpolicy :: Bool

opInstBeParams :: UncheckedDict

opinstDisks :: [IDiskParams]

opDiskTemplate :: DiskTemplate

opFileDriver :: Maybe FileDriver

opFileStorageDir :: Maybe NonEmptyString

opInstHvParams :: UncheckedDict

opHypervisor :: Maybe Hypervisor

opIallocator :: Maybe NonEmptyString

opIdentifyDefaults :: Bool

opIpCheck :: Bool

opConflictsCheck :: Bool

opInstCreateMode :: InstCreateMode

opNics :: [INicParams]

opNoInstall :: Maybe Bool

opInstOsParams :: UncheckedDict

opOsType :: Maybe NonEmptyString

opPnode :: Maybe NonEmptyString

opSnode :: Maybe NonEmptyString

opSourceHandshake :: Maybe UncheckedList

opSourceInstanceName :: Maybe NonEmptyString

opSourceShutdownTimeout :: NonNegative Int

opSourceX509Ca :: Maybe NonEmptyString

opSrcNode :: Maybe NonEmptyString

opSrcPath :: Maybe NonEmptyString

opStart :: Bool

opOpportunisticLocking :: Bool

opInstTags :: [NonEmptyString]

OpInstanceMultiAlloc

Fields

opIallocator :: Maybe NonEmptyString

opInstMultiAlloc :: UncheckedList

opOpportunisticLocking :: Bool

OpInstanceReinstall

Fields

opInstanceName :: String

opForceVariant :: Bool

opOsType :: Maybe NonEmptyString

opTempOsParams :: Maybe UncheckedDict

OpInstanceRemove

Fields

opInstanceName :: String

opShutdownTimeout :: NonNegative Int

opIgnoreFailures :: Bool

OpInstanceRename

Fields

opInstanceName :: String

opNewName :: NonEmptyString

opNameCheck :: Bool

opIpCheck :: Bool

OpInstanceStartup

Fields

opInstanceName :: String

opForce :: Bool

opIgnoreOfflineNodes :: Bool

opTempHvParams :: UncheckedDict

opTempBeParams :: UncheckedDict

opNoRemember :: Bool

opStartupPaused :: Bool

OpInstanceShutdown

Fields

opInstanceName :: String

opForce :: Bool

opIgnoreOfflineNodes :: Bool

opInstShutdownTimeout :: NonNegative Int

opNoRemember :: Bool

OpInstanceReboot

Fields

opInstanceName :: String

opShutdownTimeout :: NonNegative Int

opIgnoreSecondaries :: Bool

opRebootType :: RebootType

OpInstanceMove

Fields

opInstanceName :: String

opShutdownTimeout :: NonNegative Int

opIgnoreIpolicy :: Bool

opMoveTargetNode :: NonEmptyString

opIgnoreConsistency :: Bool

OpInstanceConsole

Fields

opInstanceName :: String

OpInstanceActivateDisks

Fields

opInstanceName :: String

opIgnoreSize :: Bool

opWaitForSync :: Bool

OpInstanceDeactivateDisks

Fields

opInstanceName :: String

opForce :: Bool

OpInstanceRecreateDisks

Fields

opInstanceName :: String

opRecreateDisksInfo :: RecreateDisksInfo

opNodes :: [NonEmptyString]

opIallocator :: Maybe NonEmptyString

OpInstanceQuery

Fields

opOutputFields :: [NonEmptyString]

opNames :: [NonEmptyString]

opUseLocking :: Bool

OpInstanceQueryData

Fields

opUseLocking :: Bool

opInstances :: [NonEmptyString]

opStatic :: Bool

OpInstanceSetParams

Fields

opInstanceName :: String

opForce :: Bool

opForceVariant :: Bool

opIgnoreIpolicy :: Bool

opInstNicChanges :: SetParamsMods INicParams

opInstDiskChanges :: SetParamsMods IDiskParams

opInstBeParams :: UncheckedDict

opRuntimeMem :: Maybe (Positive Int)

opInstHvParams :: UncheckedDict

opOptDiskTemplate :: Maybe DiskTemplate

opPnode :: Maybe NonEmptyString

opRemoteNode :: Maybe NonEmptyString

opOsName :: Maybe NonEmptyString

opInstOsParams :: UncheckedDict

opWaitForSync :: Bool

opOffline :: Maybe Bool

opConflictsCheck :: Bool

OpInstanceGrowDisk

Fields

opInstanceName :: String

opWaitForSync :: Bool

opDiskIndex :: DiskIndex

opDiskChgAmount :: NonNegative Int

opDiskChkAbsolute :: Bool

OpInstanceChangeGroup

Fields

opInstanceName :: String

opEarlyRelease :: Bool

opIallocator :: Maybe NonEmptyString

opTargetGroups :: Maybe [NonEmptyString]

OpGroupAdd

Fields

opGroupName :: NonEmptyString

opAllocPolicy :: Maybe AllocPolicy

opNdparams :: Maybe UncheckedDict

opDiskparams :: Maybe (GenericContainer DiskTemplate UncheckedDict)

opHvState :: Maybe UncheckedDict

opDiskState :: Maybe UncheckedDict

opIpolicy :: Maybe UncheckedDict

OpGroupAssignNodes

Fields

opGroupName :: NonEmptyString

opForce :: Bool

opReqNodes :: [NonEmptyString]

OpGroupQuery

Fields

opOutputFields :: [NonEmptyString]

opNames :: [NonEmptyString]

OpGroupSetParams

Fields

opGroupName :: NonEmptyString

opAllocPolicy :: Maybe AllocPolicy

opNdparams :: Maybe UncheckedDict

opDiskparams :: Maybe (GenericContainer DiskTemplate UncheckedDict)

opHvState :: Maybe UncheckedDict

opDiskState :: Maybe UncheckedDict

opIpolicy :: Maybe UncheckedDict

OpGroupRemove

Fields

opGroupName :: NonEmptyString

OpGroupRename

Fields

opGroupName :: NonEmptyString

opNewName :: NonEmptyString

OpGroupEvacuate

Fields

opGroupName :: NonEmptyString

opEarlyRelease :: Bool

opIallocator :: Maybe NonEmptyString

opTargetGroups :: Maybe [NonEmptyString]

OpOsDiagnose

Fields

opOutputFields :: [NonEmptyString]

opNames :: [NonEmptyString]

OpExtStorageDiagnose

Fields

opOutputFields :: [NonEmptyString]

opNames :: [NonEmptyString]

OpBackupQuery

Fields

opUseLocking :: Bool

opNodes :: [NonEmptyString]

OpBackupPrepare

Fields

opInstanceName :: String

opExportMode :: ExportMode

OpBackupExport

Fields

opInstanceName :: String

opShutdownTimeout :: NonNegative Int

opExportTarget :: ExportTarget

opShutdown :: Bool

opRemoveInstance :: Bool

opIgnoreRemoveFailures :: Bool

opExportMode :: ExportMode

opX509KeyName :: Maybe UncheckedList

opDestinationX509Ca :: Maybe NonEmptyString

OpBackupRemove

Fields

opInstanceName :: String

OpTestAllocator

Fields

opIAllocatorDirection :: IAllocatorTestDir

opIAllocatorMode :: IAllocatorMode

opIAllocatorReqName :: NonEmptyString

opIAllocatorNics :: [UncheckedDict]

opIAllocatorDisks :: UncheckedList

opHypervisor :: Maybe Hypervisor

opIallocator :: Maybe NonEmptyString

opInstTags :: [NonEmptyString]

opIAllocatorMem :: Maybe (NonNegative Int)

opIAllocatorVCpus :: Maybe (NonNegative Int)

opIAllocatorOs :: Maybe NonEmptyString

opDiskTemplate :: DiskTemplate

opIAllocatorInstances :: Maybe [NonEmptyString]

opIAllocatorEvacMode :: Maybe NodeEvacMode

opTargetGroups :: Maybe [NonEmptyString]

opIAllocatorSpindleUse :: NonNegative Int

opIAllocatorCount :: NonNegative Int

OpTestJqueue

Fields

opNotifyWaitlock :: Bool

opNotifyExec :: Bool

opLogMessages :: [String]

opJQueueFail :: Bool

OpTestDummy

Fields

opTestDummyResult :: UncheckedValue

opTestDummyMessages :: UncheckedValue

opTestDummyFail :: UncheckedValue

opTestDummySubmitJobs :: UncheckedValue

OpNetworkAdd

Fields

opNetworkName :: NonEmptyString

opNetworkAddress4 :: NonEmptyString

opNetworkGateway4 :: Maybe NonEmptyString

opNetworkAddress6 :: Maybe NonEmptyString

opNetworkGateway6 :: Maybe NonEmptyString

opNetMacPrefix :: Maybe NonEmptyString

opNetworkAddRsvdIps :: Maybe [NonEmptyString]

opConflictsCheck :: Bool

opInstTags :: [NonEmptyString]

OpNetworkRemove

Fields

opNetworkName :: NonEmptyString

opForce :: Bool

OpNetworkSetParams

Fields

opNetworkName :: NonEmptyString

opNetworkGateway4 :: Maybe NonEmptyString

opNetworkAddress6 :: Maybe NonEmptyString

opNetworkGateway6 :: Maybe NonEmptyString

opNetMacPrefix :: Maybe NonEmptyString

opNetworkAddRsvdIps :: Maybe [NonEmptyString]

opNetworkRemoveRsvdIps :: Maybe [NonEmptyString]

OpNetworkConnect

Fields

opGroupName :: NonEmptyString

opNetworkName :: NonEmptyString

opNetworkMode :: NICMode

opNetworkLink :: NonEmptyString

opConflictsCheck :: Bool

OpNetworkDisconnect

Fields

opGroupName :: NonEmptyString

opNetworkName :: NonEmptyString

OpNetworkQuery

Fields

opOutputFields :: [NonEmptyString]

opNames :: [NonEmptyString]

opUseLocking :: Bool

OpRestrictedCommand

Fields

opUseLocking :: Bool

opReqNodes :: [NonEmptyString]

opRestrictedCommand :: NonEmptyString

saveOpCode :: OpCode -> JSValue

toDictOpCode :: OpCode -> [(String, JSValue)]

loadOpCode :: JSValue -> Result OpCode

allOpFields :: String -> [String]

opID :: OpCode -> String

Returns the OP_ID for a given opcode value.

allOpIDs :: [String]

A list of all defined/supported opcode IDs.

opSummaryVal :: OpCode -> Maybe String

opSummary :: OpCode -> String

Computes the summary of the opcode.

data CommonOpParams

Generic/common opcode parameters.

Constructors

CommonOpParams

Fields

opDryRun :: Maybe Bool

opDebugLevel :: Maybe (NonNegative Int)

opPriority :: OpSubmitPriority

opDepends :: Maybe [JobDependency]

opComment :: Maybe String

opReason :: ReasonTrail

loadCommonOpParams :: JSValue -> Result CommonOpParams

saveCommonOpParams :: CommonOpParams -> JSValue

toDictCommonOpParams :: CommonOpParams -> [(String, JSValue)]

defOpParams :: CommonOpParams

Default common parameter values.

data MetaOpCode

The top-level opcode type.

Constructors

MetaOpCode

Fields

metaParams :: CommonOpParams

metaOpCode :: OpCode

showMeta :: MetaOpCode -> JSValue

readMeta :: JSValue -> Result MetaOpCode

wrapOpCode :: OpCode -> MetaOpCode

Wraps an OpCode with the default parameters to build a MetaOpCode.

setOpComment :: String -> MetaOpCode -> MetaOpCode

Sets the comment on a meta opcode.

setOpPriority :: OpSubmitPriority -> MetaOpCode -> MetaOpCode

Sets the priority on a meta opcode.