Ganeti.Parsers

Description

Utility functions for several parsers

This module holds the definition for some utility functions for two parsers. The parser for the procstat file and the parser for the procdiskstats file.

Utility functions

skipSpaces :: Parser ()

Our own space-skipping function, because A.skipSpace also skips newline characters. It skips ZERO or more spaces, so it does not fail if there are no spaces.

numberP :: Parser Int

A parser recognizing a number preceeded by spaces.

integerP :: Parser Integer

A parser recognizing a number preceeded by spaces.

stringP :: Parser String

A parser recognizing a word preceded by spaces, and closed by a space.