Ganeti.Path

Description

Path-related helper functions.

Documentation

pjoin :: IO String -> String -> IO String

getRootDir :: IO FilePath

addNodePrefix :: FilePath -> IO FilePath

dataDir :: IO FilePath

Directory for data.

dataDirP :: FilePath -> IO FilePath

runDir :: IO FilePath

Directory for runtime files.

logDir :: IO FilePath

Directory for log files.

socketDir :: IO FilePath

Directory for Unix sockets.

defaultMasterSocket :: IO FilePath

The default path for the master-daemon LUXI socket.

defaultQuerySocket :: IO FilePath

The default LUXI socket for queries.

confdHmacKey :: IO FilePath

Path to file containing confd's HMAC key.

clusterConfFile :: IO FilePath

Path to cluster configuration file.

watcherPauseFile :: IO FilePath

Path to the watcher pause file.

nodedCertFile :: IO FilePath

Path to the noded certificate.

nodedClientCertFile :: IO FilePath

Path to the noded client certificate.

queueDir :: IO FilePath

Job queue directory.

jobQueueSerialFile :: IO FilePath

Job queue serial file.

jobQueueLockFile :: IO FilePath

Job queue lock file

jobQueueDrainFile :: IO FilePath

Job queue drain file

jobQueueArchiveSubDir :: FilePath

Job queue archive directory.

instanceReasonDir :: IO FilePath

Directory containing the reason trails for the last change of status of instances.

getInstReasonFilename :: String -> IO FilePath

The path of the file containing the reason trail for an instance, given the instance name.