Ganeti.PyValue

Description

PyValue contains instances for the PyValue typeclass.

The typeclass PyValue converts Haskell values to Python values. This module contains instances of this typeclass for several generic types. These instances are used in the Haskell to Python generation of opcodes and constants, for example.

PyValue represents data types convertible to Python

class PyValue a where

Converts Haskell values into Python values

This is necessary for the default values of opcode parameters and return values. For example, if a default value or return type is a Data.Map, then it must be shown as a Python dictioanry.

Methods

showValue :: a -> String

showValueList :: [a] -> String

PyValue represents an unspecified value convertible to Python

data PyValueEx

Encapsulates Python default values

Constructors

forall a . PyValue a => PyValueEx a