Ganeti.Query.Instance

Description

Implementation of the Ganeti Query2 instance queries.

Documentation

type LiveInfo = (Maybe (InstanceInfo, Bool), Maybe InstanceConsoleInfo)

type Runtime = Either RpcError LiveInfo

Runtime containing the LiveInfo. See the genericQuery function in the Query.hs file for an explanation of the terms used.

fieldsMap :: FieldMap Instance Runtime

The instance fields map.

instanceAliases :: [(FieldName, FieldName)]

The instance aliases.

aliasedFields :: FieldList Instance Runtime

instanceFields :: FieldList Instance Runtime

The instance fields.

Helper functions for node property retrieval

nicDescSuffix :: String

nicAggDescPrefix :: String

getNetworkName :: ConfigData -> String -> NonEmptyString

getNicBridge :: FilledNicParams -> Maybe String

getNicVlan :: FilledNicParams -> Maybe String

fillNicParamsFromConfig :: ConfigData -> PartialNicParams -> FilledNicParams

getDefaultNicParams :: ConfigData -> FilledNicParams

getDiskSizeRequirements :: ConfigData -> Instance -> ResultEntry

getDiskSizes :: ConfigData -> Instance -> ResultEntry

getDiskSpindles :: ConfigData -> Instance -> ResultEntry

getDiskNames :: ConfigData -> Instance -> ResultEntry

getDiskUuids :: ConfigData -> Instance -> ResultEntry

getProviders :: ConfigData -> Instance -> ResultEntry

getStorageIds :: ConfigData -> Instance -> ResultEntry

getIndexedOptionalConfField :: JSON b => (ConfigData -> Instance -> ErrorResult [a]) -> (a -> Maybe b) -> Int -> FieldGetter Instance Runtime

getIndexedConfField :: JSON b => (ConfigData -> Instance -> ErrorResult [a]) -> (a -> b) -> Int -> FieldGetter Instance Runtime

getIndexedNicNetworkNameField :: Int -> FieldGetter Instance Runtime

getIndexedNicField :: JSON a => (FilledNicParams -> a) -> Int -> FieldGetter Instance Runtime

getOptionalIndexedNicField :: JSON a => (FilledNicParams -> Maybe a) -> Int -> FieldGetter Instance Runtime

getIndexedFieldWithDefault :: JSON c => (Instance -> [a]) -> (ConfigData -> Instance -> b) -> (b -> a -> b) -> (b -> Maybe c) -> Int -> FieldGetter Instance Runtime

getIndexedOptionalField :: JSON b => (Instance -> [a]) -> (a -> Maybe b) -> Int -> FieldGetter Instance Runtime

getIndexedField :: JSON b => (Instance -> [a]) -> (a -> b) -> Int -> FieldGetter Instance Runtime

maybeAt :: Int -> [a] -> Maybe a

fieldDefinitionCompleter :: PrintfArg t1 => PrintfArg t2 => FieldName -> FieldTitle -> FieldType -> FieldDoc -> t1 -> t2 -> FieldDefinition

fillIncompleteFields :: (t1 -> t2 -> FieldDefinition, t1 -> FieldGetter a b, QffMode) -> t1 -> t2 -> FieldData a b

instantiateIndexedFields :: (Show t1, Integral t1) => Int -> [(t1 -> String -> FieldDefinition, t1 -> FieldGetter a b, QffMode)] -> FieldList a b

Various helper functions for property retrieval

getPrimaryNode :: ConfigData -> Instance -> ErrorResult Node

getPrimaryNodeName :: ConfigData -> Instance -> ResultEntry

getPrimaryNodeGroup :: ConfigData -> Instance -> ErrorResult NodeGroup

getPrimaryNodeGroupName :: ConfigData -> Instance -> ResultEntry

getPrimaryNodeGroupUuid :: ConfigData -> Instance -> ResultEntry

getSecondaryNodes :: ConfigData -> Instance -> ErrorResult [Node]

getSecondaryNodeAttribute :: JSON a => (Node -> a) -> ConfigData -> Instance -> ResultEntry

getSecondaryNodeGroups :: ConfigData -> Instance -> ErrorResult [NodeGroup]

getSecondaryNodeGroupAttribute :: JSON a => (NodeGroup -> a) -> ConfigData -> Instance -> ResultEntry

beParamGetter :: String -> ConfigData -> Instance -> ResultEntry

hvParamGetter :: String -> ConfigData -> Instance -> ResultEntry

Live fields functionality

instanceLiveFieldsDefs :: [(FieldName, FieldTitle, FieldType, String, FieldDoc)]

instanceLiveFieldExtract :: FieldName -> InstanceInfo -> Instance -> JSValue

instanceLiveRpcCall :: FieldName -> Runtime -> Instance -> ResultEntry

instanceLiveFieldBuilder :: (FieldName, FieldTitle, FieldType, String, FieldDoc) -> FieldData Instance Runtime

Functionality related to status and operational status extraction

statusDocText :: String

isPrimaryOffline :: ConfigData -> Instance -> Bool

userShutdownEnabled :: ConfigData -> Bool

liveInstanceStatus :: ConfigData -> (InstanceInfo, Bool) -> Instance -> InstanceStatus

deadInstanceStatus :: ConfigData -> Instance -> InstanceStatus

determineInstanceStatus :: ConfigData -> Runtime -> Instance -> InstanceStatus

statusExtract :: ConfigData -> Runtime -> Instance -> ResultEntry

operStatusExtract :: Runtime -> Instance -> ResultEntry

consoleExtract :: Runtime -> Instance -> ResultEntry

Helper functions extracting information as necessary for the generic query

checkForNodeError :: [(String, ERpcError a)] -> String -> Maybe RpcError

findInfoInNodeResult :: Instance -> ERpcError RpcResultAllInstancesInfo -> Maybe InstanceInfo

getInstanceInfo :: [(String, ERpcError RpcResultAllInstancesInfo)] -> Instance -> ERpcError (Maybe (InstanceInfo, Bool))

Retrieves the instance information if it is present anywhere in the all instances RPC result. Notes if it originates from the primary node. An error is delivered if there is no result, and the primary node is down.

getConsoleInfo :: [(String, ERpcError RpcResultInstanceConsoleInfo)] -> Instance -> Maybe InstanceConsoleInfo

extractLiveInfo :: [(Node, ERpcError RpcResultAllInstancesInfo)] -> [(Node, ERpcError RpcResultInstanceConsoleInfo)] -> Instance -> Runtime

getAllConsoleParams :: ConfigData -> [Instance] -> ErrorResult [InstanceConsoleInfoParams]

compareParamsByNode :: InstanceConsoleInfoParams -> InstanceConsoleInfoParams -> Bool

consoleParamsToCalls :: [InstanceConsoleInfoParams] -> [(Node, RpcCallInstanceConsoleInfo)]

getHypervisorSpecs :: ConfigData -> [Instance] -> [(Hypervisor, HvParams)]

collectLiveData

Arguments

:: Bool
Live queries allowed

-> ConfigData
The cluster config

-> [String]
The requested fields

-> [Instance]
The instance objects

-> IO [(Instance, Runtime)]

Collect live data from RPC query if enabled.

getDiskTemplate :: ConfigData -> Instance -> ResultEntry