Ganeti.Query.Language

Description

Implementation of the Ganeti Query2 language.

THH declarations, that require ordering.

data ResultStatus

Status of a query field.

Constructors

RSNormal

RSUnknown

RSNoData

RSUnavail

RSOffline

resultStatusFromRaw :: forall m. Monad m => Int -> m ResultStatus

resultStatusToRaw :: ResultStatus -> Int

checkRS :: Monad m => ResultStatus -> a -> m a

Check that ResultStatus is success or fail with descriptive message.

data FieldType

Type of a query field.

Constructors

QFTUnknown

QFTText

QFTBool

QFTNumber

QFTUnit

QFTTimestamp

QFTOther

fieldTypeFromRaw :: forall m. Monad m => String -> m FieldType

fieldTypeToRaw :: FieldType -> String

data QueryTypeOp

Supported items on which Qlang works.

Constructors

QRCluster

QRInstance

QRNode

QRGroup

QROs

QRExport

QRNetwork

queryTypeOpFromRaw :: forall m. Monad m => String -> m QueryTypeOp

queryTypeOpToRaw :: QueryTypeOp -> String

data QueryTypeLuxi

Supported items on which Qlang works.

Constructors

QRLock

QRJob

queryTypeLuxiFromRaw :: forall m. Monad m => String -> m QueryTypeLuxi

queryTypeLuxiToRaw :: QueryTypeLuxi -> String

data ItemType

Overall query type.

Constructors

ItemTypeLuxi QueryTypeLuxi

ItemTypeOpCode QueryTypeOp

decodeItemType :: Monad m => JSValue -> m ItemType

Sub data types for query2 queries and responses.

type RegexType = Regex

type Fields = [String]

List of requested fields.

data Filter a

Query2 filter expression. It's a parameteric type since we can filter different "things"; e.g. field names, or actual field getters, etc.

Constructors

EmptyFilter
No filter at all

AndFilter [Filter a]
& [expression, ...]

OrFilter [Filter a]
| [expression, ...]

NotFilter (Filter a)
! expression

TrueFilter a
? field

EQFilter a FilterValue
(=|!=) field value

LTFilter a FilterValue
< field value

GTFilter a FilterValue
> field value

LEFilter a FilterValue
<= field value

GEFilter a FilterValue
>= field value

RegexpFilter a FilterRegex
=~ field regexp

ContainsFilter a FilterValue
=[] list-field value

showFilter :: JSON a => Filter a -> JSValue

readFilter :: JSON a => JSValue -> Result (Filter a)

readFilterArg :: JSON a => (Filter a -> Filter a) -> [JSValue] -> Result (Filter a)

readFilterField :: JSON a => (a -> Filter a) -> [JSValue] -> Result (Filter a)

readFilterFieldValue :: (JSON a, JSON b) => (a -> b -> Filter a) -> [JSValue] -> Result (Filter a)

readFilterArray :: JSON a => String -> [JSValue] -> Result (Filter a)

traverseFlt :: Applicative f => (a -> f b) -> Filter a -> f (Filter b)

type FilterField = String

Field name to filter on.

data FilterValue

Value to compare the field value to, for filtering purposes.

Constructors

QuotedString String

NumericValue Integer

showFilterValue :: FilterValue -> JSValue

readFilterValue :: JSValue -> Result FilterValue

data FilterRegex

Regexp to apply to the filter value, for filtering purposes. It holds both the string format, and the "compiled" format, so that we don't re-compile the regex at each match attempt.

Constructors

FilterRegex

Fields

stringRegex :: String

The string version of the regex

compiledRegex :: RegexType

The compiled regex

mkRegex :: Monad m => String -> m FilterRegex

Builder for FilterRegex. We always attempt to compile the regular expression on the initialisation of the data structure; this might fail, if the RE is not well-formed.

type FieldName = String

Name of a field.

type FieldTitle = String

Title of a field, when represented in tabular format.

type FieldDoc = String

Human redable description of a field.

data FieldDefinition

Definition of a field.

Constructors

FieldDefinition

Fields

fdefName :: FieldName

fdefTitle :: FieldTitle

fdefKind :: FieldType

fdefDoc :: FieldDoc

loadFieldDefinition :: JSValue -> Result FieldDefinition

saveFieldDefinition :: FieldDefinition -> JSValue

toDictFieldDefinition :: FieldDefinition -> [(String, JSValue)]

data ResultEntry

Constructors

ResultEntry

Fields

rentryStatus :: ResultStatus

The result status

rentryValue :: Maybe ResultValue

The (optional) result value

type ResultRow = [ResultEntry]

type ResultValue = JSValue

Value of a field, in json encoding. (its type will be depending on ResultStatus and FieldType)

Main Qlang queries and responses.

data Query

Query2 query.

Constructors

Query ItemType Fields (Filter FilterField)

data QueryResult

Query2 result.

Constructors

QueryResult

Fields

qresFields :: [FieldDefinition]

qresData :: [ResultRow]

loadQueryResult :: JSValue -> Result QueryResult

saveQueryResult :: QueryResult -> JSValue

toDictQueryResult :: QueryResult -> [(String, JSValue)]

data QueryFields

Query2 Fields query. (to get supported fields names, descriptions, and types)

Constructors

QueryFields ItemType Fields

data QueryFieldsResult

Query2 Fields result.

Constructors

QueryFieldsResult

Fields

qfieldresFields :: [FieldDefinition]

loadQueryFieldsResult :: JSValue -> Result QueryFieldsResult

saveQueryFieldsResult :: QueryFieldsResult -> JSValue

toDictQueryFieldsResult :: QueryFieldsResult -> [(String, JSValue)]