Ganeti.Query.Network

Description

Implementation of the Ganeti Query2 node group queries.

Documentation

data Runtime

There is no actual runtime.

Constructors

Runtime

networkFields :: FieldList Network Runtime

fieldsMap :: FieldMap Network Runtime

The group fields map.

getGroupConnections :: ConfigData -> String -> [(String, String, String, String)]

getGroupConnection :: String -> NodeGroup -> Maybe (String, String, String, String)

Given a network's UUID and a node group, this function assembles a tuple of the group's name, the mode and the link by which the network is connected to the group. Returns Nothing if the network is not connected to the group.

getNicMode :: PartialNicParams -> String

getNicLink :: PartialNicParams -> String

getNicVlan :: PartialNicParams -> String

getInstances :: ConfigData -> String -> [String]

instIsConnected :: String -> Instance -> Bool

Helper function that checks if an instance is linked to the given network.

getNetworkUuid :: ConfigData -> String -> Maybe String

Helper function to look up a network's UUID by its name

getReservations :: Ip4Network -> String -> [Ip4Address]

getExtReservationsString :: Network -> ResultEntry

collectLiveData :: Bool -> ConfigData -> [Network] -> IO [(Network, Runtime)]

Dummy function for collecting live data (which networks don't have).