Ganeti.Query.Node

Description

Implementation of the Ganeti Query2 node queries.

Documentation

type NodeRuntime = Either RpcError RpcResultNodeInfo

NodeRuntime is the resulting type for NodeInfo call.

nodeLiveFieldsDefs :: [(FieldName, FieldTitle, FieldType, String, FieldDoc)]

nodeLiveFieldExtract :: FieldName -> RpcResultNodeInfo -> JSValue

nodeLiveRpcCall :: FieldName -> NodeRuntime -> Node -> ResultEntry

nodeLiveFieldBuilder :: (FieldName, FieldTitle, FieldType, String, FieldDoc) -> FieldData Node NodeRuntime

nodeRoleDoc :: String

getNodePower :: ConfigData -> Node -> ResultEntry

nodeFields :: FieldList Node NodeRuntime

nodeFieldsMap :: FieldMap Node NodeRuntime

The node fields map.

maybeCollectLiveData :: Bool -> ConfigData -> [Node] -> IO [(Node, NodeRuntime)]

Collect live data from RPC query if enabled.

FIXME: Check which fields we actually need and possibly send empty hvsvgs if no info from hypervisorvolume group respectively is required