Ganeti.Query.Query

Description

Implementation of the Ganeti Query2 functionality.

Helper functions

mkUnknownFDef :: String -> FieldData a b

execGetter :: ConfigData -> b -> a -> FieldGetter a b -> ResultEntry

Main query execution

getSelectedFields :: FieldMap a b -> [String] -> FieldList a b

needsLiveData :: [FieldGetter a b] -> Bool

needsNames :: Query -> Maybe [FilterValue]

nameField :: ItemType -> FilterField

Computes the name field for different query types.

getAllQuotedStrings :: [FilterValue] -> [String]

getRequestedNames :: Query -> [String]

Checks that we have either requested a valid set of names, or we have a more complex filter.

getRequestedJobIDs :: Filter FilterField -> Result [JobId]

query

Arguments

:: ConfigData
The current configuration

-> Bool
Whether to collect live data

-> Query
The query (item, fields, filter)

-> IO (ErrorResult QueryResult)
Result

Main query execution function.

queryInner :: ConfigData -> Bool -> Query -> [String] -> IO (ErrorResult QueryResult)

queryJobs :: ConfigData -> Bool -> [FilterField] -> Filter FilterField -> IO (ErrorResult QueryResult)

fieldsExtractor :: FieldMap a b -> [FilterField] -> QueryFieldsResult

queryFields :: QueryFields -> ErrorResult QueryFieldsResult

Query fields call.

queryCompat :: QueryResult -> ErrorResult [[JSValue]]

Classic query converter. It gets a standard query result on input and computes the classic style results.