Ganeti.Rpc

Description

Implementation of the RPC client.

Base RPC functionality and types

curlOpts :: [CurlOption]

data RpcError

Data type for RPC error reporting.

Constructors

CurlLayerError String

JsonDecodeError String

RpcResultError String

OfflineNodeError

explainRpcError :: RpcError -> String

Provide explanation to RPC errors.

type ERpcError = Either RpcError

class JSON a => RpcCall a where

A generic class for RPC calls.

Methods

rpcCallName :: a -> String

Give the (Python) name of the procedure.

rpcCallTimeout :: a -> Int

Calculate the timeout value for the call execution.

rpcCallData :: Node -> a -> String

Prepare arguments of the call to be send as POST.

rpcCallAcceptOffline :: a -> Bool

Whether we accept offline nodes when making a call.

class (RpcCall a, JSON b) => Rpc a b | a -> b, b -> a where

Generic class that ensures matching RPC call with its respective result.

Methods

rpcResultFill :: a -> JSValue -> ERpcError b

Create a result based on the received HTTP response.

data HttpClientRequest

Constructors

HttpClientRequest

Fields

requestUrl :: String

requestData :: String

requestOpts :: [CurlOption]

isIpV6 :: String -> Bool

prepareUrl :: RpcCall a => Node -> a -> String

prepareHttpRequest :: RpcCall a => [CurlOption] -> Node -> a -> ERpcError HttpClientRequest

parseHttpReply :: Rpc a b => a -> ERpcError (CurlCode, String) -> ERpcError b

parseHttpResponse :: Rpc a b => a -> String -> ERpcError b

logRpcErrors :: [(a, ERpcError b)] -> IO ()

Scan the list of results produced by executeRpcCall and log all the RPC errors.

executeRpcCall :: Rpc a b => [Node] -> a -> IO [(Node, ERpcError b)]

Execute RPC call for many nodes in parallel.

sanitizeDictResults :: [(String, Result a)] -> ERpcError [(String, a)]

fromJResultToRes :: Result a -> (a -> b) -> ERpcError b

fromJSValueToRes :: JSON a => JSValue -> (a -> b) -> ERpcError b

RPC calls and results

Instance info

data RpcCallInstanceInfo

InstanceInfo Returns information about a single instance.

Constructors

RpcCallInstanceInfo

Fields

rpcCallInstInfoInstance :: String

rpcCallInstInfoHname :: Hypervisor

loadRpcCallInstanceInfo :: JSValue -> Result RpcCallInstanceInfo

saveRpcCallInstanceInfo :: RpcCallInstanceInfo -> JSValue

toDictRpcCallInstanceInfo :: RpcCallInstanceInfo -> [(String, JSValue)]

data InstanceInfo

Constructors

InstanceInfo

Fields

instInfoMemory :: Int

instInfoState :: String

instInfoVcpus :: Int

instInfoTime :: Int

loadInstanceInfo :: JSValue -> Result InstanceInfo

saveInstanceInfo :: InstanceInfo -> JSValue

toDictInstanceInfo :: InstanceInfo -> [(String, JSValue)]

data RpcResultInstanceInfo

Constructors

RpcResultInstanceInfo

Fields

rpcResInstInfoInstInfo :: Maybe InstanceInfo

loadRpcResultInstanceInfo :: JSValue -> Result RpcResultInstanceInfo

saveRpcResultInstanceInfo :: RpcResultInstanceInfo -> JSValue

toDictRpcResultInstanceInfo :: RpcResultInstanceInfo -> [(String, JSValue)]

AllInstancesInfo

data RpcCallAllInstancesInfo

AllInstancesInfo Returns information about all running instances on the given nodes

Constructors

RpcCallAllInstancesInfo

Fields

rpcCallAllInstInfoHypervisors :: [Hypervisor]

loadRpcCallAllInstancesInfo :: JSValue -> Result RpcCallAllInstancesInfo

saveRpcCallAllInstancesInfo :: RpcCallAllInstancesInfo -> JSValue

toDictRpcCallAllInstancesInfo :: RpcCallAllInstancesInfo -> [(String, JSValue)]

data RpcResultAllInstancesInfo

Constructors

RpcResultAllInstancesInfo

Fields

rpcResAllInstInfoInstances :: [(String, InstanceInfo)]

loadRpcResultAllInstancesInfo :: JSValue -> Result RpcResultAllInstancesInfo

saveRpcResultAllInstancesInfo :: RpcResultAllInstancesInfo -> JSValue

toDictRpcResultAllInstancesInfo :: RpcResultAllInstancesInfo -> [(String, JSValue)]

InstanceList

data RpcCallInstanceList

InstanceList Returns the list of running instances on the given nodes.

Constructors

RpcCallInstanceList

Fields

rpcCallInstListHypervisors :: [Hypervisor]

loadRpcCallInstanceList :: JSValue -> Result RpcCallInstanceList

saveRpcCallInstanceList :: RpcCallInstanceList -> JSValue

toDictRpcCallInstanceList :: RpcCallInstanceList -> [(String, JSValue)]

data RpcResultInstanceList

Constructors

RpcResultInstanceList

Fields

rpcResInstListInstances :: [String]

loadRpcResultInstanceList :: JSValue -> Result RpcResultInstanceList

saveRpcResultInstanceList :: RpcResultInstanceList -> JSValue

toDictRpcResultInstanceList :: RpcResultInstanceList -> [(String, JSValue)]

NodeInfo

data RpcCallNodeInfo

NodeInfo Return node information.

Constructors

RpcCallNodeInfo

Fields

rpcCallNodeInfoStorageUnits :: Map String [StorageUnit]

rpcCallNodeInfoHypervisors :: [(Hypervisor, HvParams)]

loadRpcCallNodeInfo :: JSValue -> Result RpcCallNodeInfo

saveRpcCallNodeInfo :: RpcCallNodeInfo -> JSValue

toDictRpcCallNodeInfo :: RpcCallNodeInfo -> [(String, JSValue)]

data StorageInfo

Constructors

StorageInfo

Fields

storageInfoName :: String

storageInfoType :: String

storageInfoStorageFree :: Maybe Int

storageInfoStorageSize :: Maybe Int

loadStorageInfo :: JSValue -> Result StorageInfo

saveStorageInfo :: StorageInfo -> JSValue

toDictStorageInfo :: StorageInfo -> [(String, JSValue)]

data HvInfo

We only provide common fields as described in hv_base.py.

Constructors

HvInfo

Fields

hvInfoMemoryTotal :: Int

hvInfoMemoryFree :: Int

hvInfoMemoryDom0 :: Int

hvInfoCpuTotal :: Int

hvInfoCpuNodes :: Int

hvInfoCpuSockets :: Int

hvInfoCpuDom0 :: Int

loadHvInfo :: JSValue -> Result HvInfo

saveHvInfo :: HvInfo -> JSValue

toDictHvInfo :: HvInfo -> [(String, JSValue)]

data RpcResultNodeInfo

Constructors

RpcResultNodeInfo

Fields

rpcResNodeInfoBootId :: String

rpcResNodeInfoStorageInfo :: [StorageInfo]

rpcResNodeInfoHvInfo :: [HvInfo]

loadRpcResultNodeInfo :: JSValue -> Result RpcResultNodeInfo

saveRpcResultNodeInfo :: RpcResultNodeInfo -> JSValue

toDictRpcResultNodeInfo :: RpcResultNodeInfo -> [(String, JSValue)]

Version

data RpcCallVersion

Query node version.

Constructors

RpcCallVersion

loadRpcCallVersion :: JSValue -> Result RpcCallVersion

saveRpcCallVersion :: RpcCallVersion -> JSValue

toDictRpcCallVersion :: RpcCallVersion -> [(String, JSValue)]

data RpcResultVersion

Query node reply.

Constructors

RpcResultVersion

Fields

rpcResultVersionVersion :: Int

loadRpcResultVersion :: JSValue -> Result RpcResultVersion

saveRpcResultVersion :: RpcResultVersion -> JSValue

toDictRpcResultVersion :: RpcResultVersion -> [(String, JSValue)]

StorageList

data RpcCallStorageList

Constructors

RpcCallStorageList

Fields

rpcCallStorageListSuName :: StorageType

rpcCallStorageListSuArgs :: [String]

rpcCallStorageListName :: String

rpcCallStorageListFields :: [StorageField]

loadRpcCallStorageList :: JSValue -> Result RpcCallStorageList

saveRpcCallStorageList :: RpcCallStorageList -> JSValue

toDictRpcCallStorageList :: RpcCallStorageList -> [(String, JSValue)]

data RpcResultStorageList

Constructors

RpcResultStorageList

Fields

rpcResStorageListStorage :: [[(StorageField, JSValue)]]

loadRpcResultStorageList :: JSValue -> Result RpcResultStorageList

saveRpcResultStorageList :: RpcResultStorageList -> JSValue

toDictRpcResultStorageList :: RpcResultStorageList -> [(String, JSValue)]

TestDelay

data RpcCallTestDelay

Call definition for test delay.

Constructors

RpcCallTestDelay

Fields

rpcCallTestDelayDuration :: Double

loadRpcCallTestDelay :: JSValue -> Result RpcCallTestDelay

saveRpcCallTestDelay :: RpcCallTestDelay -> JSValue

toDictRpcCallTestDelay :: RpcCallTestDelay -> [(String, JSValue)]

data RpcResultTestDelay

Result definition for test delay.

Constructors

RpcResultTestDelay

ExportList

data RpcCallExportList

Call definition for export list.

Constructors

RpcCallExportList

loadRpcCallExportList :: JSValue -> Result RpcCallExportList

saveRpcCallExportList :: RpcCallExportList -> JSValue

toDictRpcCallExportList :: RpcCallExportList -> [(String, JSValue)]

data RpcResultExportList

Result definition for export list.

Constructors

RpcResultExportList

Fields

rpcResExportListExports :: [String]

loadRpcResultExportList :: JSValue -> Result RpcResultExportList

saveRpcResultExportList :: RpcResultExportList -> JSValue

toDictRpcResultExportList :: RpcResultExportList -> [(String, JSValue)]