Ganeti.Runtime

Description

Implementation of the runtime configuration details.

Documentation

data GanetiDaemon

Constructors

GanetiMasterd

GanetiMetad

GanetiNoded

GanetiRapi

GanetiConfd

GanetiWConfd

GanetiKvmd

GanetiLuxid

GanetiMond

data MiscGroup

Constructors

DaemonsGroup

AdminGroup

data GanetiGroup

Constructors

DaemonGroup GanetiDaemon

ExtraGroup MiscGroup

data RuntimeEnts

Constructors

RuntimeEnts

Fields

reUserToUid :: Map GanetiDaemon UserID

reUidToUser :: Map UserID String

reGroupToGid :: Map GanetiGroup GroupID

reGidToGroup :: Map GroupID String

daemonName :: GanetiDaemon -> String

Returns the daemon name for a given daemon.

daemonOnlyOnMaster :: GanetiDaemon -> Bool

Returns whether the daemon only runs on the master node.

daemonLogBase :: GanetiDaemon -> String

Returns the log file base for a daemon.

daemonUser :: GanetiDaemon -> String

Returns the configured user name for a daemon.

daemonGroup :: GanetiGroup -> String

Returns the configured group for a daemon.

data ExtraLogReason

Constructors

AccessLog

ErrorLog

daemonsExtraLogbase :: GanetiDaemon -> ExtraLogReason -> String

Some daemons might require more than one logfile. Specifically, right now only the Haskell http library snap, used by the monitoring daemon, requires multiple log files.

daemonLogFile :: GanetiDaemon -> IO FilePath

Returns the log file for a daemon.

daemonsExtraLogFile :: GanetiDaemon -> ExtraLogReason -> IO FilePath

Returns the extra log files for a daemon.

daemonPidFile :: GanetiDaemon -> IO FilePath

Returns the pid file name for a daemon.

allGroups :: [GanetiGroup]

getEnts :: Error e => ResultT e IO RuntimeEnts

Computes the group/user maps.

verifyDaemonUser :: GanetiDaemon -> RuntimeEnts -> IO ()

Checks whether a daemon runs as the right user.

checkUidMatch :: String -> UserID -> UserID -> IO ()