Ganeti.Ssconf

Description

Implementation of the Ganeti Ssconf interface.

Reading individual ssconf entries

maxFileSize :: Int

sSFilePrefix :: FilePath

ssconf file prefix, re-exported from Constants.

data SSKey

Constructors

SSClusterName

SSClusterTags

SSFileStorageDir

SSSharedFileStorageDir

SSGlusterStorageDir

SSMasterCandidates

SSMasterCandidatesIps

SSMasterCandidatesCerts

SSMasterIp

SSMasterNetdev

SSMasterNetmask

SSMasterNode

SSNodeList

SSNodePrimaryIps

SSNodeSecondaryIps

SSNodeVmCapable

SSOfflineNodes

SSOnlineNodes

SSPrimaryIpFamily

SSInstanceList

SSReleaseVersion

SSHypervisorList

SSMaintainNodeHealth

SSUidPool

SSNodegroups

SSNetworks

SSEnabledUserShutdown

SSSshPorts

SSHvparamsKvm

SSHvparamsXenPvm

SSHvparamsChroot

SSHvparamsXenHvm

SSHvparamsLxc

SSHvparamsFake

sSKeyFromRaw :: forall m. Monad m => String -> m SSKey

sSKeyToRaw :: SSKey -> String

hvparamsSSKey :: Hypervisor -> SSKey

For a given hypervisor get the corresponding SSConf key that contains its parameters.

The corresponding SSKeys are generated automatically by TH, but since we don't have convenient infrastructure for generating this function, it's just manual. All constructors must be given explicitly so that adding another hypervisor will trigger incomplete pattern warning and force the corresponding addition.

keyToFilename

Arguments

:: FilePath
Config path root

-> SSKey
Ssconf key

-> FilePath
Full file name

Convert a ssconf key into a (full) file path.

catchIOErrors :: Maybe a -> IO a -> IO (Result a)

readSSConfFile :: Maybe FilePath -> Maybe String -> SSKey -> IO (Result String)

parseKeyValue :: Monad m => String -> String -> m (String, String)

parseIPFamily :: Int -> Result Family

getPrimaryIPFamily :: Maybe FilePath -> IO (Result Family)

Read the primary IP family.

parseNodesVmCapable :: String -> Result [(String, Bool)]

Parse the nodes vm capable value from a String.

getNodesVmCapable :: Maybe FilePath -> IO (Result [(String, Bool)])

Read and parse the nodes vm capable.

getMasterCandidatesIps :: Maybe FilePath -> IO (Result [String])

Read the list of IP addresses of the master candidates of the cluster.

getMasterNode :: Maybe FilePath -> IO (Result String)

Read the name of the master node.

parseHypervisorList :: String -> Result [Hypervisor]

Parse the list of enabled hypervisors from a String.

getHypervisorList :: Maybe FilePath -> IO (Result [Hypervisor])

Read and parse the list of enabled hypervisors.

parseEnabledUserShutdown :: String -> Result Bool

Parse whether user shutdown is enabled from a String.

getEnabledUserShutdown :: Maybe FilePath -> IO (Result Bool)

Read and parse whether user shutdown is enabled.

Working with the whole ssconf map

newtype SSConf

The data type used for representing the ssconf.

Constructors

SSConf

Fields

getSSConf :: Map SSKey [String]

emptySSConf :: SSConf