Ganeti.Storage.Diskstats.Parser

Description

Diskstats proc file parser

This module holds the definition of the parser that extracts status information about the disks of the system from the procdiskstats file.

Utility functions

skipSpaces :: Parser ()

numberP :: Parser Int

stringP :: Parser String

Parser implementation

oneDiskstatsParser :: Parser Diskstats

diskstatsParser :: Parser [Diskstats]

The parser for a whole diskstatus file.