Ganeti.Storage.Lvm.LVParser

Description

Logical Volumer information parser

This module holds the definition of the parser that extracts status information about the logical volumes (LVs) of the system from the output of the lvs command.

Documentation

lvsSeparator :: Char

Utility functions

skipSpaces :: Parser ()

bytesP :: Parser Int

intP :: Parser Int

stringP :: Parser String

Parser implementation

lvCommand :: String

The command providing the data, in the format the parser expects

lvParams :: [String]

The parameters for getting the data in the format the parser expects

oneLvParser :: Parser LVInfo

lvParser :: Parser [LVInfo]

The parser for a whole diskstatus file.