Ganeti.Storage.Lvm.Types

Description

LVM data types

This module holds the definition of the data types describing the status of the disks according to LVM (and particularly the lvs tool).

Documentation

data LVInfo

This is the format of the report produced by each data collector.

Constructors

LVInfo

Fields

lviUuid :: String

lviName :: String

lviAttr :: String

lviMajor :: Int

lviMinor :: Int

lviKernelMajor :: Int

lviKernelMinor :: Int

lviSize :: Int

lviSegCount :: Int

lviTags :: String

lviModules :: String

lviVgUuid :: String

lviVgName :: String

lviSegtype :: String

lviSegStart :: Int

lviSegStartPe :: Int

lviSegSize :: Int

lviSegTags :: String

lviSegPeRanges :: String

lviDevices :: String

lviInstance :: Maybe String

loadLVInfo :: JSValue -> Result LVInfo

saveLVInfo :: LVInfo -> JSValue