Ganeti.THH.Field

Description

Helpers for creating various kinds of Fields.

They aren't directly needed for the Template Haskell code in Ganeti.THH, so better keep them in a separate module.

Internal functions

numericalReadFn :: JSON a => (String -> Result a) -> [(String, JSValue)] -> JSValue -> Result a

specialNumericalField :: Name -> Field -> Field

Sets the read function to also accept string parsable by the given function.

timeAsDoubleField :: String -> Field

Creates a new mandatory field that reads time as the (floating point) number of seconds since the standard UNIX epoch, and represents it in Haskell as ClockTime.

integralField :: Q Type -> String -> Field

External functions and data types

timeStampFields :: [Field]

Timestamp fields description.

serialFields :: [Field]

Serial number fields description.

uuidFields :: [Field]

UUID fields description.

type TagSet = Set String

Tag set type alias.

tagsFields :: [Field]

Tag field description.

Fields related to POSIX data types

fileModeAsIntField :: String -> Field

Creates a new mandatory field that reads a file mode in the standard POSIX file mode representation. The Haskell type of the field is FileMode.

processIdField :: String -> Field

Creates a new mandatory field that contains a POSIX process ID.