Ganeti.UDSServer

Description

Implementation of the Ganeti Unix Domain Socket JSON server interface.

Utility functions

withTimeout :: Int -> String -> IO a -> IO a

Generic protocol functionality

data RecvResult

Result of receiving a message from the socket.

Constructors

RecvConnClosed
Connection closed

RecvError String
Any other error

RecvOk String
Successfull receive

eOM :: Word8

bEOM :: ByteString

data MsgKeys

Valid keys in the requests and responses.

Constructors

Method

Args

Success

Result

strOfKey :: MsgKeys -> String

The serialisation of MsgKeys into strings in messages.

data ServerConfig

Constructors

ServerConfig

Fields

connPermissions :: FilePermissions

connConfig :: ConnectConfig

data ConnectConfig

Constructors

ConnectConfig

Fields

recvTmo :: Int

sendTmo :: Int

data Client

A client encapsulation. Note that it has separate read and write handle. For sockets it is the same handle. It is required for bi-directional inter-process pipes though.

Constructors

Client

Fields

rsocket :: Handle

wsocket :: Handle

rbuf :: IORef ByteString

clientConfig :: ConnectConfig

data Server

A server encapsulation.

Constructors

Server

Fields

sSocket :: Socket

sPath :: FilePath

serverConfig :: ConnectConfig

Unix sockets

openClientSocket

Arguments

:: Int
connection timeout

-> FilePath
socket path

-> IO Handle

Creates a Unix socket and connects it to the specified path, where timeout specifies the connection timeout.

closeClientSocket :: Handle -> IO ()

Closes the handle. Performing the operation on a handle that has already been closed has no effect; doing so is not an error. All other operations on a closed handle will fail.

openServerSocket :: FilePath -> IO Socket

Creates a Unix socket and binds it to the specified path.

closeServerSocket :: Socket -> FilePath -> IO ()

acceptSocket :: Socket -> IO Handle

Client and server

connectClient

Arguments

:: ConnectConfig
configuration for the client

-> Int
connection timeout

-> FilePath
socket path

-> IO Client

Connects to the master daemon and returns a Client.

connectServer :: ServerConfig -> Bool -> FilePath -> IO Server

Creates and returns a server endpoint.

pipeClient :: ConnectConfig -> IO (Client, Client)

Creates a new bi-directional client pipe. The two returned clients talk to each other through the pipe.

closeServer :: MonadBase IO m => Server -> m ()

Closes a server endpoint.

acceptClient :: Server -> IO Client

Accepts a client

closeClient :: Client -> IO ()

Closes the client socket. Performing the operation on a client that has already been closed has no effect; doing so is not an error. All other operations on a closed client will fail with an exception.

clientToFd :: Client -> IO (Fd, Fd)

Extracts the read (the first) and the write (the second) file descriptor of a client. This closes the underlying Handles, therefore the original client is closed and unusable after the call.

The purpose of this function is to keep the communication channel open, while replacing a Client with some other means.

sendMsg :: Client -> String -> IO ()

Sends a message over a transport.

recvUpdate :: ConnectConfig -> Handle -> ByteString -> IO (ByteString, ByteString)

recvMsg :: Client -> IO String

Waits for a message over a transport.

recvMsgExt :: Client -> IO RecvResult

Extended wrapper over recvMsg.

buildCall

Arguments

:: (JSON mth, JSON args)

=> mth
The method

-> args
The arguments

-> String
The serialized form

Serialize a request to String.

parseCall :: (JSON mth, JSON args) => String -> Result (mth, args)

Parse the required keys out of a call.

buildResponse

Arguments

:: Bool
Success

-> JSValue
The arguments

-> String
The serialized form

Serialize the response to String.

decodeError :: JSValue -> ErrorResult JSValue

parseResponse :: String -> ErrorResult JSValue

Check that luxi responses contain the required keys and that the call was successful.

logMsg :: (Show e, JSON e, MonadLog m) => Handler i m o -> i -> GenericResult e JSValue -> m ()

prepareMsg :: JSON e => GenericResult e JSValue -> (Bool, JSValue)

Processing client requests

type HandlerResult m o = m (Bool, GenericResult GanetiException o)

data Handler i m o

Constructors

Handler

Fields

hParse :: JSValue -> JSValue -> Result i

parses method and its arguments into the input type

hInputLogShort :: i -> String

short description of an input, for the INFO logging level

hInputLogLong :: i -> String

long description of an input, for the DEBUG logging level

hExec :: i -> HandlerResult m o

executes the handler on an input

handleJsonMessage :: (JSON o, Monad m) => Handler i m o -> i -> HandlerResult m JSValue

handleRawMessage :: (JSON o, MonadLog m) => Handler i m o -> String -> m (Bool, String)

isRisky :: RecvResult -> Bool

handleClient :: (JSON o, MonadBase IO m, MonadLog m) => Handler i m o -> Client -> m Bool

clientLoop :: (JSON o, MonadBase IO m, MonadLog m) => Handler i m o -> Client -> m ()

listener :: (JSON o, MonadBaseControl IO m, MonadLog m) => Handler i m o -> Server -> m ()

Main listener loop: accepts clients, forks an I/O thread to handle that client.