Ganeti.Utils.Statistics

Description

Utility functions for statistical accumulation.

Documentation

class Show c => Stat s c | c -> s where

Typeclass describing necessary statistical accumulations functions. Types defining an instance of Stat behave as if the given statistics were computed on the list of values, but they allow a potentially more efficient update of a given value. c is the statistical accumulation data type itself while s is a type of spread values used to calculate a statistics. s defined as a type dependent from c in order to pretend ambiguity.

Methods

calculate :: [s] -> c

Calculate a statistics from the spread values list.

update :: c -> s -> s -> c

In a given statistics replace on value by another. This will only give meaningful results, if the original value was actually part of the statistics.

getValue :: c -> Double

Obtain the value of a statistics.

data SumStat

Type of statistical accumulations representing simple sum of values

Constructors

SumStat Double

data StdDevStat

Type of statistical accumulations representing values standard deviation

Constructors

StdDevStat Double Double Double

data MapStat

Type of statistical accumulations representing the amount of instances per each tags pair. See Also TagTagMap documentation.

Constructors

MapStat TagTagMap

type TagTagMap = Map (String, String) Int

Type to store the number of instances for each exclusion and location pair. This is necessary to calculate second component of location score.

data MapData

Data type used to store spread values of type TagTagMap. This data type is introduced only to defin an instance of Stat for TagTagMap.

Constructors

MapData TagTagMap

mapTmpToMap :: [MapData] -> [TagTagMap]

toDouble :: Integral a => a -> Double

Converts Integral types to Double. It's usefull than it's not enough type information in the expression to call fromIntegral directly.