Ganeti.WConfd.ConfigState

Description

Pure functions for manipulating the configuration state.

Documentation

data ConfigState

In future this data type will include the current configuration (ConfigData) and the last FStat of its file.

Constructors

ConfigState

Fields

csConfigData :: ConfigData

csConfigDataL :: Iso' ConfigState ConfigData

mkConfigState :: ConfigData -> ConfigState

Creates a new configuration state. This method will expand as more fields are added to ConfigState.

bumpSerial :: (SerialNoObjectL a, TimeStampObjectL a) => ClockTime -> a -> a

needsFullDist :: ConfigState -> ConfigState -> Bool

Given two versions of the configuration, determine if its distribution needs to be fully commited before returning the corresponding call to WConfD.