Ganeti.WConfd.Language

Description

Function related to serialisation of WConfD requests

Serialisation related to locking

data LockRequestType

Operation to be carried out on a lock (request exclusive/shared ownership, or release).

Constructors

ReqExclusive

ReqShared

ReqRelease

type GanetiLockRequest = [(GanetiLocks, LockRequestType)]

The type describing how lock update requests are passed over the wire.

toLockRequest :: (GanetiLocks, LockRequestType) -> LockRequest GanetiLocks

fromGanetiLockRequest :: GanetiLockRequest -> [LockRequest GanetiLocks]

From a GanetiLockRequest obtain a list of Ganeti.Lock.Allocation.LockRequest, suitable to updateLocks.