Ganeti.WConfd.TempRes

Description

Pure functions for manipulating reservations of temporary objects

NOTE: Reservations aren't released specifically, they're just all released at the end of a job. This could be improved in the future.

The main reservation state

Aliases to make types more meaningful:

type NodeUUID = String

type InstanceUUID = String

type DiskUUID = String

type NetworkUUID = String

type DRBDMinor = Int

type DRBDMap = Map NodeUUID (Map DRBDMinor DiskUUID)

A map of the usage of DRBD minors

type DRBDMap' = Map NodeUUID (Map DRBDMinor [DiskUUID])

The state data structure

data IPv4ResAction

Types of IPv4 reservation actions.

Constructors

IPv4Reserve

IPv4Release

data IPv4Reservation

The values stored in the IPv4 reservation table.

Constructors

IPv4Res

Fields

ipv4ResAction :: IPv4ResAction

ipv4ResNetwork :: NetworkUUID

ipv4ResAddr :: Ip4Address

newtype TempRes j a

A polymorphic data structure for managing temporary resources assigned to jobs.

Constructors

TempRes

Fields

getTempRes :: MultiMap j a

mkTempRes :: MultiMap j a -> TempRes j a

Create a temporary reservations from a given multi-map.

data TempResState

The state of the temporary reservations

Constructors

TempResState

Fields

trsDRBD :: DRBDMap

trsMACs :: TempRes ClientId MAC

trsDRBDSecrets :: TempRes ClientId DRBDSecret

trsLVs :: TempRes ClientId LogicalVolume

trsIPv4s :: TempRes ClientId IPv4Reservation

loadTempResState :: JSValue -> Result TempResState

saveTempResState :: TempResState -> JSValue

emptyTempResState :: TempResState

trsMACsL :: Lens' TempResState (TempRes ClientId MAC)

trsLVsL :: Lens' TempResState (TempRes ClientId LogicalVolume)

trsIPv4sL :: Lens' TempResState (TempRes ClientId IPv4Reservation)

trsDRBDSecretsL :: Lens' TempResState (TempRes ClientId DRBDSecret)

trsDRBDL :: Lens' TempResState DRBDMap

Utility functions

resError :: MonadError GanetiException m => String -> m a

toResError :: MonadError GanetiException m => m a -> m a

filterNested :: (Ord a, Ord b) => (c -> Bool) -> Map a (Map b c) -> Map a (Map b c)

maybeLens :: (Monoid a, Monoid b, Eq b) => Lens s t (Maybe a) (Maybe b) -> Lens s t a b

DRBD functions

computeDRBDMap' :: MonadError GanetiException m => ConfigData -> TempResState -> m DRBDMap'

Compute the map of used DRBD minor/nodes, including possible duplicates. An error is returned if the configuration isn't consistent (for example if a referenced disk is missing etc.).

computeDRBDMap :: MonadError GanetiException m => ConfigData -> TempResState -> m DRBDMap

Compute the map of used DRBD minor/nodes. Report any duplicate entries as an error.

Unlike computeDRBDMap', includes entries for all nodes, even if empty.

allocateDRBDMinor :: (MonadError GanetiException m, MonadState TempResState m) => ConfigData -> DiskUUID -> [NodeUUID] -> m [DRBDMinor]

releaseDRBDMinors :: MonadState TempResState m => DiskUUID -> m ()

Other temporary resources

isReserved :: (Ord a, Ord j) => a -> TempRes j a -> Bool

Tests if a given value is reserved for a given job.

reserve :: (MonadError GanetiException m, Show a, Ord a, Ord j) => j -> a -> TempRes j a -> m (TempRes j a)

Tries to reserve a given value for a given job.

dropReservationsFor :: (Ord a, Ord j) => j -> TempRes j a -> TempRes j a

reservedFor :: (Ord a, Ord j) => j -> TempRes j a -> Set a

reserved :: (Ord a, Ord j) => TempRes j a -> Set a

withReserved :: (MonadError GanetiException m, Show a, Ord a, Ord j) => j -> (Set a -> m a) -> TempRes j a -> m (a, TempRes j a)

generate :: (MonadError GanetiException m, Show a, Ord a, Ord j) => j -> Set a -> m (Maybe a) -> TempRes j a -> m (a, TempRes j a)

generateRand :: (MonadError GanetiException m, Show a, Ord a, Ord j, RandomGen g) => g -> j -> Set a -> (g -> (Maybe a, g)) -> TempRes j a -> m (a, TempRes j a)

stateM :: MonadState s m => (s -> m (a, s)) -> m a

modifyM :: MonadState s m => (s -> m s) -> m ()

Functions common to all reservations

dropAllReservations :: ClientId -> State TempResState ()

Removes all resources reserved by a given job.

If a new reservation resource type is added, it must be added here as well.

lookupNetwork :: MonadError GanetiException m => ConfigData -> NetworkUUID -> m Network

IDs

MAC addresses

generateMAC :: (RandomGen g, MonadError GanetiException m, Functor m) => g -> ClientId -> Maybe NetworkUUID -> ConfigData -> StateT TempResState m MAC

reserveMAC :: (MonadError GanetiException m, MonadState TempResState m, Functor m) => ClientId -> MAC -> ConfigData -> m ()

DRBD secrets

generateDRBDSecret :: (RandomGen g, MonadError GanetiException m, Functor m) => g -> ClientId -> ConfigData -> StateT TempResState m DRBDSecret

LVs

reserveLV :: (MonadError GanetiException m, MonadState TempResState m, Functor m) => ClientId -> LogicalVolume -> ConfigData -> m ()

IPv4 addresses

usedIPv4Addrs :: NetworkUUID -> Set IPv4Reservation -> Set Ip4Address

reserveIp

Arguments

:: (MonadError GanetiException m, MonadState TempResState m, Functor m)

=> ClientId

-> NetworkUUID

-> Ip4Address

-> Bool
whether to check externally reserved IPs

-> ConfigData

-> m ()

Reserve a given IPv4 address for use by an instance.

releaseIp :: (MonadError GanetiException m, MonadState TempResState m, Functor m) => ClientId -> NetworkUUID -> Ip4Address -> m ()

Give a specific IP address back to an IP pool. The IP address is returned to the IP pool designated by network id and marked as reserved.

generateIp :: (MonadError GanetiException m, MonadState TempResState m, Functor m) => ClientId -> NetworkUUID -> ConfigData -> m Ip4Address

commitIp :: (MonadError GanetiException m, Functor m) => IPv4Reservation -> ConfigData -> m ConfigData

commitReleaseIp :: (MonadError GanetiException m, Functor m) => NetworkUUID -> Ip4Address -> ConfigData -> m ConfigData

Immediately release an IP address, without using the reservations pool.

commitReservedIps :: (MonadError GanetiException m, Functor m, MonadLog m) => ClientId -> TempResState -> ConfigData -> m ConfigData

Commit all reserved/released IP address to an IP pool. The IP addresses are taken from the network's IP pool and marked as reserved/free for instances.

Note that the reservations are kept, they are supposed to be cleaned when a job finishes.

listReservedIps :: ClientId -> TempResState -> Set IPv4Reservation