Test.AutoConf

Description

Unittests for AutoConf

Documentation

isFilePath :: String -> Bool

'isFilePath x' tests whether x is a valid filepath

A valid filepath must be absolute and must not contain commas.

isGntScript :: String -> Bool

'isGntScript x' tests whether x is a valid Ganeti script

A valid Ganeti script is prefixed by gnt- and the rest of the String contains only alphabetic Chars.

isGroup :: String -> Bool

'isGroup x' tests whether x is a valid group name

A valid group name name is an alphabetic String possibly containing -.

isProgram :: String -> Bool

'isProgram x' tests whether x is a valid program name

A valid program name is an alphabetic String.

isUser :: String -> Bool

'isUser x' tests whether x is a valid username

See isGroup.

case_versionSuffix :: Assertion

case_localstatedir :: Assertion

case_sysconfdir :: Assertion

case_sshConfigDir :: Assertion

case_sshLoginUser :: Assertion

case_sshConsoleUser :: Assertion

case_exportDir :: Assertion

case_osSearchPath :: Assertion

case_esSearchPath :: Assertion

case_xenBootloader :: Assertion

case_xenConfigDir :: Assertion

case_xenKernel :: Assertion

case_xenInitrd :: Assertion

case_kvmKernel :: Assertion

case_iallocatorSearchPath :: Assertion

case_kvmPath :: Assertion

case_ipPath :: Assertion

case_socatPath :: Assertion

case_toolsdir :: Assertion

case_gntScripts :: Assertion

case_htoolsProgs :: Assertion

case_pkglibdir :: Assertion

case_sharedir :: Assertion

case_versionedsharedir :: Assertion

case_drbdBarriers :: Assertion

case_syslogUsage :: Assertion

case_daemonsGroup :: Assertion

case_adminGroup :: Assertion

case_masterdUser :: Assertion

case_masterdGroup :: Assertion

case_rapiUser :: Assertion

case_rapiGroup :: Assertion

case_confdUser :: Assertion

case_confdGroup :: Assertion

case_luxidUser :: Assertion

case_luxidGroup :: Assertion

case_nodedUser :: Assertion

case_nodedGroup :: Assertion

case_mondUser :: Assertion

case_mondGroup :: Assertion

case_diskSeparator :: Assertion

case_qemuimgPath :: Assertion

testAutoConf :: Test