Test.Ganeti.HTools.Instance

Description

Unittests for ganeti-htools.

Arbitrary instances

genInstanceSmallerThan :: Int -> Int -> Int -> Maybe Int -> Gen Instance

genInstanceSmallerThanNode :: Node -> Gen Instance

Generates an instance smaller than a node.

genInstanceMaybeBiggerThanNode :: Node -> Gen Instance

Generates an instance possibly bigger than a node.

genInstanceOnNodeList :: List -> Gen Instance

Generates an instance with nodes on a node list. The following rules are respected: 1. The instance is never bigger than its primary node 2. If possible the instance has different pnode and snode 3. Else disk templates which require secondary nodes are disabled

genInstanceList :: Gen Instance -> Gen List

Generates an instance list given an instance generator.

Test cases

prop_creat :: Instance -> Property

prop_setIdx :: Instance -> Idx -> Property

prop_setName :: Instance -> String -> Bool

prop_setAlias :: Instance -> String -> Bool

prop_setPri :: Instance -> Ndx -> Property

prop_setSec :: Instance -> Ndx -> Property

prop_setBoth :: Instance -> Ndx -> Ndx -> Bool

prop_shrinkMG :: Instance -> Property

prop_shrinkMF :: Instance -> Property

prop_shrinkCG :: Instance -> Property

prop_shrinkCF :: Instance -> Property

prop_shrinkDG :: Instance -> Property

prop_shrinkDF :: Instance -> Property

prop_setMovable :: Instance -> Bool -> Property

testHTools_Instance :: Test