Test.Ganeti.HTools.Node

Description

Unittests for ganeti-htools.

Arbitrary instances

genNode

Arguments

:: Maybe Int
Minimum node size in terms of units

-> Maybe Int
Maximum node size (when Nothing, bounded just by the max... constants)

-> Gen Node

Generates an arbitrary node based on sizing information.

genOnlineNode :: Gen Node

Helper function to generate a sane node.

genExclStorNode :: Gen Node

genMaybeExclStorNode :: Gen Node

genNodeList :: Gen Node -> Gen List

Node list generator. Given a node generator, create a random length node list. Note that real clusters always have at least one node, so we don't generate empty node lists here.

genUniqueNodeList :: Gen Node -> Gen (List, NameAssoc)

Node list generator where node names are unique

genNodeGraph :: Gen (Maybe Graph, List, List)

Test cases

prop_setAlias :: Node -> String -> Bool

prop_setOffline :: Node -> Bool -> Property

prop_setXmem :: Node -> Int -> Property

prop_setMcpu :: Node -> Double -> Property

prop_setFmemGreater :: Node -> Int -> Property

prop_setFmemExact :: Node -> Property

prop_addPri_NoN1Fail :: Property

prop_addPriFM :: Node -> Instance -> Property

prop_addPriFD :: Node -> Instance -> Property

hasInstTooManySpindles :: Instance -> Int -> Bool

prop_addPriFS :: Instance -> Property

prop_addPriFC :: Property

prop_addSec :: Node -> Instance -> Int -> Property

prop_addOfflinePri :: NonNegative Int -> NonNegative Int -> Property

prop_addOfflineSec :: NonNegative Int -> NonNegative Int -> Ndx -> Property

prop_rMem :: Instance -> Property

prop_setMdsk :: Node -> SmallRatio -> Bool

prop_tagMaps_idempotent :: Property

prop_tagMaps_reject :: Property

prop_showField :: Node -> Property

prop_computeGroups :: [Node] -> Bool

prop_addPri_idempotent :: Property

prop_addSec_idempotent :: Property

case_emptyNodeList :: Assertion

prop_numVertices :: Property

prop_numEdges :: Property

prop_nodeGraphIsColorable :: Property

prop_instanceIsEdge :: Property

prop_edgeIsInstance :: Property

testHTools_Node :: Test

List of tests for the Node module.