Test.Ganeti.Objects

Description

Unittests for ganeti-htools.

Arbitrary instances

genInstWithNets :: [String] -> Gen Instance

Generates an instance that is connected to the given networks and possibly some other networks

genInst :: Gen Instance

Generates an instance that is connected to some networks

enhanceInstWithNets :: Instance -> [String] -> Gen Instance

genDiskWithChildren :: Int -> Gen Disk

genDisk :: Gen Disk

genValidNetwork :: Gen Network

Generates a network instance with minimum netmasks of /24. Generating bigger networks slows down the tests, because long bit strings are generated for the reservations.

genBitString :: Int -> Gen AddressPool

genBitStringMaxLen :: Int -> Gen AddressPool

Generate an arbitrary string consisting of '0' and '1' of the maximum given length.

genEmptyCluster :: Int -> Gen ConfigData

Generator for config data with an empty cluster (no instances), with N defined nodes.

genConfigDataWithNetworks :: ConfigData -> Gen ConfigData

FIXME: make an even simpler base version of creating a cluster.

Generates config data with a couple of networks.

Test properties

prop_fillDict :: [(Int, Int)] -> [(Int, Int)] -> Property

prop_LogicalVolume_serialisation :: LogicalVolume -> Property

prop_LogicalVolume_deserialisationFail :: Property

prop_Disk_serialisation :: Disk -> Property

prop_Disk_array_serialisation :: Disk -> Property

prop_Node_serialisation :: Node -> Property

prop_Inst_serialisation :: Instance -> Property

prop_AddressPool_serialisation :: AddressPool -> Property

prop_Network_serialisation :: Network -> Property

prop_FilterAction_serialisation :: FilterAction -> Property

prop_FilterPredicate_serialisation :: FilterPredicate -> Property

prop_Config_serialisation :: Property

casePyCompatNetworks :: Assertion

getNetworkProperties :: Network -> (Int, Int, Network)

casePyCompatNodegroups :: Assertion

genNodeGroup :: Gen NodeGroup

prop_ip4AddressAsNum :: Ip4Address -> Property

prop_ip4AddressToNumber :: Word32 -> Property

caseNextIp4Address :: Assertion

casePyCompatInstances :: Assertion

mkLIDPlain :: String -> String -> DiskLogicalId

caseIncludeLogicalIdPlain :: Assertion

caseIncludeLogicalIdDrbd :: Assertion

caseNotIncludeLogicalIdPlain :: Assertion

testObjects :: Test