Test.Ganeti.Query.Language

Description

Unittests for ganeti-htools.

Documentation

genFilter :: Gen (Filter FilterField)

Custom Filter generator (top-level), which enforces a (sane) limit on the depth of the generated filters.

genFilter' :: Int -> Gen (Filter FilterField)

genJSValue :: Gen JSValue

Generates an arbitrary JSValue. We do this via a function a not via arbitrary instance since that would require us to define an arbitrary for JSValue, which can be recursive, entering the usual problems with that; so we only generate the base types, not the recursive ones, and not JSNull, which we can't use in a RSNormal ResultEntry.

genResultEntry :: Gen ResultEntry

prop_filter_serialisation :: Property

prop_filterregex_instances :: FilterRegex -> Property

prop_resultstatus_serialisation :: ResultStatus -> Property

prop_fieldtype_serialisation :: FieldType -> Property

prop_fielddef_serialisation :: FieldDefinition -> Property

prop_resultentry_serialisation :: Property

prop_fieldsresult_serialisation :: Property

prop_itemtype_serialisation :: ItemType -> Property

testQuery_Language :: Test