Test.Ganeti.SlotMap

Description

Unittests for the SlotMap.

Documentation

genSlotLimit :: Gen Int

Generates a number typical for the limit of a Slot. Useful for constructing resource bounds when not directly constructing the relevant Slots.

genSlotCount :: Gen Int

resampleFittingSlot :: Slot -> Gen Slot

type TestKey = String

genTestKey :: Gen TestKey

Generates short strings used as SlotMap keys.

We limit ourselves to a small set of key strings with high probability to increase the chance that SlotMaps actually have more than one slot taken.

listSizeGen :: Gen Int

genSlotMap :: Ord a => Gen a -> Gen (SlotMap a)

genCountMap :: Ord a => Gen a -> Gen (CountMap a)

overfullKeys :: Ord a => SlotMap a -> Set a

Tells which keys of a SlotMap are overfull.

genFittingSlotMap :: Ord a => Gen a -> Gen (SlotMap a)

Test cases

case_isOverfull :: Assertion

case_occupySlots_examples :: Assertion

keyUnion :: Ord a => Map a b -> Map a c -> Set a

prop_occupySlots :: Property

case_hasSlotsFor_examples :: Assertion

prop_hasSlotsFor_fitting :: Property

prop_hasSlotsFor :: Property

testSlotMap :: Test