Test.Ganeti.TestHelper

Description

Unittest helpers for TemplateHaskell components.

Documentation

propPrefix :: String

casePrefix :: String

case2Pfx :: String

simplifyName :: String -> String -> String

runProp :: Testable prop => String -> prop -> Test

runCase :: String -> Assertion -> Test

run :: Name -> Q Exp

mapSlashes :: String -> String

testSuite :: String -> [Name] -> Q [Dec]

Builds a test suite.

mkConsArbitrary :: (Name, [a]) -> Exp

conInfo :: Con -> (Name, [Type])

mkRegularArbitrary :: Name -> [Con] -> Q [Dec]

genArbitrary :: Name -> Q [Dec]

Builds a default Arbitrary instance for a type. This requires that all members are of types that already have Arbitrary instances, and that the arbitrary instances are well behaved (w.r.t. recursive data structures, or similar concerns). In that sense, this is not appropriate for all data types, just those that are simple but very repetitive or have many simple fields.