Ganeti.BasicTypes

Documentation

data GenericResult a b

Generic monad for our error handling mechanisms.

Constructors

Bad a

Ok b

genericResult :: (a -> c) -> (b -> c) -> GenericResult a b -> c

Sum type structure of GenericResult.

type Result = GenericResult String

Type alias for a string Result.

class FromString a where

Type class for things that can be built from strings.

Methods

mkFromString :: String -> a

newtype ResultT a m b

This is a monad transformation for Result. It's implementation is based on the implementations of MaybeT and ErrorT.

Constructors

ResultT

Fields

runResultT :: m (GenericResult a b)

resultT :: Monad m => GenericResult a b -> ResultT a m b

Lift a Result value to a ResultT.

isOk :: GenericResult a b -> Bool

Simple checker for whether a GenericResult is OK.

isBad :: GenericResult a b -> Bool

Simple checker for whether a GenericResult is a failure.

eitherToResult :: Either a b -> GenericResult a b

Converter from Either to GenericResult.

annotateResult :: String -> Result a -> Result a

Annotate a Result with an ownership information.

iterateOk :: (a -> GenericResult b a) -> a -> [a]

Iterate while Ok.

Misc functionality

select

Arguments

:: a
default result

-> [(Bool, a)]
list of "condition, result"

-> a
first result which has a True condition, or default

Return the first result with a True condition, or the default otherwise.

runListHead :: a -> (b -> a) -> [b] -> a

Apply a function to the first element of a list, return the default value, if the list is empty. This is just a convenient combination of maybe and listToMaybe.

Lookup of partial names functionality

data MatchPriority

The priority of a match in a lookup result.

Constructors

ExactMatch

MultipleMatch

PartialMatch

FailMatch

data LookupResult

The result of a name lookup in a list.

Constructors

LookupResult

Fields

lrMatchPriority :: MatchPriority

The result type | Matching value (for ExactMatch, PartialMatch), Lookup string otherwise

lrContent :: String

prefixMatch

Arguments

:: String
Lookup

-> String
Full name

-> Bool
Whether there is a prefix match

Check for prefix matches in names. Implemented in Ganeti core utils.text.MatchNameComponent as the regexp r^%s(..*)?$ % re.escape(key)

goodMatchPriority :: MatchPriority -> Bool

Is the lookup priority a good one?

goodLookupResult :: LookupResult -> Bool

Is the lookup result an actual match?

compareNameComponent

Arguments

:: String
Canonical (target) name

-> String
Partial (lookup) name

-> LookupResult
Result of the lookup

Compares a canonical name and a lookup string.

chooseLookupResult :: String -> String -> LookupResult -> LookupResult

lookupName

Arguments

:: [String]
List of keys

-> String
Lookup string

-> LookupResult
Result of the lookup

Find the canonical name for a lookup string in a list of names.

newtype ListSet a

Wrapper for a Haskell Set

This type wraps a Set and it is used in the Haskell to Python opcode generation to transform a Haskell Set into a Python list without duplicate elements.

Constructors

ListSet

Fields

unListSet :: Set a

emptyListSet :: ListSet a