Ganeti.Confd.ClientFunctions

Description

Some utility functions, based on the Confd client, providing data in a ready-to-use way.

Documentation

getInstances :: String -> Maybe String -> Maybe Int -> IO (Result ([Instance], [Instance]))

Get the list of instances the given node is ([primary], [secondary]) for. The server address and the server port parameters are mainly intended for testing purposes. If they are Nothing, the default values will be used.