Ganeti.Config

Description

Implementation of the Ganeti configuration database.

Documentation

type LinkIpMap = Map String (Map String String)

Type alias for the link and ip map.

class NdParamObject a where

Type class denoting objects which have node parameters.

Methods

getNdParamsOf :: ConfigData -> a -> Maybe FilledNDParams

readConfig :: FilePath -> IO String

parseConfig :: String -> Result ConfigData

loadConfig :: FilePath -> IO (Result ConfigData)

Wrapper over readConfig and parseConfig.

Query functions

computeDiskNodes :: Disk -> Set String

instDiskNodes :: Instance -> Set String

instNodes :: Instance -> Set String

Computes all nodes of an instance.

instSecondaryNodes :: Instance -> Set String

getNodeInstances :: ConfigData -> String -> ([Instance], [Instance])

Get instances of a given node. The node is specified through its UUID.

getNodeRole :: ConfigData -> Node -> NodeRole

Computes the role of a node.

getDefaultNicLink :: ConfigData -> String

Returns the default cluster link.

getDefaultHypervisor :: ConfigData -> Hypervisor

Returns the default cluster hypervisor.

getInstancesIpByLink :: LinkIpMap -> String -> [String]

Returns instances of a given link.

getItem :: String -> String -> Map String a -> ErrorResult a

getNode :: ConfigData -> String -> ErrorResult Node

Looks up a node by name or uuid.

getInstance :: ConfigData -> String -> ErrorResult Instance

Looks up an instance by name or uuid.

getGroup :: ConfigData -> String -> ErrorResult NodeGroup

Looks up a node group by name or uuid.

getGroupNdParams :: ConfigData -> NodeGroup -> FilledNDParams

Computes a node group's node params.

getGroupIpolicy :: ConfigData -> NodeGroup -> FilledIPolicy

Computes a node group's ipolicy.

getGroupDiskParams :: ConfigData -> NodeGroup -> DiskParams

Computes a group's (merged) disk params.

getGroupNodes :: ConfigData -> String -> [Node]

Get nodes of a given node group.

getGroupInstances :: ConfigData -> String -> ([Instance], [Instance])

Get (primary, secondary) instances of a given node group.

getNetwork :: ConfigData -> String -> ErrorResult Network

Looks up a network. If looking up by uuid fails, we look up by name.

getInstPrimaryNode :: ConfigData -> String -> ErrorResult Node

Looks up an instance's primary node.

getDrbdMinorsForNode :: String -> Disk -> [(Int, String)]

rolePrimary :: String

roleSecondary :: String

getInstMinorsForNode

Arguments

:: String
The UUID of a node.

-> Instance

-> [(String, Int, String, String, String, String)]

Gets the list of DRBD minors for an instance that are related to a given node.

buildLinkIpInstnameMap :: ConfigData -> LinkIpMap

Builds link -> ip -> instname map.

TODO: improve this by splitting it into multiple independent functions:

abstract the "fetch instance with filled params" functionality

abstsract the [instance] -> [(nic, instance_name)] part

etc.

getGroupOfNode :: ConfigData -> Node -> Maybe NodeGroup

Returns a node's group, with optional failure if we can't find it (configuration corrupt).

getNodeNdParams :: ConfigData -> Node -> Maybe FilledNDParams

Returns a node's ndparams, filled.