Ganeti.ConstantUtils

Description

ConstantUtils contains the helper functions for constants

This module cannot be merged with Utils because it would create a circular dependency if imported, for example, from Constants.

Documentation

newtype PythonChar

PythonChar wraps a Python char

Constructors

PythonChar

Fields

unPythonChar :: Char

data PythonNone

PythonNone wraps Python None

Constructors

PythonNone

newtype FrozenSet a

FrozenSet wraps a Haskell Set

See PyValue instance for FrozenSet.

Constructors

FrozenSet

Fields

unFrozenSet :: Set a

mkSet :: Ord a => [a] -> FrozenSet a

toList :: FrozenSet a -> [a]

union :: Ord a => FrozenSet a -> FrozenSet a -> FrozenSet a

difference :: Ord a => FrozenSet a -> FrozenSet a -> FrozenSet a

data Protocol

Protocol represents the protocols used by the daemons

Constructors

Tcp

Udp

exitFailure :: Int

Failure exit code

These are defined here and not in Constants together with the other exit codes in order to avoid a circular dependency between Constants and Runtime

devConsole :: String

Console device

This is defined here and not in Constants order to avoid a circular dependency between Constants and Logging

randomUuidFile :: String

Random uuid generator

This is defined here and not in Constants order to avoid a circular dependendy between Constants and Types

Priority levels

priorityLow :: Int

priorityNormal :: Int

priorityHigh :: Int

buildVersion :: Int -> Int -> Int -> Int

Calculates int version number from major, minor and revision numbers.

confdProtocolVersion :: Int

Confd protocol version

This is defined here in order to avoid a circular dependency between Types and Constants.

Confd request query fields

confdReqqLink :: String

confdReqqIp :: String

confdReqqIplist :: String

confdReqqFields :: String

ISpec

ispecMemSize :: String

ispecCpuCount :: String

ispecDiskCount :: String

ispecDiskSize :: String

ispecNicCount :: String

ispecSpindleUse :: String

ispecsMinmax :: String

ispecsStd :: String

ipolicyDts :: String

ipolicyVcpuRatio :: String

ipolicySpindleRatio :: String

ipolicyDefaultsVcpuRatio :: Double

ipolicyDefaultsSpindleRatio :: Double