Ganeti.DataCollectors.CLI

Description

Implementation of DataCollectors CLI functions.

This module holds the common command-line related functions for the collector binaries.

Data types

data Options

Command line options structure.

Constructors

Options

Fields

optShowHelp :: Bool

Just show the help

optShowComp :: Bool

Just show the completion info

optShowVer :: Bool

Just show the program version

optDrbdStatus :: Maybe FilePath

Path to the file containing DRBD status information

optDrbdPairing :: Maybe FilePath

Path to the file containing pairings between instances and DRBD minors

optNode :: Maybe String

Info are requested for this node

optConfdAddr :: Maybe String

IP address of the Confd server

optConfdPort :: Maybe Int

The port of the Confd server to connect to

optInputFile :: Maybe FilePath

Path to the file containing the information to be parsed

optInstances :: Maybe FilePath

Path to the file contained a serialized list of instances as in: ([Primary], [Secondary])

defaultOptions :: Options

Default values for the command line options.

type OptType = GenericOptType Options

Abbreviation for the option type.

Command line options

oDrbdPairing :: OptType

oDrbdStatus :: OptType

oNode :: OptType

oConfdAddr :: OptType

oConfdPort :: OptType

oInputFile :: OptType

oInstances :: OptType

genericOptions :: [GenericOptType Options]

Generic options.