Ganeti.DataCollectors.Types

Description

Implementation of the Ganeti data collector types.

Documentation

data DCReport

This is the format of the report produced by each data collector.

Constructors

DCReport

Fields

dcReportName :: String

dcReportVersion :: String

dcReportFormatVersion :: Int

dcReportTimestamp :: Integer

dcReportData :: JSValue

loadDCReport :: JSValue -> Result DCReport

saveDCReport :: DCReport -> JSValue

toDictDCReport :: DCReport -> [(String, JSValue)]

buildReport :: String -> Maybe String -> Int -> JSValue -> IO DCReport

Utility function for building a report automatically adding the current timestamp (rounded up to seconds). If the version is not specified, it will be set to the value indicating a builtin collector.