Ganeti.JQueue

Description

Implementation of the job queue.

Data types

type Timestamp = (Int, Int)

The ganeti queue timestamp type

noTimestamp :: Timestamp

Missing timestamp type.

data InputOpCode

An input opcode.

Constructors

ValidOpCode MetaOpCode
OpCode was parsed successfully

InvalidOpCode JSValue
Invalid opcode

invalidOp :: String

extractOpSummary :: InputOpCode -> String

Tries to extract the opcode summary from an InputOpCode. This duplicates some functionality from the opSummary function in Ganeti.OpCodes.

data QueuedOpCode

Constructors

QueuedOpCode

Fields

qoInput :: InputOpCode

qoStatus :: OpStatus

qoResult :: JSValue

qoLog :: [(Int, Timestamp, ELogType, JSValue)]

qoPriority :: Int

qoStartTimestamp :: Maybe Timestamp

qoExecTimestamp :: Maybe Timestamp

qoEndTimestamp :: Maybe Timestamp

loadQueuedOpCode :: JSValue -> Result QueuedOpCode

saveQueuedOpCode :: QueuedOpCode -> JSValue

toDictQueuedOpCode :: QueuedOpCode -> [(String, JSValue)]

data QueuedJob

Constructors

QueuedJob

Fields

qjId :: JobId

qjOps :: [QueuedOpCode]

qjReceivedTimestamp :: Maybe Timestamp

qjStartTimestamp :: Maybe Timestamp

qjEndTimestamp :: Maybe Timestamp

loadQueuedJob :: JSValue -> Result QueuedJob

saveQueuedJob :: QueuedJob -> JSValue

toDictQueuedJob :: QueuedJob -> [(String, JSValue)]

jobFilePrefix :: String

jobFileName :: JobId -> FilePath

Computes the filename for a given job ID.

parseJobFileId :: Monad m => FilePath -> m JobId

liveJobFile :: FilePath -> JobId -> FilePath

Computes the full path to a live job.

archivedJobFile :: FilePath -> JobId -> FilePath

Computes the full path to an archives job. BROKEN.

opStatusToJob :: OpStatus -> JobStatus

calcJobStatus :: QueuedJob -> JobStatus

Computes a queued job's status.

opStatusFinalized :: OpStatus -> Bool

Determine whether an opcode status is finalized.

calcJobPriority :: QueuedJob -> Int

Compute a job's priority.

ignoreIOError :: a -> Bool -> String -> IOError -> IO a

allArchiveDirs :: FilePath -> IO [FilePath]

determineJobDirectories :: FilePath -> Bool -> IO [FilePath]

Build list of directories containing job files. Note: compared to the Python version, this doesn't ignore a potential lost+found file.

sequencer :: [Either IOError [JobId]] -> Either IOError [[JobId]]

seqFolder :: Either IOError [[JobId]] -> Either IOError [JobId] -> Either IOError [[JobId]]

getJobIDs :: [FilePath] -> IO (Either IOError [JobId])

Computes the list of all jobs in the given directories.

sortJobIDs :: [JobId] -> [JobId]

Sorts the a list of job IDs.

getDirJobIDs :: FilePath -> IO (Either IOError [JobId])

readJobDataFromDisk :: FilePath -> Bool -> JobId -> IO (Maybe (String, Bool))

noSuchJob :: Result (QueuedJob, Bool)

Failed to load job error.

loadJobFromDisk :: FilePath -> Bool -> JobId -> IO (Result (QueuedJob, Bool))

Loads a job from disk.