Ganeti.Luxi

Description

Implementation of the Ganeti LUXI interface.

Utility functions

withTimeout :: Int -> String -> IO a -> IO a

Generic protocol functionality

data RecvResult

Result of receiving a message from the socket.

Constructors

RecvConnClosed
Connection closed

RecvError String
Any other error

RecvOk String
Successfull receive

data LuxiOp

Currently supported Luxi operations and JSON serialization.

Constructors

Query ItemType [String] (Filter FilterField)

QueryFields ItemType [String]

QueryNodes [String] [String] Bool

QueryGroups [String] [String] Bool

QueryNetworks [String] [String] Bool

QueryInstances [String] [String] Bool

QueryJobs [JobId] [String]

QueryExports [String] Bool

QueryConfigValues [String]

QueryClusterInfo

QueryTags TagKind String

SubmitJob [MetaOpCode]

SubmitJobToDrainedQueue [MetaOpCode]

SubmitManyJobs [[MetaOpCode]]

WaitForJobChange JobId [String] JSValue JSValue Int

ArchiveJob JobId

AutoArchiveJobs Int Int

CancelJob JobId

ChangeJobPriority JobId Int

SetDrainFlag Bool

SetWatcherPause Double

data LuxiReq

Constructors

ReqQuery

ReqQueryFields

ReqQueryNodes

ReqQueryGroups

ReqQueryNetworks

ReqQueryInstances

ReqQueryJobs

ReqQueryExports

ReqQueryConfigValues

ReqQueryClusterInfo

ReqQueryTags

ReqSubmitJob

ReqSubmitJobToDrainedQueue

ReqSubmitManyJobs

ReqWaitForJobChange

ReqArchiveJob

ReqAutoArchiveJobs

ReqCancelJob

ReqChangeJobPriority

ReqSetDrainFlag

ReqSetWatcherPause

luxiReqFromRaw :: forall m. Monad m => String -> m LuxiReq

luxiReqToRaw :: LuxiReq -> String

opToArgs :: LuxiOp -> JSValue

opToDict :: LuxiOp -> [(String, JSValue)]

allLuxiCalls :: [String]

List of all defined Luxi calls.

strOfOp :: LuxiOp -> String

The serialisation of LuxiOps into strings in messages.

data LuxiCall

Constructors

LuxiCall LuxiReq JSValue

eOM :: Word8

bEOM :: ByteString

data MsgKeys

Constructors

Method

Args

Success

Result

strOfKey :: MsgKeys -> String

data Client

Luxi client encapsulation.

Constructors

Client

Fields

socket :: Handle

rbuf :: IORef ByteString

getClient :: String -> IO Client

Connects to the master daemon and returns a luxi Client.

getServer :: Bool -> FilePath -> IO Socket

Creates and returns a server endpoint.

closeServer :: FilePath -> Socket -> IO ()

Closes a server endpoint. FIXME: this should be encapsulated into a nicer type.

acceptClient :: Socket -> IO Client

Accepts a client

closeClient :: Client -> IO ()

Closes the client socket.

sendMsg :: Client -> String -> IO ()

Sends a message over a luxi transport.

recvUpdate :: Handle -> ByteString -> IO (ByteString, ByteString)

recvMsg :: Client -> IO String

Waits for a message over a luxi transport.

recvMsgExt :: Client -> IO RecvResult

Extended wrapper over recvMsg.

buildCall

Arguments

:: LuxiOp
The method

-> String
The serialized form

Serialize a request to String.

buildResponse

Arguments

:: Bool
Success

-> JSValue
The arguments

-> String
The serialized form

Serialize the response to String.

validateCall :: String -> Result LuxiCall

Check that luxi request contains the required keys and parse it.

decodeCall :: LuxiCall -> Result LuxiOp

Converts Luxi call arguments into a LuxiOp data structure.

This is currently hand-coded until we make it more uniform so that it can be generated using TH.

validateResult :: String -> ErrorResult JSValue

decodeError :: JSValue -> ErrorResult JSValue

callMethod :: LuxiOp -> Client -> IO (ErrorResult JSValue)

Generic luxi method call.

parseSubmitJobResult :: JSValue -> ErrorResult JobId

submitManyJobs :: Client -> [[MetaOpCode]] -> IO (ErrorResult [JobId])

Specialized submitManyJobs call.

queryJobsStatus :: Client -> [JobId] -> IO (ErrorResult [JobStatus])

Custom queryJobs call.