Ganeti.Network

Description

Implementation of the Ganeti network objects.

This is does not (yet) cover all methods that are provided in the corresponding python implementation (network.py).

Documentation

data AddressPool

Constructors

AddressPool

Fields

network :: Network

reservations :: Vector Bool

extReservations :: Vector Bool

createAddressPool :: Network -> Maybe AddressPool

Create an address pool from a network.

networkIsValid :: Network -> Bool

Checks the consistency of the network object. So far, only checks the length of the reservation strings.

sameLength :: Maybe String -> Maybe String -> Bool

maybeStr2BitVec :: Maybe String -> Vector Bool

bitStringToBitVector :: String -> Vector Bool

Converts a string to a bit vector. The character '0' is interpreted as False, all others as True.

allReservations :: AddressPool -> Vector Bool

Get a bit vector of all reservations (internal and external) combined.

getReservedCount :: AddressPool -> Int

Get the count of reserved addresses.

getFreeCount :: AddressPool -> Int

Get the count of free addresses.

isFull :: AddressPool -> Bool

Check whether the network is full.

getMap :: AddressPool -> String

Return a textual representation of the network's occupation status.