Ganeti.Objects

Description

Implementation of the Ganeti config objects.

Some object fields are not implemented yet, and as such they are commented out below.

Generic definitions

fillDict :: Ord k => Map k v -> Map k v -> [k] -> Map k v

Fills one map with keys from the other map, if not already existing. Mirrors objects.py:FillDict.

type HvParams = Container JSValue

The hypervisor parameter type. This is currently a simple map, without type checking on key/value pairs.

type OsParams = Container String

The OS parameters type. This is, and will remain, a string container, since the keys are dynamically declared by the OSes, and the values are always strings.

type OsParamsPrivate = Container (Private String)

class TimeStampObject a where

Class of objects that have timestamps.

Methods

cTimeOf :: a -> ClockTime

mTimeOf :: a -> ClockTime

class UuidObject a where

Class of objects that have an UUID.

Methods

uuidOf :: a -> String

class SerialNoObject a where

Class of object that have a serial number.

Methods

serialOf :: a -> Int

class TagsObject a where

Class of objects that have tags.

Methods

tagsOf :: a -> Set String

Network definitions

Ipv4 types

data Ip4Address

Constructors

Ip4Address Word8 Word8 Word8 Word8

mkIp4Address :: (Word8, Word8, Word8, Word8) -> Ip4Address

readIp4Address :: (Applicative m, Monad m) => String -> m Ip4Address

ip4AddressToList :: Ip4Address -> [Word8]

ip4AddressToNumber :: Ip4Address -> Integer

Converts an address into its ordinal number. This is needed for indexing IP adresses in reservation pools.

ip4AddressFromNumber :: Integer -> Ip4Address

Converts a number into an address. This is needed for indexing IP adresses in reservation pools.

nextIp4Address :: Ip4Address -> Ip4Address

data Ip4Network

Custom type for an IPv4 network.

Constructors

Ip4Network

Fields

ip4netAddr :: Ip4Address

ip4netMask :: Word8

mkIp4Network :: Ip4Address -> Word8 -> Ip4Network

Address pools

newtype AddressPool

Currently address pools just wrap a reservation BitArray.

In future, Network might be extended to include several address pools and address pools might include their own ranges of addresses.

Constructors

AddressPool

Fields

apReservations :: BitArray

Ganeti "network" config object.

data Network

Constructors

Network

Fields

networkName :: NonEmptyString

networkMacPrefix :: Maybe String

networkNetwork :: Ip4Network

networkNetwork6 :: Maybe String

networkGateway :: Maybe Ip4Address

networkGateway6 :: Maybe String

networkReservations :: Maybe AddressPool

networkExtReservations :: Maybe AddressPool

networkUuid :: String

networkCtime :: ClockTime

networkMtime :: ClockTime

networkSerial :: Int

networkTags :: TagSet

loadNetwork :: JSValue -> Result Network

saveNetwork :: Network -> JSValue

NIC definitions

data FilledNicParams

Constructors

FilledNicParams

Fields

nicpMode :: NICMode

nicpLink :: String

nicpVlan :: String

data PartialNicParams

Constructors

PartialNicParams

Fields

nicpModeP :: Maybe NICMode

nicpLinkP :: Maybe String

nicpVlanP :: Maybe String

allNicParamFields :: [String]

fillNicParams :: FilledNicParams -> PartialNicParams -> FilledNicParams

loadPartialNicParams :: JSValue -> Result PartialNicParams

savePartialNicParams :: PartialNicParams -> JSValue

loadFilledNicParams :: JSValue -> Result FilledNicParams

saveFilledNicParams :: FilledNicParams -> JSValue

data PartialNic

Constructors

PartialNic

Fields

nicMac :: String

nicIp :: Maybe String

nicNicparams :: PartialNicParams

nicNetwork :: Maybe String

nicName :: Maybe String

nicUuid :: String

loadPartialNic :: JSValue -> Result PartialNic

savePartialNic :: PartialNic -> JSValue

type MicroSeconds = Integer

Datacollector definitions

data DataCollectorConfig

The configuration regarding a single data collector.

Constructors

DataCollectorConfig

Fields

dataCollectorActive :: Bool

dataCollectorInterval :: MicroSeconds

loadDataCollectorConfig :: JSValue -> Result DataCollectorConfig

saveDataCollectorConfig :: DataCollectorConfig -> JSValue

Disk definitions

devType :: String

type DiskParams = Container JSValue

type DRBDSecret = String

An alias for DRBD secrets

data LogicalVolume

Constructors

LogicalVolume

Fields

lvGroup :: String

lvVolume :: String

data DiskLogicalId

The disk configuration type. This includes the disk type itself, for a more complete consistency. Note that since in the Python code-base there's no authoritative place where we document the logical id, this is probably a good reference point.

Constructors

LIDPlain LogicalVolume
Volume group, logical volume

LIDDrbd8 String String Int Int Int (Private DRBDSecret)
NodeA, NodeB, Port, MinorA, MinorB, Secret

LIDFile FileDriver String
Driver, path

LIDSharedFile FileDriver String
Driver, path

LIDGluster FileDriver String
Driver, path

LIDBlockDev BlockDriver String
Driver, path (must be under /dev)

LIDRados String String
Unused, path

LIDExt String String
ExtProvider, unique name

lidDiskType :: DiskLogicalId -> DiskTemplate

lidEncodeType :: DiskLogicalId -> [(String, JSValue)]

encodeDLId :: DiskLogicalId -> JSValue

encodeFullDLId :: DiskLogicalId -> (JSValue, [(String, JSValue)])

decodeDLId :: [(String, JSValue)] -> JSValue -> Result DiskLogicalId

data Disk

Disk data structure.

This is declared manually as it's a recursive structure, and our TH code currently can't build it.

Constructors

Disk

Fields

diskLogicalId :: DiskLogicalId

diskChildren :: [Disk]

diskIvName :: String

diskSize :: Int

diskMode :: DiskMode

diskName :: Maybe String

diskSpindles :: Maybe Int

diskParams :: Maybe DiskParams

diskUuid :: String

diskSerial :: Int

diskCtime :: ClockTime

diskMtime :: ClockTime

loadDisk :: JSValue -> Result Disk

saveDisk :: Disk -> JSValue

includesLogicalId :: LogicalVolume -> Disk -> Bool

Determines whether a disk or one of his children has the given logical id (determined by the volume group name and by the logical volume name). This can be true only for DRBD or LVM disks.

Instance definitions

data FilledBeParams

Constructors

FilledBeParams

Fields

bepMinmem :: Int

bepMaxmem :: Int

bepVcpus :: Int

bepAutoBalance :: Bool

bepAlwaysFailover :: Bool

bepSpindleUse :: Int

data PartialBeParams

Constructors

PartialBeParams

Fields

bepMinmemP :: Maybe Int

bepMaxmemP :: Maybe Int

bepVcpusP :: Maybe Int

bepAutoBalanceP :: Maybe Bool

bepAlwaysFailoverP :: Maybe Bool

bepSpindleUseP :: Maybe Int

allBeParamFields :: [String]

fillBeParams :: FilledBeParams -> PartialBeParams -> FilledBeParams

loadPartialBeParams :: JSValue -> Result PartialBeParams

savePartialBeParams :: PartialBeParams -> JSValue

loadFilledBeParams :: JSValue -> Result FilledBeParams

saveFilledBeParams :: FilledBeParams -> JSValue

data Instance

Constructors

Instance

Fields

instName :: String

instPrimaryNode :: String

instOs :: String

instHypervisor :: Hypervisor

instHvparams :: HvParams

instBeparams :: PartialBeParams

instOsparams :: OsParams

instOsparamsPrivate :: OsParamsPrivate

instAdminState :: AdminState

instAdminStateSource :: AdminStateSource

instNics :: [PartialNic]

instDisks :: [String]

instDiskTemplate :: DiskTemplate

instDisksActive :: Bool

instNetworkPort :: Maybe Int

instCtime :: ClockTime

instMtime :: ClockTime

instUuid :: String

instSerial :: Int

instTags :: TagSet

loadInstance :: JSValue -> Result Instance

saveInstance :: Instance -> JSValue

IPolicy definitions

data FilledISpecParams

Constructors

FilledISpecParams

Fields

ispecMemorySize :: Int

ispecDiskSize :: Int

ispecDiskCount :: Int

ispecCpuCount :: Int

ispecNicCount :: Int

ispecSpindleUse :: Int

data PartialISpecParams

Constructors

PartialISpecParams

Fields

ispecMemorySizeP :: Maybe Int

ispecDiskSizeP :: Maybe Int

ispecDiskCountP :: Maybe Int

ispecCpuCountP :: Maybe Int

ispecNicCountP :: Maybe Int

ispecSpindleUseP :: Maybe Int

allISpecParamFields :: [String]

fillISpecParams :: FilledISpecParams -> PartialISpecParams -> FilledISpecParams

loadPartialISpecParams :: JSValue -> Result PartialISpecParams

savePartialISpecParams :: PartialISpecParams -> JSValue

loadFilledISpecParams :: JSValue -> Result FilledISpecParams

saveFilledISpecParams :: FilledISpecParams -> JSValue

data MinMaxISpecs

Constructors

MinMaxISpecs

Fields

mmisMinSpec :: FilledISpecParams

mmisMaxSpec :: FilledISpecParams

loadMinMaxISpecs :: JSValue -> Result MinMaxISpecs

saveMinMaxISpecs :: MinMaxISpecs -> JSValue

data PartialIPolicy

Custom partial ipolicy. This is not built via buildParam since it has a special 2-level inheritance mode.

Constructors

PartialIPolicy

Fields

ipolicyMinMaxISpecsP :: Maybe [MinMaxISpecs]

ipolicyStdSpecP :: Maybe PartialISpecParams

ipolicySpindleRatioP :: Maybe Double

ipolicyVcpuRatioP :: Maybe Double

ipolicyDiskTemplatesP :: Maybe [DiskTemplate]

loadPartialIPolicy :: JSValue -> Result PartialIPolicy

savePartialIPolicy :: PartialIPolicy -> JSValue

data FilledIPolicy

Custom filled ipolicy. This is not built via buildParam since it has a special 2-level inheritance mode.

Constructors

FilledIPolicy

Fields

ipolicyMinMaxISpecs :: [MinMaxISpecs]

ipolicyStdSpec :: FilledISpecParams

ipolicySpindleRatio :: Double

ipolicyVcpuRatio :: Double

ipolicyDiskTemplates :: [DiskTemplate]

loadFilledIPolicy :: JSValue -> Result FilledIPolicy

saveFilledIPolicy :: FilledIPolicy -> JSValue

fillIPolicy :: FilledIPolicy -> PartialIPolicy -> FilledIPolicy

Custom filler for the ipolicy types.

Node definitions

data FilledNDParams

Constructors

FilledNDParams

Fields

ndpOobProgram :: String

ndpSpindleCount :: Int

ndpExclusiveStorage :: Bool

ndpOvs :: Bool

ndpOvsName :: String

ndpOvsLink :: String

ndpSshPort :: Int

ndpCpuSpeed :: Double

data PartialNDParams

Constructors

PartialNDParams

Fields

ndpOobProgramP :: Maybe String

ndpSpindleCountP :: Maybe Int

ndpExclusiveStorageP :: Maybe Bool

ndpOvsP :: Maybe Bool

ndpOvsNameP :: Maybe String

ndpOvsLinkP :: Maybe String

ndpSshPortP :: Maybe Int

ndpCpuSpeedP :: Maybe Double

allNDParamFields :: [String]

fillNDParams :: FilledNDParams -> PartialNDParams -> FilledNDParams

loadPartialNDParams :: JSValue -> Result PartialNDParams

savePartialNDParams :: PartialNDParams -> JSValue

loadFilledNDParams :: JSValue -> Result FilledNDParams

saveFilledNDParams :: FilledNDParams -> JSValue

type DiskState = Container JSValue

type HypervisorState = Container JSValue

data Node

Constructors

Node

Fields

nodeName :: String

nodePrimaryIp :: String

nodeSecondaryIp :: String

nodeMasterCandidate :: Bool

nodeOffline :: Bool

nodeDrained :: Bool

nodeGroup :: String

nodeMasterCapable :: Bool

nodeVmCapable :: Bool

nodeNdparams :: PartialNDParams

nodePowered :: Bool

nodeHvStateStatic :: HypervisorState

nodeDiskStateStatic :: DiskState

nodeCtime :: ClockTime

nodeMtime :: ClockTime

nodeUuid :: String

nodeSerial :: Int

nodeTags :: TagSet

loadNode :: JSValue -> Result Node

saveNode :: Node -> JSValue

NodeGroup definitions

type GroupDiskParams = Container DiskParams

The cluster/group disk parameters type.

type Networks = Container PartialNicParams

data NodeGroup

Constructors

NodeGroup

Fields

groupName :: String

groupMembers :: [String]

groupNdparams :: PartialNDParams

groupAllocPolicy :: AllocPolicy

groupIpolicy :: PartialIPolicy

groupDiskparams :: GroupDiskParams

groupNetworks :: Networks

groupHvStateStatic :: HypervisorState

groupDiskStateStatic :: DiskState

groupCtime :: ClockTime

groupMtime :: ClockTime

groupUuid :: String

groupSerial :: Int

groupTags :: TagSet

loadNodeGroup :: JSValue -> Result NodeGroup

saveNodeGroup :: NodeGroup -> JSValue

Job scheduler filtering definitions

data FilterAction

Actions that can be performed when a filter matches.

Constructors

Accept

Pause

Reject

Continue

RateLimit Int

data FilterPredicate

Constructors

FPJobId (Filter FilterField)

FPOpCode (Filter FilterField)

FPReason (Filter FilterField)

data FilterRule

Constructors

FilterRule

Fields

frWatermark :: JobId

frPriority :: NonNegative Int

frPredicates :: [FilterPredicate]

frAction :: FilterAction

frReasonTrail :: ReasonTrail

frUuid :: String

loadFilterRule :: JSValue -> Result FilterRule

saveFilterRule :: FilterRule -> JSValue

filterRuleOrder :: FilterRule -> FilterRule -> Ordering

Order in which filter rules are evaluated, according to `doc/design-optables.rst`. For FilterRule fields not specified as important for the order, we choose an arbitrary ordering effect (after the ones from the spec).

The Ord instance for FilterRule agrees with this function. Yet it is recommended to use this function instead of compare to be explicit that the spec order is used.

data IpFamily

IP family type

Constructors

IpFamilyV4

IpFamilyV6

ipFamilyFromRaw :: forall m. Monad m => Int -> m IpFamily

ipFamilyToRaw :: IpFamily -> Int

ipFamilyToVersion :: IpFamily -> Int

Conversion from IP family to IP version. This is needed because Python uses both, depending on context.

type ClusterHvParams = GenericContainer Hypervisor HvParams

Cluster HvParams (hvtype to hvparams mapping).

type OsHvParams = Container ClusterHvParams

Cluster Os-HvParams (os to hvparams mapping).

type ClusterBeParams = Container FilledBeParams

Cluser BeParams.

type ClusterOsParams = Container OsParams

Cluster OsParams.

type ClusterOsParamsPrivate = Container (Private OsParams)

type ClusterNicParams = Container FilledNicParams

Cluster NicParams.

type UidRange = (Int, Int)

A low-high UID ranges.

formatUidRange :: UidRange -> String

type UidPool = [UidRange]

Cluster UID Pool, list (low, high) UID ranges.

type IAllocatorParams = Container JSValue

The iallocator parameters type.

type CandidateCertificates = Container String

Cluster definitions

data Cluster

Constructors

Cluster

Fields

clusterRsahostkeypub :: String

clusterDsahostkeypub :: Maybe String

clusterHighestUsedPort :: Int

clusterTcpudpPortPool :: [Int]

clusterMacPrefix :: String

clusterVolumeGroupName :: Maybe String

clusterReservedLvs :: [String]

clusterDrbdUsermodeHelper :: Maybe String

clusterMasterNode :: String

clusterMasterIp :: String

clusterMasterNetdev :: String

clusterMasterNetmask :: Int

clusterUseExternalMipScript :: Bool

clusterClusterName :: String

clusterFileStorageDir :: String

clusterSharedFileStorageDir :: String

clusterGlusterStorageDir :: String

clusterEnabledHypervisors :: [Hypervisor]

clusterHvparams :: ClusterHvParams

clusterOsHvp :: OsHvParams

clusterBeparams :: ClusterBeParams

clusterOsparams :: ClusterOsParams

clusterOsparamsPrivateCluster :: ClusterOsParamsPrivate

clusterNicparams :: ClusterNicParams

clusterNdparams :: FilledNDParams

clusterDiskparams :: GroupDiskParams

clusterCandidatePoolSize :: Int

clusterModifyEtcHosts :: Bool

clusterModifySshSetup :: Bool

clusterMaintainNodeHealth :: Bool

clusterUidPool :: UidPool

clusterDefaultIallocator :: String

clusterDefaultIallocatorParams :: IAllocatorParams

clusterHiddenOs :: [String]

clusterBlacklistedOs :: [String]

clusterPrimaryIpFamily :: IpFamily

clusterPreallocWipeDisks :: Bool

clusterIpolicy :: FilledIPolicy

clusterHvStateStatic :: HypervisorState

clusterDiskStateStatic :: DiskState

clusterEnabledDiskTemplates :: [DiskTemplate]

clusterCandidateCerts :: CandidateCertificates

clusterMaxRunningJobs :: Int

clusterMaxTrackedJobs :: Int

clusterInstallImage :: String

clusterInstanceCommunicationNetwork :: String

clusterZeroingImage :: String

clusterCompressionTools :: [String]

clusterEnabledUserShutdown :: Bool

clusterDataCollectors :: Container DataCollectorConfig

clusterCtime :: ClockTime

clusterMtime :: ClockTime

clusterUuid :: String

clusterSerial :: Int

clusterTags :: TagSet

loadCluster :: JSValue -> Result Cluster

saveCluster :: Cluster -> JSValue

ConfigData definitions

data ConfigData

Constructors

ConfigData

Fields

configVersion :: Int

configCluster :: Cluster

configNodes :: Container Node

configNodegroups :: Container NodeGroup

configInstances :: Container Instance

configNetworks :: Container Network

configDisks :: Container Disk

configFilters :: Container FilterRule

configCtime :: ClockTime

configMtime :: ClockTime

configSerial :: Int

loadConfigData :: JSValue -> Result ConfigData

saveConfigData :: ConfigData -> JSValue

Master network parameters

data MasterNetworkParameters

Constructors

MasterNetworkParameters

Fields

masterNetworkParametersUuid :: String

masterNetworkParametersIp :: String

masterNetworkParametersNetmask :: Int

masterNetworkParametersNetdev :: String

masterNetworkParametersIpFamily :: IpFamily

loadMasterNetworkParameters :: JSValue -> Result MasterNetworkParameters

saveMasterNetworkParameters :: MasterNetworkParameters -> JSValue