Ganeti.Objects

Description

Implementation of the Ganeti config objects.

Some object fields are not implemented yet, and as such they are commented out below.

Generic definitions

fillDict :: Ord k => Map k v -> Map k v -> [k] -> Map k v

Fills one map with keys from the other map, if not already existing. Mirrors objects.py:FillDict.

type HvParams = Container JSValue

The hypervisor parameter type. This is currently a simple map, without type checking on key/value pairs.

type OsParams = Container String

The OS parameters type. This is, and will remain, a string container, since the keys are dynamically declared by the OSes, and the values are always strings.

class TimeStampObject a where

Class of objects that have timestamps.

Methods

cTimeOf :: a -> Double

mTimeOf :: a -> Double

class UuidObject a where

Class of objects that have an UUID.

Methods

uuidOf :: a -> String

class SerialNoObject a where

Class of object that have a serial number.

Methods

serialOf :: a -> Int

class TagsObject a where

Class of objects that have tags.

Methods

tagsOf :: a -> Set String

Network definitions

Ipv4 types

data Ip4Address

Custom type for a simple IPv4 address.

Constructors

Ip4Address Word8 Word8 Word8 Word8

readIp4Address :: (Applicative m, Monad m) => String -> m Ip4Address

Parses an IPv4 address from a string.

nextIp4Address :: Ip4Address -> Ip4Address

"Next" address implementation for IPv4 addresses.

Note that this loops! Note also that this is a very dumb implementation.

data Ip4Network

Custom type for an IPv4 network.

Constructors

Ip4Network Ip4Address Word8

Ganeti "network" config object.

data Network

Constructors

Network

Fields

networkName :: NonEmptyString

networkMacPrefix :: Maybe String

networkNetwork :: Ip4Network

networkNetwork6 :: Maybe String

networkGateway :: Maybe Ip4Address

networkGateway6 :: Maybe String

networkReservations :: Maybe String

networkExtReservations :: Maybe String

networkUuid :: String

networkCtime :: Double

networkMtime :: Double

networkSerial :: Int

networkTags :: TagSet

loadNetwork :: JSValue -> Result Network

saveNetwork :: Network -> JSValue

toDictNetwork :: Network -> [(String, JSValue)]

NIC definitions

data FilledNicParams

Constructors

FilledNicParams

Fields

nicpMode :: NICMode

nicpLink :: String

nicpVlan :: String

data PartialNicParams

Constructors

PartialNicParams

Fields

nicpModeP :: Maybe NICMode

nicpLinkP :: Maybe String

nicpVlanP :: Maybe String

allNicParamFields :: [String]

fillNicParams :: FilledNicParams -> PartialNicParams -> FilledNicParams

loadPartialNicParams :: JSValue -> Result PartialNicParams

savePartialNicParams :: PartialNicParams -> JSValue

toDictPartialNicParams :: PartialNicParams -> [(String, JSValue)]

loadFilledNicParams :: JSValue -> Result FilledNicParams

saveFilledNicParams :: FilledNicParams -> JSValue

toDictFilledNicParams :: FilledNicParams -> [(String, JSValue)]

data PartialNic

Constructors

PartialNic

Fields

nicMac :: String

nicIp :: Maybe String

nicNicparams :: PartialNicParams

nicNetwork :: Maybe String

nicName :: Maybe String

nicUuid :: String

loadPartialNic :: JSValue -> Result PartialNic

savePartialNic :: PartialNic -> JSValue

toDictPartialNic :: PartialNic -> [(String, JSValue)]

Disk definitions

devType :: String

data DiskLogicalId

The disk configuration type. This includes the disk type itself, for a more complete consistency. Note that since in the Python code-base there's no authoritative place where we document the logical id, this is probably a good reference point.

Constructors

LIDPlain String String
Volume group, logical volume

LIDDrbd8 String String Int Int Int String
NodeA, NodeB, Port, MinorA, MinorB, Secret

LIDFile FileDriver String
Driver, path

LIDSharedFile FileDriver String
Driver, path

LIDBlockDev BlockDriver String
Driver, path (must be under /dev)

LIDRados String String
Unused, path

LIDExt String String
ExtProvider, unique name

lidDiskType :: DiskLogicalId -> DiskTemplate

lidEncodeType :: DiskLogicalId -> [(String, JSValue)]

encodeDLId :: DiskLogicalId -> JSValue

encodeFullDLId :: DiskLogicalId -> (JSValue, [(String, JSValue)])

decodeDLId :: [(String, JSValue)] -> JSValue -> Result DiskLogicalId

data Disk

Disk data structure.

This is declared manually as it's a recursive structure, and our TH code currently can't build it.

Constructors

Disk

Fields

diskLogicalId :: DiskLogicalId

diskChildren :: [Disk]

diskIvName :: String

diskSize :: Int

diskMode :: DiskMode

diskName :: Maybe String

diskSpindles :: Maybe Int

diskUuid :: String

loadDisk :: JSValue -> Result Disk

saveDisk :: Disk -> JSValue

toDictDisk :: Disk -> [(String, JSValue)]

includesLogicalId :: String -> String -> Disk -> Bool

Determines whether a disk or one of his children has the given logical id (determined by the volume group name and by the logical volume name). This can be true only for DRBD or LVM disks.

Instance definitions

data FilledBeParams

Constructors

FilledBeParams

Fields

bepMinmem :: Int

bepMaxmem :: Int

bepVcpus :: Int

bepAutoBalance :: Bool

bepAlwaysFailover :: Bool

bepSpindleUse :: Int

data PartialBeParams

Constructors

PartialBeParams

Fields

bepMinmemP :: Maybe Int

bepMaxmemP :: Maybe Int

bepVcpusP :: Maybe Int

bepAutoBalanceP :: Maybe Bool

bepAlwaysFailoverP :: Maybe Bool

bepSpindleUseP :: Maybe Int

allBeParamFields :: [String]

fillBeParams :: FilledBeParams -> PartialBeParams -> FilledBeParams

loadPartialBeParams :: JSValue -> Result PartialBeParams

savePartialBeParams :: PartialBeParams -> JSValue

toDictPartialBeParams :: PartialBeParams -> [(String, JSValue)]

loadFilledBeParams :: JSValue -> Result FilledBeParams

saveFilledBeParams :: FilledBeParams -> JSValue

toDictFilledBeParams :: FilledBeParams -> [(String, JSValue)]

data Instance

Constructors

Instance

Fields

instName :: String

instPrimaryNode :: String

instOs :: String

instHypervisor :: Hypervisor

instHvparams :: HvParams

instBeparams :: PartialBeParams

instOsparams :: OsParams

instAdminState :: AdminState

instAdminStateSource :: AdminStateSource

instNics :: [PartialNic]

instDisks :: [Disk]

instDiskTemplate :: DiskTemplate

instDisksActive :: Bool

instNetworkPort :: Maybe Int

instCtime :: Double

instMtime :: Double

instUuid :: String

instSerial :: Int

instTags :: TagSet

loadInstance :: JSValue -> Result Instance

saveInstance :: Instance -> JSValue

toDictInstance :: Instance -> [(String, JSValue)]

getDiskSizeRequirements :: Instance -> Int

Retrieves the real disk size requirements for all the disks of the instance. This includes the metadata etc. and is different from the values visible to the instance.

IPolicy definitions

data FilledISpecParams

Constructors

FilledISpecParams

Fields

ispecMemorySize :: Int

ispecDiskSize :: Int

ispecDiskCount :: Int

ispecCpuCount :: Int

ispecNicCount :: Int

ispecSpindleUse :: Int

data PartialISpecParams

Constructors

PartialISpecParams

Fields

ispecMemorySizeP :: Maybe Int

ispecDiskSizeP :: Maybe Int

ispecDiskCountP :: Maybe Int

ispecCpuCountP :: Maybe Int

ispecNicCountP :: Maybe Int

ispecSpindleUseP :: Maybe Int

allISpecParamFields :: [String]

fillISpecParams :: FilledISpecParams -> PartialISpecParams -> FilledISpecParams

loadPartialISpecParams :: JSValue -> Result PartialISpecParams

savePartialISpecParams :: PartialISpecParams -> JSValue

toDictPartialISpecParams :: PartialISpecParams -> [(String, JSValue)]

loadFilledISpecParams :: JSValue -> Result FilledISpecParams

saveFilledISpecParams :: FilledISpecParams -> JSValue

toDictFilledISpecParams :: FilledISpecParams -> [(String, JSValue)]

data MinMaxISpecs

Constructors

MinMaxISpecs

Fields

mmisMinSpec :: FilledISpecParams

mmisMaxSpec :: FilledISpecParams

loadMinMaxISpecs :: JSValue -> Result MinMaxISpecs

saveMinMaxISpecs :: MinMaxISpecs -> JSValue

toDictMinMaxISpecs :: MinMaxISpecs -> [(String, JSValue)]

data PartialIPolicy

Custom partial ipolicy. This is not built via buildParam since it has a special 2-level inheritance mode.

Constructors

PartialIPolicy

Fields

ipolicyMinMaxISpecsP :: Maybe [MinMaxISpecs]

ipolicyStdSpecP :: Maybe PartialISpecParams

ipolicySpindleRatioP :: Maybe Double

ipolicyVcpuRatioP :: Maybe Double

ipolicyDiskTemplatesP :: Maybe [DiskTemplate]

loadPartialIPolicy :: JSValue -> Result PartialIPolicy

savePartialIPolicy :: PartialIPolicy -> JSValue

toDictPartialIPolicy :: PartialIPolicy -> [(String, JSValue)]

data FilledIPolicy

Custom filled ipolicy. This is not built via buildParam since it has a special 2-level inheritance mode.

Constructors

FilledIPolicy

Fields

ipolicyMinMaxISpecs :: [MinMaxISpecs]

ipolicyStdSpec :: FilledISpecParams

ipolicySpindleRatio :: Double

ipolicyVcpuRatio :: Double

ipolicyDiskTemplates :: [DiskTemplate]

loadFilledIPolicy :: JSValue -> Result FilledIPolicy

saveFilledIPolicy :: FilledIPolicy -> JSValue

toDictFilledIPolicy :: FilledIPolicy -> [(String, JSValue)]

fillIPolicy :: FilledIPolicy -> PartialIPolicy -> FilledIPolicy

Custom filler for the ipolicy types.

Node definitions

data FilledNDParams

Constructors

FilledNDParams

Fields

ndpOobProgram :: String

ndpSpindleCount :: Int

ndpExclusiveStorage :: Bool

ndpOvs :: Bool

ndpOvsName :: String

ndpOvsLink :: String

ndpSshPort :: Int

data PartialNDParams

Constructors

PartialNDParams

Fields

ndpOobProgramP :: Maybe String

ndpSpindleCountP :: Maybe Int

ndpExclusiveStorageP :: Maybe Bool

ndpOvsP :: Maybe Bool

ndpOvsNameP :: Maybe String

ndpOvsLinkP :: Maybe String

ndpSshPortP :: Maybe Int

allNDParamFields :: [String]

fillNDParams :: FilledNDParams -> PartialNDParams -> FilledNDParams

loadPartialNDParams :: JSValue -> Result PartialNDParams

savePartialNDParams :: PartialNDParams -> JSValue

toDictPartialNDParams :: PartialNDParams -> [(String, JSValue)]

loadFilledNDParams :: JSValue -> Result FilledNDParams

saveFilledNDParams :: FilledNDParams -> JSValue

toDictFilledNDParams :: FilledNDParams -> [(String, JSValue)]

data Node

Constructors

Node

Fields

nodeName :: String

nodePrimaryIp :: String

nodeSecondaryIp :: String

nodeMasterCandidate :: Bool

nodeOffline :: Bool

nodeDrained :: Bool

nodeGroup :: String

nodeMasterCapable :: Bool

nodeVmCapable :: Bool

nodeNdparams :: PartialNDParams

nodePowered :: Bool

nodeCtime :: Double

nodeMtime :: Double

nodeUuid :: String

nodeSerial :: Int

nodeTags :: TagSet

loadNode :: JSValue -> Result Node

saveNode :: Node -> JSValue

toDictNode :: Node -> [(String, JSValue)]

NodeGroup definitions

type DiskParams = Container (Container JSValue)

The disk parameters type.

type Networks = Container PartialNicParams

data NodeGroup

Constructors

NodeGroup

Fields

groupName :: String

groupMembers :: [String]

groupNdparams :: PartialNDParams

groupAllocPolicy :: AllocPolicy

groupIpolicy :: PartialIPolicy

groupDiskparams :: DiskParams

groupNetworks :: Networks

groupCtime :: Double

groupMtime :: Double

groupUuid :: String

groupSerial :: Int

groupTags :: TagSet

loadNodeGroup :: JSValue -> Result NodeGroup

saveNodeGroup :: NodeGroup -> JSValue

toDictNodeGroup :: NodeGroup -> [(String, JSValue)]

data IpFamily

IP family type

Constructors

IpFamilyV4

IpFamilyV6

ipFamilyFromRaw :: forall m. Monad m => Int -> m IpFamily

ipFamilyToRaw :: IpFamily -> Int

ipFamilyToVersion :: IpFamily -> Int

Conversion from IP family to IP version. This is needed because Python uses both, depending on context.

type ClusterHvParams = Container HvParams

Cluster HvParams (hvtype to hvparams mapping).

type OsHvParams = Container ClusterHvParams

Cluster Os-HvParams (os to hvparams mapping).

type ClusterBeParams = Container FilledBeParams

Cluser BeParams.

type ClusterOsParams = Container OsParams

Cluster OsParams.

type ClusterNicParams = Container FilledNicParams

Cluster NicParams.

type UidPool = [(Int, Int)]

type IAllocatorParams = Container JSValue

The iallocator parameters type.

type CandidateCertificates = Container String

Cluster definitions

data Cluster

Constructors

Cluster

Fields

clusterRsahostkeypub :: String

clusterDsahostkeypub :: Maybe String

clusterHighestUsedPort :: Int

clusterTcpudpPortPool :: [Int]

clusterMacPrefix :: String

clusterVolumeGroupName :: Maybe String

clusterReservedLvs :: [String]

clusterDrbdUsermodeHelper :: Maybe String

clusterMasterNode :: String

clusterMasterIp :: String

clusterMasterNetdev :: String

clusterMasterNetmask :: Int

clusterUseExternalMipScript :: Bool

clusterClusterName :: String

clusterFileStorageDir :: String

clusterSharedFileStorageDir :: String

clusterGlusterStorageDir :: String

clusterEnabledHypervisors :: [Hypervisor]

clusterHvparams :: ClusterHvParams

clusterOsHvp :: OsHvParams

clusterBeparams :: ClusterBeParams

clusterOsparams :: ClusterOsParams

clusterNicparams :: ClusterNicParams

clusterNdparams :: FilledNDParams

clusterDiskparams :: DiskParams

clusterCandidatePoolSize :: Int

clusterModifyEtcHosts :: Bool

clusterModifySshSetup :: Bool

clusterMaintainNodeHealth :: Bool

clusterUidPool :: UidPool

clusterDefaultIallocator :: String

clusterDefaultIallocatorParams :: IAllocatorParams

clusterHiddenOs :: [String]

clusterBlacklistedOs :: [String]

clusterPrimaryIpFamily :: IpFamily

clusterPreallocWipeDisks :: Bool

clusterIpolicy :: FilledIPolicy

clusterEnabledDiskTemplates :: [DiskTemplate]

clusterCandidateCerts :: CandidateCertificates

clusterMaxRunningJobs :: Int

clusterEnabledUserShutdown :: Bool

clusterCtime :: Double

clusterMtime :: Double

clusterUuid :: String

clusterSerial :: Int

clusterTags :: TagSet

loadCluster :: JSValue -> Result Cluster

saveCluster :: Cluster -> JSValue

toDictCluster :: Cluster -> [(String, JSValue)]

ConfigData definitions

data ConfigData

Constructors

ConfigData

Fields

configVersion :: Int

configCluster :: Cluster

configNodes :: Container Node

configNodegroups :: Container NodeGroup

configInstances :: Container Instance

configNetworks :: Container Network

configSerial :: Int

loadConfigData :: JSValue -> Result ConfigData

saveConfigData :: ConfigData -> JSValue

toDictConfigData :: ConfigData -> [(String, JSValue)]