Ganeti.OpParams

Description

Implementation of opcodes parameters.

These are defined in a separate module only due to TemplateHaskell stage restrictions - expressions defined in the current module can't be passed to splices. So we have to either parameters/repeat each parameter definition multiple times, or separate them into this module.

Helper functions and types

Type aliases

booleanField :: String -> Field

defaultFalse :: String -> Field

defaultTrue :: String -> Field

stringField :: String -> Field

optionalStringField :: String -> Field

optionalNEStringField :: String -> Field

type UncheckedValue = JSValue

type UncheckedDict = JSObject JSValue

type UncheckedList = [JSValue]

forceNonNeg :: (Num a, Ord a, Show a) => a -> NonNegative a

Tags

data TagType

Data type representing what items do the tag operations apply to.

Constructors

TagTypeInstance

TagTypeNode

TagTypeGroup

TagTypeCluster

tagTypeFromRaw :: forall m. Monad m => String -> m TagType

tagTypeToRaw :: TagType -> String

data TagObject

Data type holding a tag object (type and object name).

Constructors

TagInstance String

TagNode String

TagGroup String

TagCluster

tagTypeOf :: TagObject -> TagType

tagNameOf :: TagObject -> Maybe String

Gets the potential tag object name.

tagObjectFrom :: Monad m => TagType -> JSValue -> m TagObject

Builds a TagObject from a tag type and name.

tagNameField :: String

encodeTagObject :: TagObject -> (JSValue, [(String, JSValue)])

Custom encoder for TagObject as represented in an opcode.

decodeTagObject :: Monad m => [(String, JSValue)] -> JSValue -> m TagObject

Custom decoder for TagObject as represented in an opcode.

Disks

data ReplaceDisksMode

Replace disks type.

Constructors

ReplaceOnPrimary

ReplaceOnSecondary

ReplaceNewSecondary

ReplaceAuto

replaceDisksModeFromRaw :: forall m. Monad m => String -> m ReplaceDisksMode

replaceDisksModeToRaw :: ReplaceDisksMode -> String

newtype DiskIndex

Disk index type (embedding constraints on the index value via a smart constructor).

Constructors

DiskIndex

Fields

unDiskIndex :: Int

mkDiskIndex :: Monad m => Int -> m DiskIndex

Smart constructor for DiskIndex.

I* param types

data DiskAccess

Type holding disk access modes.

Constructors

DiskReadOnly

DiskReadWrite

diskAccessFromRaw :: forall m. Monad m => String -> m DiskAccess

diskAccessToRaw :: DiskAccess -> String

data INicParams

NIC modification definition.

Constructors

INicParams

Fields

inicMac :: Maybe NonEmptyString

inicIp :: Maybe String

inicMode :: Maybe NonEmptyString

inicLink :: Maybe NonEmptyString

loadINicParams :: JSValue -> Result INicParams

saveINicParams :: INicParams -> JSValue

toDictINicParams :: INicParams -> [(String, JSValue)]

data IDiskParams

Disk modification definition. FIXME: disksize should be VTYPE_UNIT.

Constructors

IDiskParams

Fields

idiskSize :: Maybe Int

idiskMode :: Maybe DiskAccess

idiskAdopt :: Maybe NonEmptyString

idiskVg :: Maybe NonEmptyString

idiskMetavg :: Maybe NonEmptyString

loadIDiskParams :: JSValue -> Result IDiskParams

saveIDiskParams :: IDiskParams -> JSValue

toDictIDiskParams :: IDiskParams -> [(String, JSValue)]

data RecreateDisksInfo

Disk changes type for OpInstanceRecreateDisks. This is a bit strange, because the type in Python is something like Either [DiskIndex] [DiskChanges], but we can't represent the type of an empty list in JSON, so we have to add a custom case for the empty list.

Constructors

RecreateDisksAll

RecreateDisksIndices (NonEmpty DiskIndex)

RecreateDisksParams (NonEmpty (DiskIndex, IDiskParams))

readRecreateDisks :: JSValue -> Result RecreateDisksInfo

data DdmOldChanges

Simple type for old-style ddm changes.

Constructors

DdmOldIndex (NonNegative Int)

DdmOldMod DdmSimple

readDdmOldChanges :: JSValue -> Result DdmOldChanges

data SetParamsMods a

Instance disk or nic modifications.

Constructors

SetParamsEmpty

SetParamsDeprecated (NonEmpty (DdmOldChanges, a))

SetParamsNew (NonEmpty (DdmFull, Int, a))

readSetParams :: JSON a => JSValue -> Result (SetParamsMods a)

data ExportTarget

Custom type for target_node parameter of OpBackupExport, which varies depending on mode. FIXME: this uses an UncheckedList since we don't care about individual rows (just like the Python code tests). But the proper type could be parsed if we wanted.

Constructors

ExportTargetLocal NonEmptyString

ExportTargetRemote UncheckedList

readExportTarget :: JSValue -> Result ExportTarget

Parameters

pInstanceName :: Field

A required instance name (for single-instance LUs).

pInstances :: Field

A list of instances.

pName :: Field

A generic name.

pTagsList :: Field

Tags list.

pTagsObject :: Field

Tags object.

pOutputFields :: Field

Selected output fields.

pShutdownTimeout :: Field

How long to wait for instance to shut down.

pShutdownTimeout' :: Field

Another name for the shutdown timeout, because we like to be inconsistent.

pShutdownInstance :: Field

Whether to shutdown the instance in backup-export.

pForce :: Field

Whether to force the operation.

pIgnoreOfflineNodes :: Field

Whether to ignore offline nodes.

pNodeName :: Field

A required node name (for single-node LUs).

pNodeNames :: Field

List of nodes.

pGroupName :: Field

A required node group name (for single-group LUs).

pMigrationMode :: Field

Migration type (live/non-live).

pMigrationLive :: Field

Obsolete 'live' migration mode (boolean).

pMigrationCleanup :: Field

Migration cleanup parameter.

pForceVariant :: Field

Whether to force an unknown OS variant.

pWaitForSync :: Field

Whether to wait for the disk to synchronize.

pWaitForSyncFalse :: Field

Whether to wait for the disk to synchronize (defaults to false).

pIgnoreConsistency :: Field

Whether to ignore disk consistency

pStorageName :: Field

Storage name.

pUseLocking :: Field

Whether to use synchronization.

pOpportunisticLocking :: Field

Whether to employ opportunistic locking for nodes, meaning nodes already locked by another opcode won't be considered for instance allocation (only when an iallocator is used).

pNameCheck :: Field

Whether to check name.

pNodeGroupAllocPolicy :: Field

Instance allocation policy.

pGroupNodeParams :: Field

Default node parameters for group.

pQueryWhat :: Field

Resource(s) to query for.

pEarlyRelease :: Field

Whether to release locks as soon as possible.

pIpCheck :: Field

Whether to ensure instance's IP address is inactive.

pIpConflictsCheck :: Field

Check for conflicting IPs.

pNoRemember :: Field

Do not remember instance state changes.

pMigrationTargetNode :: Field

Target node for instance migration/failover.

pMoveTargetNode :: Field

Target node for instance move (required).

pStartupPaused :: Field

Pause instance at startup.

pVerbose :: Field

Verbose mode.

Parameters for cluster verification

pDebugSimulateErrors :: Field

Whether to simulate errors (useful for debugging).

pErrorCodes :: Field

Error codes.

pSkipChecks :: Field

Which checks to skip.

pIgnoreErrors :: Field

List of error codes that should be treated as warnings.

pOptGroupName :: Field

Optional group name.

pDiskParams :: Field

Disk templates' parameter defaults.

Parameters for node resource model

pHvState :: Field

Set hypervisor states.

pDiskState :: Field

Set disk states.

pIgnoreIpolicy :: Field

Whether to ignore ipolicy violations.

pAllowRuntimeChgs :: Field

Allow runtime changes while migrating.

type TestClusterOsListItem = (DdmSimple, NonEmptyString)

type TestClusterOsList = [TestClusterOsListItem]

pInstDisks :: Field

List of instance disks.

pDiskTemplate :: Field

Instance disk template.

pOptDiskTemplate :: Field

Instance disk template.

pFileDriver :: Field

File driver.

pFileStorageDir :: Field

Directory for storing file-backed disks.

pVgName :: Field

Volume group name.

pEnabledHypervisors :: Field

List of enabled hypervisors.

pHypervisor :: Field

Selected hypervisor for an instance.

pClusterHvParams :: Field

Cluster-wide hypervisor parameters, hypervisor-dependent.

pInstHvParams :: Field

Instance hypervisor parameters.

pClusterBeParams :: Field

Cluster-wide beparams.

pInstBeParams :: Field

Instance beparams.

pResetDefaults :: Field

Reset instance parameters to default if equal.

pOsHvp :: Field

Cluster-wide per-OS hypervisor parameter defaults.

pClusterOsParams :: Field

Cluster-wide OS parameter defaults.

pInstOsParams :: Field

Instance OS parameters.

pTempOsParams :: Field

Temporary OS parameters (currently only in reinstall, might be added to install as well).

pTempHvParams :: Field

Temporary hypervisor parameters, hypervisor-dependent.

pTempBeParams :: Field

Temporary backend parameters.

pCandidatePoolSize :: Field

Candidate pool size.

pUidPool :: Field

Set UID pool, must be list of lists describing UID ranges (two items, start and end inclusive.

pAddUids :: Field

Extend UID pool, must be list of lists describing UID ranges (two items, start and end inclusive.

pRemoveUids :: Field

Shrink UID pool, must be list of lists describing UID ranges (two items, start and end inclusive) to be removed.

pMaintainNodeHealth :: Field

Whether to automatically maintain node health.

pPreallocWipeDisks :: Field

Whether to wipe disks before allocating them to instances.

pNicParams :: Field

Cluster-wide NIC parameter defaults.

pInstNics :: Field

Instance NIC definitions.

pNdParams :: Field

Cluster-wide node parameter defaults.

pIpolicy :: Field

Cluster-wide ipolicy specs.

pDrbdHelper :: Field

DRBD helper program.

pDefaultIAllocator :: Field

Default iallocator for cluster.

pMasterNetdev :: Field

Master network device.

pMasterNetmask :: Field

Netmask of the master IP.

pReservedLvs :: Field

List of reserved LVs.

pHiddenOs :: Field

Modify list of hidden operating systems: each modification must have two items, the operation and the OS name; the operation can be add or remove.

pBlacklistedOs :: Field

Modify list of blacklisted operating systems: each modification must have two items, the operation and the OS name; the operation can be add or remove.

pUseExternalMipScript :: Field

Whether to use an external master IP address setup script.

pQueryFields :: Field

Requested fields.

pQueryFilter :: Field

Query filter.

pOobCommand :: Field

OOB command to run.

pOobTimeout :: Field

Timeout before the OOB helper will be terminated.

pIgnoreStatus :: Field

Ignores the node offline status for power off.

pPowerDelay :: Field

Time in seconds to wait between powering on nodes.

pPrimaryIp :: Field

Primary IP address.

pSecondaryIp :: Field

Secondary IP address.

pReadd :: Field

Whether node is re-added to cluster.

pNodeGroup :: Field

Initial node group.

pMasterCapable :: Field

Whether node can become master or master candidate.

pVmCapable :: Field

Whether node can host instances.

pNames :: Field

List of names.

pNodes :: Field

List of node names.

pRequiredNodes :: Field

Required list of node names.

pStorageType :: Field

Storage type.

pStorageChanges :: Field

Storage changes (unchecked).

pMasterCandidate :: Field

Whether the node should become a master candidate.

pOffline :: Field

Whether the node should be marked as offline.

pDrained :: Field

Whether the node should be marked as drained.

pAutoPromote :: Field

Whether node(s) should be promoted to master candidate if necessary.

pPowered :: Field

Whether the node should be marked as powered

pIallocator :: Field

Iallocator for deciding the target node for shared-storage instances during migrate and failover.

pRemoteNode :: Field

New secondary node.

pEvacMode :: Field

Node evacuation mode.

pInstCreateMode :: Field

Instance creation mode.

pNoInstall :: Field

Do not install the OS (will disable automatic start).

pInstOs :: Field

OS type for instance installation.

pPrimaryNode :: Field

Primary node for an instance.

pSecondaryNode :: Field

Secondary node for an instance.

pSourceHandshake :: Field

Signed handshake from source (remote import only).

pSourceInstance :: Field

Source instance name (remote import only).

pSourceShutdownTimeout :: Field

How long source instance was given to shut down (remote import only). FIXME: non-negative int, whereas the constant is a plain int.

pSourceX509Ca :: Field

Source X509 CA in PEM format (remote import only).

pSrcNode :: Field

Source node for import.

pSrcPath :: Field

Source directory for import.

pStartInstance :: Field

Whether to start instance after creation.

pInstTags :: Field

Instance tags. FIXME: unify/simplify with pTags, once that migrates to NonEmpty String.

pMultiAllocInstances :: Field

Unchecked list of OpInstanceCreate, used in OpInstanceMultiAlloc.

pIgnoreFailures :: Field

Ignore failures parameter.

pNewName :: Field

New instance or cluster name.

pIgnoreSecondaries :: Field

Whether to start the instance even if secondary disks are failing.

pRebootType :: Field

How to reboot the instance.

pIgnoreDiskSize :: Field

Whether to ignore recorded disk size.

pRecreateDisksInfo :: Field

Disk list for recreate disks.

pStatic :: Field

Whether to only return configuration data without querying nodes.

pInstParamsNicChanges :: Field

InstanceSetParams NIC changes.

pInstParamsDiskChanges :: Field

InstanceSetParams Disk changes.

pRuntimeMem :: Field

New runtime memory.

pOsNameChange :: Field

Change the instance's OS without reinstalling the instance

pDiskIndex :: Field

Disk index for e.g. grow disk.

pDiskChgAmount :: Field

Disk amount to add or grow to.

pDiskChgAbsolute :: Field

Whether the amount parameter is an absolute target or a relative one.

pTargetGroups :: Field

Destination group names or UUIDs (defaults to "all but current group".

pExportMode :: Field

Export mode field.

pExportTargetNode :: Field

Export target_node field, depends on mode.

pRemoveInstance :: Field

Whether to remove instance after export.

pIgnoreRemoveFailures :: Field

Whether to ignore failures while removing instances.

pX509KeyName :: Field

Name of X509 key (remote export only).

pX509DestCA :: Field

Destination X509 CA (remote export only).

pTagSearchPattern :: Field

Search pattern (regular expression). FIXME: this should be compiled at load time?

pRestrictedCommand :: Field

Restricted command name.

pReplaceDisksMode :: Field

Replace disks mode.

pReplaceDisksList :: Field

List of disk indices.

pAllowFailover :: Field

Whether do allow failover in migrations.

Test opcode parameters

pDelayDuration :: Field

Duration parameter for OpTestDelay.

pDelayOnMaster :: Field

on_master field for OpTestDelay.

pDelayOnNodes :: Field

on_nodes field for OpTestDelay.

pDelayRepeat :: Field

Repeat parameter for OpTestDelay.

pIAllocatorDirection :: Field

IAllocator test direction.

pIAllocatorMode :: Field

IAllocator test mode.

pIAllocatorReqName :: Field

IAllocator target name (new instance, node to evac, etc.).

pIAllocatorNics :: Field

Custom OpTestIAllocator nics.

pIAllocatorDisks :: Field

Custom OpTestAllocator disks.

pIAllocatorMemory :: Field

IAllocator memory field.

pIAllocatorVCpus :: Field

IAllocator vcpus field.

pIAllocatorOs :: Field

IAllocator os field.

pIAllocatorInstances :: Field

IAllocator instances field.

pIAllocatorEvacMode :: Field

IAllocator evac mode.

pIAllocatorSpindleUse :: Field

IAllocator spindle use.

pIAllocatorCount :: Field

IAllocator count field.

pJQueueNotifyWaitLock :: Field

OpTestJqueue notify_waitlock.

pJQueueNotifyExec :: Field

OpTestJQueue notify_exec.

pJQueueLogMessages :: Field

OpTestJQueue log_messages.

pJQueueFail :: Field

OpTestJQueue fail attribute.

pTestDummyResult :: Field

OpTestDummy result field.

pTestDummyMessages :: Field

OpTestDummy messages field.

pTestDummyFail :: Field

OpTestDummy fail field.

pTestDummySubmitJobs :: Field

OpTestDummy submit_jobs field.

Network parameters

pNetworkName :: Field

Network name.

pNetworkAddress4 :: Field

Network address (IPv4 subnet). FIXME: no real type for this.

pNetworkGateway4 :: Field

Network gateway (IPv4 address). FIXME: no real type for this.

pNetworkAddress6 :: Field

Network address (IPv6 subnet). FIXME: no real type for this.

pNetworkGateway6 :: Field

Network gateway (IPv6 address). FIXME: no real type for this.

pNetworkMacPrefix :: Field

Network specific mac prefix (that overrides the cluster one).

pNetworkAddRsvdIps :: Field

Network add reserved IPs.

pNetworkRemoveRsvdIps :: Field

Network remove reserved IPs.

pNetworkMode :: Field

Network mode when connecting to a group.

pNetworkLink :: Field

Network link when connecting to a group.

Common opcode parameters

pDryRun :: Field

Run checks only, don't execute.

pDebugLevel :: Field

Debug level.

pOpPriority :: Field

Opcode priority. Note: python uses a separate constant, we're using the actual value we know it's the default.

pDependencies :: Field

Job dependencies.

pComment :: Field

Comment field.

Entire opcode parameter list

dOldQuery :: [Field]

Old-style query opcode, with locking.

dOldQueryNoLocking :: [Field]

Old-style query opcode, without locking.