Ganeti.Types

Description

Some common Ganeti types.

This holds types common to both core work, and to htools. Types that are very core specific (e.g. configuration objects) should go in Objects, while types that are specific to htools in-memory representation should go into Types.

Generic types

newtype NonNegative a

Type that holds a non-negative value.

Constructors

NonNegative

Fields

fromNonNegative :: a

mkNonNegative :: (Monad m, Num a, Ord a, Show a) => a -> m (NonNegative a)

Smart constructor for NonNegative.

newtype Positive a

Type that holds a positive value.

Constructors

Positive

Fields

fromPositive :: a

mkPositive :: (Monad m, Num a, Ord a, Show a) => a -> m (Positive a)

Smart constructor for Positive.

newtype Negative a

Type that holds a negative value.

Constructors

Negative

Fields

fromNegative :: a

mkNegative :: (Monad m, Num a, Ord a, Show a) => a -> m (Negative a)

Smart constructor for Negative.

newtype NonEmpty a

Type that holds a non-null list.

Constructors

NonEmpty

Fields

fromNonEmpty :: [a]

mkNonEmpty :: Monad m => [a] -> m (NonEmpty a)

Smart constructor for NonEmpty.

type NonEmptyString = NonEmpty Char

A simple type alias for non-empty strings.

Ganeti types

data DiskTemplate

Instance disk template type.

Constructors

DTDiskless

DTFile

DTSharedFile

DTPlain

DTBlock

DTDrbd8

DTRbd

DTExt

diskTemplateFromRaw :: forall m. Monad m => String -> m DiskTemplate

diskTemplateToRaw :: DiskTemplate -> String

data AllocPolicy

The Group allocation policy type.

Note that the order of constructors is important as the automatic Ord instance will order them in the order they are defined, so when changing this data type be careful about the interaction with the desired sorting order.

Constructors

AllocPreferred

AllocLastResort

AllocUnallocable

allocPolicyFromRaw :: forall m. Monad m => String -> m AllocPolicy

allocPolicyToRaw :: AllocPolicy -> String

data InstanceStatus

The Instance real state type. FIXME: this could be improved to just wrap a NormalState AdminStatus | ErrorState ErrorCondition.

Constructors

StatusDown

StatusOffline

ErrorDown

ErrorUp

NodeDown

NodeOffline

Running

WrongNode

instanceStatusFromRaw :: forall m. Monad m => String -> m InstanceStatus

instanceStatusToRaw :: InstanceStatus -> String

data MigrationMode

Migration mode.

Constructors

MigrationLive

MigrationNonLive

migrationModeFromRaw :: forall m. Monad m => String -> m MigrationMode

migrationModeToRaw :: MigrationMode -> String

data VerifyOptionalChecks

Verify optional checks.

Constructors

VerifyNPlusOneMem

verifyOptionalChecksFromRaw :: forall m. Monad m => String -> m VerifyOptionalChecks

verifyOptionalChecksToRaw :: VerifyOptionalChecks -> String

data CVErrorCode

Cluster verify error codes.

Constructors

CvECLUSTERCFG

CvECLUSTERCERT

CvECLUSTERFILECHECK

CvECLUSTERDANGLINGNODES

CvECLUSTERDANGLINGINST

CvEINSTANCEBADNODE

CvEINSTANCEDOWN

CvEINSTANCELAYOUT

CvEINSTANCEMISSINGDISK

CvEINSTANCEFAULTYDISK

CvEINSTANCEWRONGNODE

CvEINSTANCESPLITGROUPS

CvEINSTANCEPOLICY

CvENODEDRBD

CvENODEDRBDHELPER

CvENODEFILECHECK

CvENODEHOOKS

CvENODEHV

CvENODELVM

CvENODEN1

CvENODENET

CvENODEOS

CvENODEORPHANINSTANCE

CvENODEORPHANLV

CvENODERPC

CvENODESSH

CvENODEVERSION

CvENODESETUP

CvENODETIME

CvENODEOOBPATH

CvENODEUSERSCRIPTS

CvENODEFILESTORAGEPATHS

cVErrorCodeFromRaw :: forall m. Monad m => String -> m CVErrorCode

cVErrorCodeToRaw :: CVErrorCode -> String

data DdmSimple

Dynamic device modification, just add/remove version.

Constructors

DdmSimpleAdd

DdmSimpleRemove

ddmSimpleFromRaw :: forall m. Monad m => String -> m DdmSimple

ddmSimpleToRaw :: DdmSimple -> String

data DdmFull

Dynamic device modification, all operations version.

Constructors

DdmFullAdd

DdmFullRemove

DdmFullModify

ddmFullFromRaw :: forall m. Monad m => String -> m DdmFull

ddmFullToRaw :: DdmFull -> String

data Hypervisor

Hypervisor type definitions.

Constructors

Kvm

XenPvm

Chroot

XenHvm

Lxc

Fake

hypervisorFromRaw :: forall m. Monad m => String -> m Hypervisor

hypervisorToRaw :: Hypervisor -> String

data OobCommand

Oob command type.

Constructors

OobHealth

OobPowerCycle

OobPowerOff

OobPowerOn

OobPowerStatus

oobCommandFromRaw :: forall m. Monad m => String -> m OobCommand

oobCommandToRaw :: OobCommand -> String

data StorageType

Storage type.

Constructors

StorageFile

StorageLvmPv

StorageLvmVg

storageTypeFromRaw :: forall m. Monad m => String -> m StorageType

storageTypeToRaw :: StorageType -> String

data NodeEvacMode

Node evac modes.

Constructors

NEvacPrimary

NEvacSecondary

NEvacAll

nodeEvacModeFromRaw :: forall m. Monad m => String -> m NodeEvacMode

nodeEvacModeToRaw :: NodeEvacMode -> String

data FileDriver

The file driver type.

Constructors

FileLoop

FileBlktap

fileDriverFromRaw :: forall m. Monad m => String -> m FileDriver

fileDriverToRaw :: FileDriver -> String

data InstCreateMode

The instance create mode.

Constructors

InstCreate

InstImport

InstRemoteImport

instCreateModeFromRaw :: forall m. Monad m => String -> m InstCreateMode

instCreateModeToRaw :: InstCreateMode -> String

data RebootType

Reboot type.

Constructors

RebootSoft

RebootHard

RebootFull

rebootTypeFromRaw :: forall m. Monad m => String -> m RebootType

rebootTypeToRaw :: RebootType -> String

data ExportMode

Export modes.

Constructors

ExportModeLocal

ExportModeRemove

exportModeFromRaw :: forall m. Monad m => String -> m ExportMode

exportModeToRaw :: ExportMode -> String

data IAllocatorTestDir

IAllocator run types (OpTestIAllocator).

Constructors

IAllocatorDirIn

IAllocatorDirOut

iAllocatorTestDirFromRaw :: forall m. Monad m => String -> m IAllocatorTestDir

iAllocatorTestDirToRaw :: IAllocatorTestDir -> String

data IAllocatorMode

IAllocator mode. FIXME: use this in HTools.Backend.IAlloc.

Constructors

IAllocatorAlloc

IAllocatorMultiAlloc

IAllocatorReloc

IAllocatorNodeEvac

IAllocatorChangeGroup

iAllocatorModeFromRaw :: forall m. Monad m => String -> m IAllocatorMode

iAllocatorModeToRaw :: IAllocatorMode -> String

data NICMode

Netork mode.

Constructors

NMBridged

NMRouted

NMOvs

nICModeFromRaw :: forall m. Monad m => String -> m NICMode

nICModeToRaw :: NICMode -> String

data JobStatus

The JobStatus data type. Note that this is ordered especially such that greater/lesser comparison on values of this type makes sense.

Constructors

JOB_STATUS_QUEUED

JOB_STATUS_WAITING

JOB_STATUS_CANCELING

JOB_STATUS_RUNNING

JOB_STATUS_CANCELED

JOB_STATUS_SUCCESS

JOB_STATUS_ERROR

jobStatusFromRaw :: forall m. Monad m => String -> m JobStatus

jobStatusToRaw :: JobStatus -> String

data FinalizedJobStatus

Finalized job status.

Constructors

JobStatusCanceled

JobStatusSuccessful

JobStatusFailed

finalizedJobStatusFromRaw :: forall m. Monad m => String -> m FinalizedJobStatus

finalizedJobStatusToRaw :: FinalizedJobStatus -> String

newtype JobId

The Ganeti job type.

Constructors

JobId

Fields

fromJobId :: Int

makeJobId :: Monad m => Int -> m JobId

Builds a job ID.

makeJobIdS :: Monad m => String -> m JobId

Builds a job ID from a string.

parseJobId :: Monad m => JSValue -> m JobId

type RelativeJobId = Negative Int

Relative job ID type alias.

data JobIdDep

Job ID dependency.

Constructors

JobDepRelative RelativeJobId

JobDepAbsolute JobId

data JobDependency

Job Dependency type.

Constructors

JobDependency JobIdDep [FinalizedJobStatus]

data OpSubmitPriority

Valid opcode priorities for submit.

Constructors

OpPrioLow

OpPrioNormal

OpPrioHigh

opSubmitPriorityFromRaw :: forall m. Monad m => Int -> m OpSubmitPriority

opSubmitPriorityToRaw :: OpSubmitPriority -> Int

parseSubmitPriority :: Monad m => String -> m OpSubmitPriority

Parse submit priorities from a string.

fmtSubmitPriority :: OpSubmitPriority -> String

Format a submit priority as string.

data OpStatus

Our ADT for the OpCode status at runtime (while in a job).

Constructors

OP_STATUS_QUEUED

OP_STATUS_WAITING

OP_STATUS_CANCELING

OP_STATUS_RUNNING

OP_STATUS_CANCELED

OP_STATUS_SUCCESS

OP_STATUS_ERROR

opStatusFromRaw :: forall m. Monad m => String -> m OpStatus

opStatusToRaw :: OpStatus -> String

data ELogType

Type for the job message type.

Constructors

ELogMessage

ELogRemoteImport

ELogJqueueTest

eLogTypeFromRaw :: forall m. Monad m => String -> m ELogType

eLogTypeToRaw :: ELogType -> String