Ganeti.UDSServer

Description

Implementation of the Ganeti Unix Domain Socket JSON server interface.

Utility functions

withTimeout :: Int -> String -> IO a -> IO a

Generic protocol functionality

data RecvResult

Result of receiving a message from the socket.

Constructors

RecvConnClosed
Connection closed

RecvError String
Any other error

RecvOk String
Successfull receive

eOM :: Word8

bEOM :: ByteString

data MsgKeys

Valid keys in the requests and responses.

Constructors

Method

Args

Success

Result

strOfKey :: MsgKeys -> String

The serialisation of MsgKeys into strings in messages.

data ConnectConfig

Constructors

ConnectConfig

Fields

connDaemon :: GanetiDaemon

recvTmo :: Int

sendTmo :: Int

data Client

A client encapsulation.

Constructors

Client

Fields

socket :: Handle

rbuf :: IORef ByteString

clientConfig :: ConnectConfig

data Server

A server encapsulation.

Constructors

Server

Fields

sSocket :: Socket

sPath :: FilePath

serverConfig :: ConnectConfig

Unix sockets

openClientSocket

Arguments

:: Int
connection timeout

-> FilePath
socket path

-> IO Handle

Creates a Unix socket and connects it to the specified path, where timeout specifies the connection timeout.

closeClientSocket :: Handle -> IO ()

openServerSocket :: FilePath -> IO Socket

Creates a Unix socket and binds it to the specified path.

closeServerSocket :: Socket -> FilePath -> IO ()

acceptSocket :: Socket -> IO Handle

Client and server

connectClient

Arguments

:: ConnectConfig
configuration for the client

-> Int
connection timeout

-> FilePath
socket path

-> IO Client

Connects to the master daemon and returns a Client.

connectServer :: ConnectConfig -> Bool -> FilePath -> IO Server

Creates and returns a server endpoint.

closeServer :: Server -> IO ()

Closes a server endpoint.

acceptClient :: Server -> IO Client

Accepts a client

closeClient :: Client -> IO ()

Closes the client socket.

sendMsg :: Client -> String -> IO ()

Sends a message over a transport.

recvUpdate :: ConnectConfig -> Handle -> ByteString -> IO (ByteString, ByteString)

recvMsg :: Client -> IO String

Waits for a message over a transport.

recvMsgExt :: Client -> IO RecvResult

Extended wrapper over recvMsg.

parseCall :: (JSON mth, JSON args) => String -> Result (mth, args)

Parse the required keys out of a call.

buildResponse

Arguments

:: Bool
Success

-> JSValue
The arguments

-> String
The serialized form

Serialize the response to String.

logMsg :: (Show e, JSON e, MonadLog m) => Handler i o -> i -> GenericResult e JSValue -> m ()

prepareMsg :: JSON e => GenericResult e JSValue -> (Bool, JSValue)

Processing client requests

type HandlerResult o = IO (Bool, GenericResult GanetiException o)

data Handler i o

Constructors

Handler

Fields

hParse :: JSValue -> JSValue -> Result i

parses method and its arguments into the input type

hInputLogShort :: i -> String

short description of an input, for the INFO logging level

hInputLogLong :: i -> String

long description of an input, for the DEBUG logging level

hExec :: i -> HandlerResult o

executes the handler on an input

handleJsonMessage :: JSON o => Handler i o -> i -> HandlerResult JSValue

handleRawMessage :: JSON o => Handler i o -> String -> IO (Bool, String)

handleClient :: JSON o => Handler i o -> Client -> IO Bool

clientLoop :: JSON o => Handler i o -> Client -> IO ()

listener :: JSON o => Handler i o -> Server -> IO ()

Main listener loop: accepts clients, forks an I/O thread to handle that client.