Ganeti.WConfd.ConfigModifications

Description

The WConfd functions for direct configuration manipulation

This module contains the client functions exported by WConfD for specific configuration manipulation.

Documentation

addInstance :: Instance -> ClientId -> WConfdMonad Bool

Add a new instance to the configuration, release DRBD minors, and commit temporary IPs, all while temporarily holding the config lock. Return True upon success and False if the config lock was not available and the client should retry.

TODO: add verifications to this call; the client should have a lock on the name of the instance, and no instance with the given uuid should exist.

The list of functions exported to RPC.

exportedFunctions :: [Name]