Test.Ganeti.Objects

Description

Unittests for ganeti-htools.

Arbitrary instances

genValidNetwork :: Gen Network

Generates a network instance with minimum netmasks of /24. Generating bigger networks slows down the tests, because long bit strings are generated for the reservations.

genBitString :: Int -> Gen String

genBitStringMaxLen :: Int -> Gen String

Generate an arbitrary string consisting of '0' and '1' of the maximum given length.

genEmptyCluster :: Int -> Gen ConfigData

Generator for config data with an empty cluster (no instances), with N defined nodes.

Test properties

prop_fillDict :: [(Int, Int)] -> [(Int, Int)] -> Property

prop_Disk_serialisation :: Disk -> Property

prop_Node_serialisation :: Node -> Property

prop_Inst_serialisation :: Instance -> Property

prop_Network_serialisation :: Network -> Property

prop_Config_serialisation :: Property

case_py_compat_networks :: Assertion

getNetworkProperties :: Network -> (Int, Int, Network)

case_py_compat_nodegroups :: Assertion

genNodeGroup :: Gen NodeGroup

testObjects :: Test