Test.Ganeti.TestCommon

Description

Unittest helpers for ganeti-htools.

Constants

maxMem :: Int

Maximum memory (1TiB, somewhat random value).

maxDsk :: Int

Maximum disk (8TiB, somewhat random value).

maxCpu :: Int

Max CPUs (1024, somewhat random value).

maxVcpuRatio :: Double

Max vcpu ratio (random value).

maxSpindleRatio :: Double

Max spindle ratio (random value).

maxNodes :: Int

Max nodes, used just to limit arbitrary instances for smaller opcode definitions (e.g. list of nodes in OpTestDelay).

maxOpCodes :: Int

Max opcodes or jobs in a submit job and submit many jobs.

Helper functions

(==?) :: (Show a, Eq a) => a -> a -> Property

Checks for equality with proper annotation. The first argument is the computed value, the second one the expected value.

(/=?) :: (Show a, Eq a) => a -> a -> Property

Checks for inequality with proper annotation. The first argument is the computed value, the second one the expected (not equal) value.

failTest :: String -> Property

Show a message and fail the test.

passTest :: Property

A True property.

pythonCmd :: IO String

Return the python binary to use. If the PYTHON environment variable is defined, use its value, otherwise use just "python".

runPython :: String -> String -> IO (ExitCode, String, String)

Run Python with an expression, returning the exit code, standard output and error.

checkPythonResult :: (ExitCode, String, String) -> IO String

Check python exit code, and fail via HUnit assertions if non-zero. Otherwise, return the standard output.

Arbitrary instances

newtype DNSChar

Defines a DNS name.

Constructors

DNSChar

Fields

dnsGetChar :: Char

genName :: Gen String

Generates a single name component.

genFQDN :: Gen String

Generates an entire FQDN.

genMaybe :: Gen a -> Gen (Maybe a)

Combinator that generates a Maybe using a sub-combinator.

newtype TagChar

Constructors

TagChar

Fields

tagGetChar :: Char

tagChar :: String

genTag :: Gen [TagChar]

genTags :: Gen [String]

Generates a list of tags (correctly upper bounded).

genFields :: Gen [String]

Generates a fields list. This uses the same character set as a DNS name (just for simplicity).

genUniquesList :: (Eq a, Arbitrary a, Ord a) => Int -> Gen a -> Gen [a]

Generates a list of a given size with non-duplicate elements.

newtype SmallRatio

Constructors

SmallRatio Double

genSetHelper :: Ord a => [a] -> Maybe Int -> Gen (Set a)

genSet :: (Ord a, Bounded a, Enum a) => Maybe Int -> Gen (Set a)

Generates a set of arbitrary elements.

genIp4Addr :: Gen NonEmptyString

Generate an arbitrary IPv4 address in textual form (non empty).

genIp4AddrStr :: Gen String

Generate an arbitrary IPv4 address in textual form.

genIp4NetWithNetmask :: Int -> Gen NonEmptyString

Generates an arbitrary IPv4 address with a given netmask in textual form.

genIp4Net :: Gen NonEmptyString

Generate an arbitrary IPv4 network in textual form.

netmask2NumHosts :: Int -> Int

Helper function to compute the number of hosts in a network given the netmask. (For IPv4 only.)

genIp6Addr :: Gen String

Generates an arbitrary IPv6 network address in textual form. The generated address is not simpflified, e. g. an address like 2607:f0d0:1002:0051:0000:0000:0000:0004 does not become 2607:f0d0:1002:51::4

genIp6Net :: Gen String

Generates an arbitrary IPv6 network in textual form.

Helper functions

testSerialisation :: (Eq a, Show a, JSON a) => a -> Property

Checks for serialisation idempotence.

resultProp :: Show a => GenericResult a b -> PropertyM IO b

Result to PropertyM IO.

getSourceDir :: IO FilePath

testDataFilename :: String -> String -> IO FilePath

readTestData :: String -> IO String

Returns the content of the specified haskell test data file.

genSample :: Gen a -> IO a

Generate arbitrary values in the IO monad. This is a simple wrapper over sample'.